Struct bevy_ecs::archetype::Archetype

pub struct Archetype { /* private fields */ }

Metadata for a single archetype within a World.

For more information, see the module level documentation.

Implementations

impl Archetype

pub fn id(&self) -> ArchetypeId

Fetches the ID for the archetype.

pub fn table_id(&self) -> TableId

Fetches the archetype’s Table ID.

pub fn entities(&self) -> &[ArchetypeEntity]

Fetches the entities contained in this archetype.

pub fn table_components(&self) -> impl Iterator<Item = ComponentId> + '_

Gets an iterator of all of the components stored in Tables.

All of the IDs are unique.

pub fn sparse_set_components(&self) -> impl Iterator<Item = ComponentId> + '_

Gets an iterator of all of the components stored in ComponentSparseSets.

All of the IDs are unique.

pub fn components(&self) -> impl Iterator<Item = ComponentId> + '_

Gets an iterator of all of the components in the archetype.

All of the IDs are unique.

pub fn component_count(&self) -> usize

Returns the total number of components in the archetype

pub fn edges(&self) -> &Edges

Fetches a immutable reference to the archetype’s Edges, a cache of archetypal relationships.

pub fn entity_table_row(&self, row: ArchetypeRow) -> TableRow

Fetches the row in the Table where the components for the entity at index is stored.

An entity’s archetype row can be fetched from EntityLocation::archetype_row, which can be retrieved from Entities::get.

Panics

This function will panic if index >= self.len().

pub fn len(&self) -> usize

Gets the total number of entities that belong to the archetype.

pub fn is_empty(&self) -> bool

Checks if the archetype has any entities.

pub fn contains(&self, component_id: ComponentId) -> bool

Checks if the archetype contains a specific component. This runs in O(1) time.

pub fn get_storage_type(&self, component_id: ComponentId) -> Option<StorageType>

Gets the type of storage where a component in the archetype can be found. Returns None if the component is not part of the archetype. This runs in O(1) time.

pub fn get_archetype_component_id( &self, component_id: ComponentId ) -> Option<ArchetypeComponentId>

Fetches the corresponding ArchetypeComponentId for a component in the archetype. Returns None if the component is not part of the archetype. This runs in O(1) time.

pub fn has_add_hook(&self) -> bool

Returns true if any of the components in this archetype have on_add hooks

pub fn has_insert_hook(&self) -> bool

Returns true if any of the components in this archetype have on_insert hooks

pub fn has_remove_hook(&self) -> bool

Returns true if any of the components in this archetype have on_remove hooks

pub fn has_add_observer(&self) -> bool

Returns true if any of the components in this archetype have at least one OnAdd observer

pub fn has_insert_observer(&self) -> bool

Returns true if any of the components in this archetype have at least one OnInsert observer

pub fn has_remove_observer(&self) -> bool

Returns true if any of the components in this archetype have at least one OnRemove observer