Struct bevy_ecs::archetype::Archetypes

pub struct Archetypes { /* private fields */ }

The backing store of all Archetypes within a World.

For more information, see the module level documentation.

Implementations

impl Archetypes

pub fn generation(&self) -> ArchetypeGeneration

Returns the “generation”, a handle to the current highest archetype ID.

This can be used with the Index Archetypes implementation to iterate over newly introduced Archetypes since the last time this function was called.

pub fn len(&self) -> usize

Fetches the total number of Archetypes within the world.

pub fn empty(&self) -> &Archetype

Fetches an immutable reference to the archetype without any components.

Shorthand for archetypes.get(ArchetypeId::EMPTY).unwrap()

pub fn get(&self, id: ArchetypeId) -> Option<&Archetype>

Fetches an immutable reference to an Archetype using its ID. Returns None if no corresponding archetype exists.

pub fn iter(&self) -> impl Iterator<Item = &Archetype>

Returns a read-only iterator over all archetypes.

pub fn archetype_components_len(&self) -> usize

Returns the number of components that are stored in archetypes. Note that if some component T is stored in more than one archetype, it will be counted once for each archetype it’s present in.