bevy_ecs::bundle

Struct Bundles

pub struct Bundles { /* private fields */ }

Metadata for bundles. Stores a BundleInfo for each type of Bundle in a given world.

Implementations

impl Bundles

pub fn len(&self) -> usize

The total number of Bundle registered in Storages.

pub fn is_empty(&self) -> bool

Returns true if no Bundle registered in Storages.

pub fn iter(&self) -> impl Iterator<Item = &BundleInfo>

Iterate over BundleInfo.

pub fn get(&self, bundle_id: BundleId) -> Option<&BundleInfo>

Gets the metadata associated with a specific type of bundle. Returns None if the bundle is not registered with the world.

pub fn get_id(&self, type_id: TypeId) -> Option<BundleId>

Gets the value identifying a specific type of bundle. Returns None if the bundle does not exist in the world, or if type_id does not correspond to a type of bundle.