bevy_ecs::bundle

Trait DynamicBundle

pub trait DynamicBundle: Sized {
    type Effect;

    // Required methods
    unsafe fn get_components(
        ptr: MovingPtr<'_, Self>,
        func: &mut impl FnMut(StorageType, OwningPtr<'_>),
    );
    unsafe fn apply_effect(
        ptr: MovingPtr<'_, MaybeUninit<Self>>,
        entity: &mut EntityWorldMut<'_>,
    );
}

The parts from Bundle that don’t require statically knowing the components of the bundle.

Required Associated Types

type Effect

An operation on the entity that happens after inserting this bundle.

Required Methods

unsafe fn get_components( ptr: MovingPtr<'_, Self>, func: &mut impl FnMut(StorageType, OwningPtr<'_>), )

Moves the components out of the bundle.

Safety

For callers:

Must be called exactly once before apply_effect

The StorageType argument passed into func must be correct for the component being fetched.

apply_effect must be called exactly once after this has been called if Effect: !NoBundleEffect

For implementors:

Implementors of this function must convert ptr into pointers to individual components stored within Self and call func on each of them in exactly the same order as Bundle::get_component_ids and BundleFromComponents::from_components.

If any part of ptr is to be accessed in apply_effect, it must not be dropped at any point in this function. Calling bevy_ptr::deconstruct_moving_ptr in this function automatically ensures this.

unsafe fn apply_effect( ptr: MovingPtr<'_, MaybeUninit<Self>>, entity: &mut EntityWorldMut<'_>, )

Applies the after-effects of spawning this bundle.

This is applied after all residual changes to the World, including flushing the internal command queue.

Safety

For callers:

Must be called exactly once after get_components has been called.

ptr must point to the instance of Self that get_components was called on, all of fields that were moved out of in get_components will not be valid anymore.

For implementors:

If any part of ptr is to be accessed in this function, it must not be dropped at any point in get_components. Calling bevy_ptr::deconstruct_moving_ptr in get_components automatically ensures this is the case.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.