bevy_ecs

Module change_detection

Types that detect when their internal data mutate.

Structs

MaybeLocation

A value that contains a T if the track_location feature is enabled, and is a ZST if it is not.

Mut

Unique mutable borrow of an entity’s component or of a resource.

MutUntyped

Unique mutable borrow of resources or an entity’s component.

NonSendMut

Unique borrow of a non-Send resource.

Ref

Shared borrow of an entity’s component with access to change detection. Similar to Mut but is immutable and so doesn’t require unique access.

Res

Shared borrow of a Resource.

ResMut

Unique mutable borrow of a Resource.

Constants

CHECK_TICK_THRESHOLD

The (arbitrarily chosen) minimum number of world tick increments between check_tick scans.

MAX_CHANGE_AGE

The maximum change tick difference that won’t overflow before the next check_tick scan.

Traits

DetectChanges

Types that can read change detection information. This change detection is controlled by DetectChangesMut types such as ResMut.

DetectChangesMut

Types that implement reliable change detection.