bevy_ecs::change_detection

Struct MaybeLocation

pub struct MaybeLocation<T: ?Sized = &'static Location<'static>> { /* private fields */ }

A value that contains a T if the track_location feature is enabled, and is a ZST if it is not.

The overall API is similar to Option, but whether the value is Some or None is set at compile time and is the same for all values.

If the track_location feature is disabled, then all functions on this type that return an MaybeLocation will have an empty body and should be removed by the optimizer.

This allows code to be written that will be checked by the compiler even when the feature is disabled, but that will be entirely removed during compilation.

Implementations

impl<T> MaybeLocation<T>

pub const fn new(_value: T) -> Self
where T: Copy,

Constructs a new MaybeLocation that wraps the given value.

This may only accept Copy types, since it needs to drop the value if the track_location feature is disabled, and non-Copy types cannot be dropped in const context. Use new_with if you need to construct a non-Copy value.

See also

new_with to initialize using a closure.

new_with_flattened to initialize using a closure that returns an Option<MaybeLocation<T>>.

pub fn new_with(_f: impl FnOnce() -> T) -> Self

Constructs a new MaybeLocation that wraps the result of the given closure.

See also

new to initialize using a value.

new_with_flattened to initialize using a closure that returns an Option<MaybeLocation<T>>.

pub fn map<U>(self, _f: impl FnOnce(T) -> U) -> MaybeLocation<U>

Maps an MaybeLocation<T> to MaybeLocation<U> by applying a function to a contained value.

pub fn zip<U>(self, _other: MaybeLocation<U>) -> MaybeLocation<(T, U)>

Converts a pair of MaybeLocation values to an MaybeLocation of a tuple.

pub fn unwrap_or_default(self) -> T
where T: Default,

Returns the contained value or a default. If the track_location feature is enabled, this always returns the contained value. If it is disabled, this always returns T::Default().

pub fn into_option(self) -> Option<T>

Converts an MaybeLocation to an Option to allow run-time branching. If the track_location feature is enabled, this always returns Some. If it is disabled, this always returns None.

impl<T> MaybeLocation<Option<T>>

pub fn new_with_flattened(_f: impl FnOnce() -> Option<MaybeLocation<T>>) -> Self

Constructs a new MaybeLocation that wraps the result of the given closure. If the closure returns Some, it unwraps the inner value.

See also

new to initialize using a value.

new_with to initialize using a closure.

pub fn transpose(self) -> Option<MaybeLocation<T>>

Transposes a MaybeLocation of an Option into an Option of a MaybeLocation.

This can be useful if you want to use the ? operator to exit early if the track_location feature is enabled but the value is not found.

If the track_location feature is enabled, this returns Some if the inner value is Some and None if the inner value is None.

If it is disabled, this always returns Some.

Example

let mut world = World::new();
let entity = world.spawn(()).id();
let location: MaybeLocation<Option<&'static Location<'static>>> =
    world.entities().entity_get_spawned_or_despawned_by(entity);
let location: MaybeLocation<&'static Location<'static>> = location.transpose()?;

See also

into_option to convert to an Option<Option<T>>. When used with Option::flatten, this will have a similar effect, but will return None when the track_location feature is disabled.

impl<T> MaybeLocation<&T>

pub const fn copied(&self) -> MaybeLocation<T>
where T: Copy,

Maps an MaybeLocation<&T> to an MaybeLocation<T> by copying the contents.

impl<T> MaybeLocation<&mut T>

pub const fn copied(&self) -> MaybeLocation<T>
where T: Copy,

Maps an MaybeLocation<&mut T> to an MaybeLocation<T> by copying the contents.

pub fn assign(&mut self, _value: MaybeLocation<T>)

Assigns the contents of an MaybeLocation<T> to an MaybeLocation<&mut T>.

impl<T: ?Sized> MaybeLocation<T>

pub const fn as_ref(&self) -> MaybeLocation<&T>

Converts from &MaybeLocation<T> to MaybeLocation<&T>.

pub const fn as_mut(&mut self) -> MaybeLocation<&mut T>

Converts from &mut MaybeLocation<T> to MaybeLocation<&mut T>.

pub fn as_deref(&self) -> MaybeLocation<&T::Target>
where T: Deref,

Converts from &MaybeLocation<T> to MaybeLocation<&T::Target>.

pub fn as_deref_mut(&mut self) -> MaybeLocation<&mut T::Target>
where T: DerefMut,

Converts from &mut MaybeLocation<T> to MaybeLocation<&mut T::Target>.

impl MaybeLocation

pub const fn caller() -> Self

Returns the source location of the caller of this function. If that function’s caller is annotated then its call location will be returned, and so on up the stack to the first call within a non-tracked function body.