Struct bevy_ecs::change_detection::Mut

pub struct Mut<'w, T: ?Sized> { /* private fields */ }

Unique mutable borrow of an entity’s component or of a resource.

This can be used in queries to opt into change detection on both their mutable and immutable forms, as opposed to &mut T, which only provides access to change detection while in its mutable form:

#[derive(Component, Clone)]
struct Name(String);

#[derive(Component, Clone, Copy)]
struct Health(f32);

#[derive(Component, Clone, Copy)]
struct Position {
    x: f32,
    y: f32,
};

#[derive(Component, Clone, Copy)]
struct Player {
    id: usize,
};

#[derive(QueryData)]
#[query_data(mutable)]
struct PlayerQuery {
    id: &'static Player,

    // Reacting to `PlayerName` changes is expensive, so we need to enable change detection when reading it.
    name: Mut<'static, Name>,

    health: &'static mut Health,
    position: &'static mut Position,
}

fn update_player_avatars(players_query: Query<PlayerQuery>) {
    // The item returned by the iterator is of type `PlayerQueryReadOnlyItem`.
    for player in players_query.iter() {
        if player.name.is_changed() {
            // Update the player's name. This clones a String, and so is more expensive.
            update_player_name(player.id, player.name.clone());
        }

        // Update the health bar.
        update_player_health(player.id, *player.health);

        // Update the player's position.
        update_player_position(player.id, *player.position);
    }
}

Implementations

impl<'w, T: ?Sized> Mut<'w, T>

pub fn new( value: &'w mut T, added: &'w mut Tick, last_changed: &'w mut Tick, last_run: Tick, this_run: Tick ) -> Self

Creates a new change-detection enabled smart pointer. In almost all cases you do not need to call this method manually, as instances of Mut will be created by engine-internal code.

Many use-cases of this method would be better served by Mut::map_unchanged or Mut::reborrow.

value - The value wrapped by this smart pointer.

added - A Tick that stores the tick when the wrapped value was created.

last_changed - A Tick that stores the last time the wrapped value was changed. This will be updated to the value of change_tick if the returned smart pointer is modified.

last_run - A Tick, occurring before this_run, which is used as a reference to determine whether the wrapped value is newly added or changed.

this_run - A Tick corresponding to the current point in time – “now”.

impl<'w, T: ?Sized> Mut<'w, T>

pub fn into_inner(self) -> &'w mut T

Consume self and return a mutable reference to the contained value while marking self as “changed”.

pub fn reborrow(&mut self) -> Mut<'_, T>

Returns a Mut<> with a smaller lifetime. This is useful if you have &mut Mut <T>, but you need a Mut<T>.

pub fn map_unchanged<U: ?Sized>( self, f: impl FnOnce(&mut T) -> &mut U ) -> Mut<'w, U>

Maps to an inner value by applying a function to the contained reference, without flagging a change.

You should never modify the argument passed to the closure – if you want to modify the data without flagging a change, consider using DetectChangesMut::bypass_change_detection to make your intent explicit.

// When run, zeroes the translation of every entity.
fn reset_positions(mut transforms: Query<&mut Transform>) {
    for transform in &mut transforms {
        // We pinky promise not to modify `t` within the closure.
        // Breaking this promise will result in logic errors, but will never cause undefined behavior.
        let mut translation = transform.map_unchanged(|t| &mut t.translation);
        // Only reset the translation if it isn't already zero;
        translation.set_if_neq(Vec2::ZERO);
    }
}

pub fn as_deref_mut(&mut self) -> Mut<'_, <T as Deref>::Target>
where T: DerefMut,

Allows you access to the dereferenced value of this pointer without immediately triggering change detection.