Struct bevy_ecs::change_detection::MutUntyped

pub struct MutUntyped<'w> { /* private fields */ }

Unique mutable borrow of resources or an entity’s component.

Similar to Mut, but not generic over the component type, instead exposing the raw pointer as a *mut ().

Usually you don’t need to use this and can instead use the APIs returning a Mut, but in situations where the types are not known at compile time or are defined outside of rust this can be used.

Implementations

impl<'w> MutUntyped<'w>

pub fn into_inner(self) -> PtrMut<'w>

Returns the pointer to the value, marking it as changed.

In order to avoid marking the value as changed, you need to call bypass_change_detection.

pub fn reborrow(&mut self) -> MutUntyped<'_>

Returns a MutUntyped with a smaller lifetime. This is useful if you have &mut MutUntyped, but you need a MutUntyped.

pub fn has_changed_since(&self, tick: Tick) -> bool

Returns true if this value was changed or mutably dereferenced either since a specific change tick.

pub fn as_mut(&mut self) -> PtrMut<'_>

Returns a pointer to the value without taking ownership of this smart pointer, marking it as changed.

In order to avoid marking the value as changed, you need to call bypass_change_detection.

pub fn as_ref(&self) -> Ptr<'_>

Returns an immutable pointer to the value without taking ownership.

pub fn map_unchanged<T: ?Sized>( self, f: impl FnOnce(PtrMut<'w>) -> &'w mut T ) -> Mut<'w, T>

Turn this MutUntyped into a Mut by mapping the inner PtrMut to another value, without flagging a change. This function is the untyped equivalent of Mut::map_unchanged.

You should never modify the argument passed to the closure – if you want to modify the data without flagging a change, consider using bypass_change_detection to make your intent explicit.

If you know the type of the value you can do

// SAFETY: ptr is of type `u8`
mut_untyped.map_unchanged(|ptr| unsafe { ptr.deref_mut::<u8>() });

If you have a ReflectFromPtr that you know belongs to this MutUntyped, you can do

// SAFETY: from the context it is known that `ReflectFromPtr` was made for the type of the `MutUntyped`
mut_untyped.map_unchanged(|ptr| unsafe { reflect_from_ptr.as_reflect_mut(ptr) });

pub unsafe fn with_type<T>(self) -> Mut<'w, T>

Transforms this MutUntyped into a Mut<T> with the same lifetime.

Safety

T must be the erased pointee type for this MutUntyped.