bevy_ecs::change_detection

Struct NonSendMut

pub struct NonSendMut<'w, T: ?Sized + 'static> { /* private fields */ }

Unique borrow of a non-Send resource.

Only Send resources may be accessed with the ResMut SystemParam. In case that the resource does not implement Send, this SystemParam wrapper can be used. This will instruct the scheduler to instead run the system on the main thread so that it doesn’t send the resource over to another thread.

This SystemParam fails validation if non-send resource doesn’t exist. This will cause a panic, but can be configured to do nothing or warn once.

Use Option<NonSendMut<T>> instead if the resource might not always exist.

Implementations

impl<'w, T: ?Sized> NonSendMut<'w, T>

pub fn into_inner(self) -> &'w mut T

Consume self and return a mutable reference to the contained value while marking self as “changed”.

pub fn reborrow(&mut self) -> Mut<'_, T>

Returns a Mut<> with a smaller lifetime. This is useful if you have &mut NonSendMut <T>, but you need a Mut<T>.

pub fn map_unchanged<U: ?Sized>( self, f: impl FnOnce(&mut T) -> &mut U, ) -> Mut<'w, U>

Maps to an inner value by applying a function to the contained reference, without flagging a change.

You should never modify the argument passed to the closure – if you want to modify the data without flagging a change, consider using DetectChangesMut::bypass_change_detection to make your intent explicit.

// When run, zeroes the translation of every entity.
fn reset_positions(mut transforms: Query<&mut Transform>) {
    for transform in &mut transforms {
        // We pinky promise not to modify `t` within the closure.
        // Breaking this promise will result in logic errors, but will never cause undefined behavior.
        let mut translation = transform.map_unchanged(|t| &mut t.translation);
        // Only reset the translation if it isn't already zero;
        translation.set_if_neq(Vec2::ZERO);
    }
}

pub fn filter_map_unchanged<U: ?Sized>( self, f: impl FnOnce(&mut T) -> Option<&mut U>, ) -> Option<Mut<'w, U>>

Optionally maps to an inner value by applying a function to the contained reference. This is useful in a situation where you need to convert a Mut<T> to a Mut<U>, but only if T contains U.

As with map_unchanged, you should never modify the argument passed to the closure.

pub fn try_map_unchanged<U: ?Sized, E>( self, f: impl FnOnce(&mut T) -> Result<&mut U, E>, ) -> Result<Mut<'w, U>, E>

Optionally maps to an inner value by applying a function to the contained reference, returns an error on failure. This is useful in a situation where you need to convert a Mut<T> to a Mut<U>, but only if T contains U.

As with map_unchanged, you should never modify the argument passed to the closure.

pub fn as_deref_mut(&mut self) -> Mut<'_, <T as Deref>::Target>
where T: DerefMut,

Allows you access to the dereferenced value of this pointer without immediately triggering change detection.