bevy_ecs::change_detection

Struct Res

pub struct Res<'w, T: ?Sized + Resource> { /* private fields */ }

Shared borrow of a Resource.

See the Resource documentation for usage.

If you need a unique mutable borrow, use ResMut instead.

Panics

Panics when used as a SystemParameter if the resource does not exist.

Use Option<Res<T>> instead if the resource might not always exist.

Implementations

impl<'w, T: Resource> Res<'w, T>

pub fn clone(this: &Self) -> Self

Copies a reference to a resource.

Note that unless you actually need an instance of Res<T>, you should prefer to just convert it to &T which can be freely copied.

pub fn into_inner(self) -> &'w T

Due to lifetime limitations of the Deref trait, this method can be used to obtain a reference of the Resource with a lifetime bound to 'w instead of the lifetime of the struct itself.