bevy_ecs::component

Function component_clone_via_clone

pub fn component_clone_via_clone<C: Clone + Component>(
    source: &SourceComponent<'_>,
    ctx: &mut ComponentCloneCtx<'_, '_>,
)

Component clone handler function implemented using the Clone trait. Can be set as clone handler for the specific component it is implemented for. It will panic if set as handler for any other component.