Struct bevy_ecs::component::ComponentTicks

pub struct ComponentTicks { /* private fields */ }

Records when a component or resource was added and when it was last mutably dereferenced (or added).

Implementations

impl ComponentTicks

pub fn is_added(&self, last_run: Tick, this_run: Tick) -> bool

Returns true if the component or resource was added after the system last ran.

pub fn is_changed(&self, last_run: Tick, this_run: Tick) -> bool

Returns true if the component or resource was added or mutably dereferenced after the system last ran.

pub fn last_changed_tick(&self) -> Tick

Returns the tick recording the time this component or resource was most recently changed.

pub fn added_tick(&self) -> Tick

Returns the tick recording the time this component or resource was added.

pub fn set_changed(&mut self, change_tick: Tick)

Manually sets the change tick.

This is normally done automatically via the DerefMut implementation on Mut<T>, ResMut<T>, etc. However, components and resources that make use of interior mutability might require manual updates.

Example

let world: World = unimplemented!();
let component_ticks: ComponentTicks = unimplemented!();

component_ticks.set_changed(world.read_change_tick());