bevy_ecs::component

Struct ComponentsQueuedRegistrator

pub struct ComponentsQueuedRegistrator<'w> { /* private fields */ }

A type that enables queuing registration in Components.

Note

These queued registrations return ComponentIds. These ids are not yet valid, but they will become valid when either ComponentsRegistrator::apply_queued_registrations is called or the same registration is made directly. In either case, the returned ComponentIds will be correct, but they are not correct yet.

Generally, that means these ComponentIds can be safely used for read-only purposes. Modifying the contents of the world through these ComponentIds directly without waiting for them to be fully registered and without then confirming that they have been fully registered is not supported. Hence, extra care is needed with these ComponentIds to ensure all safety rules are followed.

As a rule of thumb, if you have mutable access to ComponentsRegistrator, prefer to use that instead. Use this only if you need to know the id of a component but do not need to modify the contents of the world based on that id.

Implementations

impl<'w> ComponentsQueuedRegistrator<'w>

pub unsafe fn new(components: &'w Components, ids: &'w ComponentIds) -> Self

Constructs a new ComponentsQueuedRegistrator.

Safety

The Components and ComponentIds must match. For example, they must be from the same world.

pub fn queue_register_component<T: Component>(&self) -> ComponentId

This is a queued version of ComponentsRegistrator::register_component. This will reserve an id and queue the registration. These registrations will be carried out at the next opportunity.

If this has already been registered or queued, this returns the previous ComponentId.

Note

Technically speaking, the returned ComponentId is not valid, but it will become valid later. See type level docs for details.

pub fn queue_register_component_with_descriptor( &self, descriptor: ComponentDescriptor, ) -> ComponentId

This is a queued version of ComponentsRegistrator::register_component_with_descriptor. This will reserve an id and queue the registration. These registrations will be carried out at the next opportunity.

Note

Technically speaking, the returned ComponentId is not valid, but it will become valid later. See type level docs for details.

pub fn queue_register_resource<T: Resource>(&self) -> ComponentId

This is a queued version of ComponentsRegistrator::register_resource. This will reserve an id and queue the registration. These registrations will be carried out at the next opportunity.

If this has already been registered or queued, this returns the previous ComponentId.

Note

Technically speaking, the returned ComponentId is not valid, but it will become valid later. See type level docs for details.

pub fn queue_register_non_send<T: Any>(&self) -> ComponentId

This is a queued version of ComponentsRegistrator::register_non_send. This will reserve an id and queue the registration. These registrations will be carried out at the next opportunity.

If this has already been registered or queued, this returns the previous ComponentId.

Note

Technically speaking, the returned ComponentId is not valid, but it will become valid later. See type level docs for details.

pub fn queue_register_resource_with_descriptor( &self, descriptor: ComponentDescriptor, ) -> ComponentId

This is a queued version of ComponentsRegistrator::register_resource_with_descriptor. This will reserve an id and queue the registration. These registrations will be carried out at the next opportunity.

Note

Technically speaking, the returned ComponentId is not valid, but it will become valid later. See type level docs for details.