bevy_ecs::component

Struct ComponentsRegistrator

pub struct ComponentsRegistrator<'w> { /* private fields */ }

A Components wrapper that enables additional features, like registration.

Implementations

impl<'w> ComponentsRegistrator<'w>

pub unsafe fn new( components: &'w mut Components, ids: &'w mut ComponentIds, ) -> Self

Constructs a new ComponentsRegistrator.

Safety

The Components and ComponentIds must match. For example, they must be from the same world.

pub fn as_queued(&self) -> ComponentsQueuedRegistrator<'_>

Converts this ComponentsRegistrator into a ComponentsQueuedRegistrator. This is intended for use to pass this value to a function that requires ComponentsQueuedRegistrator. It is generally not a good idea to queue a registration when you can instead register directly on this type.

pub fn apply_queued_registrations(&mut self)

Applies every queued registration. This ensures that every valid ComponentId is registered, enabling retrieving ComponentInfo, etc.

pub fn register_component<T: Component>(&mut self) -> ComponentId

Registers a Component of type T with this instance. If a component of this type has already been registered, this will return the ID of the pre-existing component.

See also

Components::component_id()

ComponentsRegistrator::register_component_with_descriptor()

pub fn register_component_with_descriptor( &mut self, descriptor: ComponentDescriptor, ) -> ComponentId

Registers a component described by descriptor.

Note

If this method is called multiple times with identical descriptors, a distinct ComponentId will be created for each one.

See also

Components::component_id()

ComponentsRegistrator::register_component()

pub fn register_resource<T: Resource>(&mut self) -> ComponentId

Registers a Resource of type T with this instance. If a resource of this type has already been registered, this will return the ID of the pre-existing resource.

See also

Components::resource_id()

ComponentsRegistrator::register_resource_with_descriptor()

pub fn register_non_send<T: Any>(&mut self) -> ComponentId

Registers a non-send resource of type T with this instance. If a resource of this type has already been registered, this will return the ID of the pre-existing resource.

pub fn register_resource_with_descriptor( &mut self, descriptor: ComponentDescriptor, ) -> ComponentId

Registers a Resource described by descriptor.

Note

If this method is called multiple times with identical descriptors, a distinct ComponentId will be created for each one.

See also

Components::resource_id()

ComponentsRegistrator::register_resource()

pub fn any_queued_mut(&mut self) -> bool

Equivalent of Components::any_queued_mut

pub fn num_queued_mut(&mut self) -> usize

Equivalent of Components::any_queued_mut