bevy_ecs::entity

Module index_map

Contains the EntityIndexMap type, an IndexMap pre-configured to use EntityHash hashing.

This module is a lightweight wrapper around indexmap’s IndexMap that is more performant for Entity keys.

Structs

Drain

A draining iterator over the entries of an EntityIndexMap.

EntityIndexMap

A IndexMap pre-configured to use EntityHash hashing.

IntoIter

An owning iterator over the entries of an IndexMap.

IntoKeys

An owning iterator over the keys of an EntityIndexMap.

Iter

An iterator over the entries of an EntityIndexMap.

IterMut

A mutable iterator over the entries of an EntityIndexMap.

Keys

An iterator over the keys of an EntityIndexMap.

Slice

A dynamically-sized slice of key-value pairs in an EntityIndexMap.