bevy_ecs::entity::index_map

Struct Slice

pub struct Slice<V, S = EntityHash>(/* private fields */);

A dynamically-sized slice of key-value pairs in an EntityIndexMap.

Equivalent to an indexmap::map::Slice<V> whose source IndexMap uses EntityHash.

Implementations

impl<V> Slice<V>

pub const fn new<'a>() -> &'a Self

Returns an empty slice.

Equivalent to map::Slice::new.

pub fn new_mut<'a>() -> &'a mut Self

Returns an empty mutable slice.

Equivalent to map::Slice::new_mut.

pub const unsafe fn from_slice_unchecked(slice: &Slice<Entity, V>) -> &Self

Constructs a entity::index_map::Slice from a indexmap::map::Slice unsafely.

Safety

slice must stem from an IndexMap using EntityHash.

pub const unsafe fn from_slice_unchecked_mut( slice: &mut Slice<Entity, V>, ) -> &mut Self

Constructs a entity::index_map::Slice from a indexmap::map::Slice unsafely.

Safety

slice must stem from an IndexMap using EntityHash.

pub const fn as_inner(&self) -> &Slice<Entity, V>

Casts self to the inner slice.

pub unsafe fn from_boxed_slice_unchecked( slice: Box<Slice<Entity, V>>, ) -> Box<Self>

Constructs a boxed entity::index_map::Slice from a boxed indexmap::map::Slice unsafely.

Safety

slice must stem from an IndexMap using EntityHash.

pub fn as_boxed_inner(self: &Box<Self>) -> &Box<Slice<Entity, V>>

Casts a reference to self to the inner slice.

pub fn into_boxed_inner(self: Box<Self>) -> Box<Slice<Entity, V>>

Casts self to the inner slice.

pub fn get_index_mut(&mut self, index: usize) -> Option<(&Entity, &mut V)>

Get a key-value pair by index, with mutable access to the value.

Equivalent to map::Slice::get_index_mut.

pub fn get_range<R: RangeBounds<usize>>(&self, range: R) -> Option<&Self>

Returns a slice of key-value pairs in the given range of indices.

Equivalent to map::Slice::get_range.

pub fn get_range_mut<R: RangeBounds<usize>>( &mut self, range: R, ) -> Option<&mut Self>

Returns a mutable slice of key-value pairs in the given range of indices.

Equivalent to map::Slice::get_range_mut.

pub fn first_mut(&mut self) -> Option<(&Entity, &mut V)>

Get the first key-value pair, with mutable access to the value.

Equivalent to map::Slice::first_mut.

pub fn last_mut(&mut self) -> Option<(&Entity, &mut V)>

Get the last key-value pair, with mutable access to the value.

Equivalent to map::Slice::last_mut.

pub fn split_at(&self, index: usize) -> (&Self, &Self)

Divides one slice into two at an index.

Equivalent to map::Slice::split_at.

pub fn split_at_mut(&mut self, index: usize) -> (&mut Self, &mut Self)

Divides one mutable slice into two at an index.

Equivalent to map::Slice::split_at_mut.

pub fn split_first(&self) -> Option<((&Entity, &V), &Self)>

Returns the first key-value pair and the rest of the slice, or None if it is empty.

Equivalent to map::Slice::split_first.

pub fn split_first_mut(&mut self) -> Option<((&Entity, &mut V), &mut Self)>

Returns the first key-value pair and the rest of the slice, with mutable access to the value, or None if it is empty.

Equivalent to map::Slice::split_first_mut.

pub fn split_last(&self) -> Option<((&Entity, &V), &Self)>

Returns the last key-value pair and the rest of the slice, or None if it is empty.

Equivalent to map::Slice::split_last.

pub fn split_last_mut(&mut self) -> Option<((&Entity, &mut V), &mut Self)>

Returns the last key-value pair and the rest of the slice, with mutable access to the value, or None if it is empty.

Equivalent to map::Slice::split_last_mut.

pub fn iter(&self) -> Iter<'_, V>

Return an iterator over the key-value pairs of the map slice.

Equivalent to map::Slice::iter.

pub fn iter_mut(&mut self) -> IterMut<'_, V>

Return an iterator over the key-value pairs of the map slice.

Equivalent to map::Slice::iter_mut.

pub fn keys(&self) -> Keys<'_, V>

Return an iterator over the keys of the map slice.

Equivalent to map::Slice::keys.

pub fn into_keys(self: Box<Self>) -> IntoKeys<V>

Return an owning iterator over the keys of the map slice.

Equivalent to map::Slice::into_keys.

pub fn values_mut(&mut self) -> ValuesMut<'_, Entity, V>

Return an iterator over mutable references to the the values of the map slice.

Equivalent to map::Slice::values_mut.

pub fn into_values(self: Box<Self>) -> IntoValues<Entity, V>

Return an owning iterator over the values of the map slice.

Equivalent to map::Slice::into_values.