bevy_ecs::entity::index_set

Struct Slice

pub struct Slice<S = EntityHash>(/* private fields */);

A dynamically-sized slice of values in an EntityIndexSet.

Equivalent to an indexmap::set::Slice<V> whose source IndexSet uses EntityHash.

Implementations

impl Slice

pub const fn new<'a>() -> &'a Self

Returns an empty slice.

Equivalent to set::Slice::new.

pub const unsafe fn from_slice_unchecked(slice: &Slice<Entity>) -> &Self

Constructs a entity::index_set::Slice from a indexmap::set::Slice unsafely.

Safety

slice must stem from an IndexSet using EntityHash.

pub const unsafe fn from_slice_unchecked_mut( slice: &mut Slice<Entity>, ) -> &mut Self

Constructs a entity::index_set::Slice from a indexmap::set::Slice unsafely.

Safety

slice must stem from an IndexSet using EntityHash.

pub const fn as_inner(&self) -> &Slice<Entity>

Casts self to the inner slice.

pub unsafe fn from_boxed_slice_unchecked(slice: Box<Slice<Entity>>) -> Box<Self>

Constructs a boxed entity::index_set::Slice from a boxed indexmap::set::Slice unsafely.

Safety

slice must stem from an IndexSet using EntityHash.

pub fn as_boxed_inner(self: &Box<Self>) -> &Box<Slice<Entity>>

Casts a reference to self to the inner slice.

pub fn into_boxed_inner(self: Box<Self>) -> Box<Slice<Entity>>

Casts self to the inner slice.

pub fn get_range<R: RangeBounds<usize>>(&self, range: R) -> Option<&Self>

Returns a slice of values in the given range of indices.

Equivalent to set::Slice::get_range.

pub fn split_at(&self, index: usize) -> (&Self, &Self)

Divides one slice into two at an index.

Equivalent to set::Slice::split_at.

pub fn split_first(&self) -> Option<(&Entity, &Self)>

Returns the first value and the rest of the slice, or None if it is empty.

Equivalent to set::Slice::split_first.

pub fn split_last(&self) -> Option<(&Entity, &Self)>

Returns the last value and the rest of the slice, or None if it is empty.

Equivalent to set::Slice::split_last.

pub fn iter(&self) -> Iter<'_>

Return an iterator over the values of the set slice.

Equivalent to set::Slice::iter.