Struct bevy_ecs::entity::Entities

pub struct Entities { /* private fields */ }

Implementations

impl Entities

pub fn reserve_entities(&self, count: u32) -> ReserveEntitiesIterator<'_>

Reserve entity IDs concurrently.

Storage for entity generation and location is lazily allocated by calling flush.

pub fn reserve_entity(&self) -> Entity

Reserve one entity ID concurrently.

Equivalent to self.reserve_entities(1).next().unwrap(), but more efficient.

pub fn alloc(&mut self) -> Entity

Allocate an entity ID directly.

pub fn alloc_at(&mut self, entity: Entity) -> Option<EntityLocation>

Allocate a specific entity ID, overwriting its generation.

Returns the location of the entity currently using the given ID, if any. Location should be written immediately.

pub fn alloc_at_without_replacement(
&mut self,
entity: Entity
) -> AllocAtWithoutReplacement

Allocate a specific entity ID, overwriting its generation.

Returns the location of the entity currently using the given ID, if any.

pub fn free(&mut self, entity: Entity) -> Option<EntityLocation>

Destroy an entity, allowing it to be reused.

Must not be called while reserved entities are awaiting flush().

pub fn reserve(&mut self, additional: u32)

Ensure at least n allocations can succeed without reallocating.

pub fn contains(&self, entity: Entity) -> bool

Returns true if the Entities contains entity.

pub fn clear(&mut self)

pub fn get(&self, entity: Entity) -> Option<EntityLocation>

Returns Ok(Location { archetype: Archetype::invalid(), index: undefined }) for pending entities.

pub fn resolve_from_id(&self, id: u32) -> Option<Entity>

Get the Entity with a given id, if it exists in this Entities collection Returns None if this Entity is outside of the range of currently reserved Entities

Note: This method may return Entities which are currently free Note that contains will correctly return false for freed entities, since it checks the generation

pub unsafe fn flush(&mut self, init: impl FnMut(Entity, &mut EntityLocation))

Allocates space for entities previously reserved with reserve_entity or reserve_entities, then initializes each one using the supplied function.

Safety

Flush must set the entity location to the correct ArchetypeId for the given Entity each time init is called. This can be ArchetypeId::INVALID, provided the Entity has not been assigned to an Archetype.

pub fn flush_as_invalid(&mut self)

pub fn meta_len(&self) -> usize

Accessor for getting the length of the vec in self.meta

pub fn len(&self) -> u32

pub fn is_empty(&self) -> bool