Struct bevy_ecs::entity::EntityHasher

pub struct EntityHasher { /* private fields */ }

A very fast hash that is only designed to work on generational indices like Entity. It will panic if attempting to hash a type containing non-u64 fields.

This is heavily optimized for typical cases, where you have mostly live entities, and works particularly well for contiguous indices.

If you have an unusual case โ say all your indices are multiples of 256 or most of the entities are dead generations โ then you might want also to try AHasher for a slower hash computation but fewer lookup conflicts.