Struct bevy_ecs::entity::EntityMap

pub struct EntityMap { /* private fields */ }

A mapping from one set of entities to another.

The API generally follows HashMap, but each Entity is returned by value, as they are Copy.

This is typically used to coordinate data transfer between sets of entities, such as between a scene and the world or over the network. This is required as Entity identifiers are opaque; you cannot and do not want to reuse identifiers directly.

On its own, an EntityMap is not capable of allocating new entity identifiers, which is needed to map references to entities that lie outside the source entity set. To do this, an EntityMap can be wrapped in an EntityMapper which scopes it to a particular destination World and allows new identifiers to be allocated. This functionality can be accessed through Self::world_scope().

Implementations

impl EntityMap

pub fn insert(&mut self, from: Entity, to: Entity) -> Option<Entity>

Inserts an entities pair into the map.

If the map did not have from present, None is returned.

If the map did have from present, the value is updated, and the old value is returned.

pub fn remove(&mut self, entity: Entity) -> Option<Entity>

Removes an entity from the map, returning the mapped value of it if the entity was previously in the map.

pub fn entry(&mut self, entity: Entity) -> Entry<'_, Entity, Entity>

Gets the given entity’s corresponding entry in the map for in-place manipulation.

pub fn get(&self, entity: Entity) -> Option<Entity>

Returns the corresponding mapped entity.

pub fn keys(&self) -> impl Iterator<Item = Entity> + '_

An iterator visiting all keys in arbitrary order.

pub fn values(&self) -> impl Iterator<Item = Entity> + '_

An iterator visiting all values in arbitrary order.

pub fn len(&self) -> usize

Returns the number of elements in the map.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

pub fn iter(&self) -> impl Iterator<Item = (Entity, Entity)> + '_

An iterator visiting all (key, value) pairs in arbitrary order.

pub fn world_scope<R>( &mut self, world: &mut World, f: impl FnOnce(&mut World, &mut EntityMapper<'_>) -> R ) -> R

Creates an EntityMapper from this EntityMap and scoped to the provided World, then calls the provided function with it. This allows one to allocate new entity references in the provided World that are guaranteed to never point at a living entity now or in the future. This functionality is useful for safely mapping entity identifiers that point at entities outside the source world. The passed function, f, is called within the scope of the passed world. Its return value is then returned from world_scope as the generic type parameter R.