Struct bevy_ecs::entity::EntityMapper

pub struct EntityMapper<'m> { /* private fields */ }

A wrapper for HashMap<Entity, Entity>, augmenting it with the ability to allocate new Entity references in a destination world. These newly allocated references are guaranteed to never point to any living entity in that world.

References are allocated by returning increasing generations starting from an internally initialized base Entity. After it is finished being used by MapEntities implementations, this entity is despawned and the requisite number of generations reserved.

Implementations

impl<'m> EntityMapper<'m>

pub fn get_or_reserve(&mut self, entity: Entity) -> Entity

Returns the corresponding mapped entity or reserves a new dead entity ID if it is absent.

pub fn get_map(&'m self) -> &'m HashMap<Entity, Entity>

Gets a reference to the underlying HashMap<Entity, Entity>.

pub fn get_map_mut(&'m mut self) -> &'m mut HashMap<Entity, Entity>

Gets a mutable reference to the underlying HashMap<Entity, Entity>.

pub fn world_scope<R>( entity_map: &'m mut HashMap<Entity, Entity>, world: &mut World, f: impl FnOnce(&mut World, &mut Self) -> R ) -> R

Creates an EntityMapper from a provided World and HashMap<Entity, Entity>, then calls the provided function with it. This allows one to allocate new entity references in this World that are guaranteed to never point at a living entity now or in the future. This functionality is useful for safely mapping entity identifiers that point at entities outside the source world. The passed function, f, is called within the scope of this world. Its return value is then returned from world_scope as the generic type parameter R.