bevy_ecs::entity

Struct SceneEntityMapper

pub struct SceneEntityMapper<'m> { /* private fields */ }

A wrapper for EntityHashMap<Entity>, augmenting it with the ability to allocate new Entity references in a destination world. These newly allocated references are guaranteed to never point to any living entity in that world.

References are allocated by returning increasing generations starting from an internally initialized base Entity. After it is finished being used, this entity is despawned and the requisite number of generations reserved.

Implementations

impl<'m> SceneEntityMapper<'m>

pub fn get_map(&'m self) -> &'m EntityHashMap<Entity>

Gets a reference to the underlying EntityHashMap<Entity>.

pub fn get_map_mut(&'m mut self) -> &'m mut EntityHashMap<Entity>

Gets a mutable reference to the underlying EntityHashMap<Entity>.

pub fn new(map: &'m mut EntityHashMap<Entity>, world: &mut World) -> Self

Creates a new SceneEntityMapper, spawning a temporary base Entity in the provided World

pub fn finish(self, world: &mut World)

Reserves the allocated references to dead entities within the world. This frees the temporary base Entity while reserving extra generations. Because this makes the SceneEntityMapper unable to safely allocate any more references, this method takes ownership of self in order to render it unusable.

pub fn world_scope<R>( entity_map: &'m mut EntityHashMap<Entity>, world: &mut World, f: impl FnOnce(&mut World, &mut Self) -> R, ) -> R

Creates an SceneEntityMapper from a provided World and EntityHashMap<Entity>, then calls the provided function with it. This allows one to allocate new entity references in this World that are guaranteed to never point at a living entity now or in the future. This functionality is useful for safely mapping entity identifiers that point at entities outside the source world. The passed function, f, is called within the scope of this world. Its return value is then returned from world_scope as the generic type parameter R.