bevy_ecs::entity

Module unique_slice

A wrapper around entity slices with a uniqueness invariant.

Structs

UniqueEntityEquivalentSlice

A slice that contains only unique entities.

UniqueEntityEquivalentSliceIter

An iterator that yields &UniqueEntityEquivalentSlice. Note that an entity may appear in multiple slices, depending on the wrapped iterator.

UniqueEntityEquivalentSliceIterMut

An iterator that yields &mut UniqueEntityEquivalentSlice. Note that an entity may appear in multiple slices, depending on the wrapped iterator.

Functions

cast_slice_of_mut_unique_entity_slice_mut⚠

Casts a mutable slice of mutable entity slices to a slice of mutable UniqueEntityEquivalentSlices.

cast_slice_of_unique_entity_slice⚠

Casts a slice of entity slices to a slice of UniqueEntityEquivalentSlices.

cast_slice_of_unique_entity_slice_mut⚠

Casts a mutable slice of entity slices to a slice of UniqueEntityEquivalentSlices.

from_mut

Converts a reference to T into a slice of length 1 (without copying).

from_raw_parts⚠

Forms a slice from a pointer and a length.

from_raw_parts_mut⚠

Performs the same functionality as from_raw_parts, except that a mutable slice is returned.

from_ref

Converts a reference to T into a slice of length 1 (without copying).

Type Aliases

ChunkBy

An iterator over slice in (non-overlapping) chunks separated by a predicate.

ChunkByMut

An iterator over slice in (non-overlapping) mutable chunks separated by a predicate.

Chunks

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the beginning of the slice.

ChunksExact

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the beginning of the slice.

ChunksExactMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the beginning of the slice.

ChunksMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the beginning of the slice.

Iter

Immutable slice iterator.

IterMut

Mutable slice iterator.

RChunks

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the end of the slice.

RChunksExact

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the end of the slice.

RChunksExactMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the end of the slice.

RChunksMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the end of the slice.

RSplit

An iterator over subslices separated by elements that match a predicate function, starting from the end of the slice.

RSplitMut

An iterator over the subslices of the vector which are separated by elements that match pred, starting from the end of the slice.

RSplitN

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits, starting from the end of the slice.

RSplitNMut

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits, starting from the end of the slice.

Split

An iterator over subslices separated by elements that match a predicate function.

SplitInclusive

An iterator over subslices separated by elements that match a predicate function.

SplitInclusiveMut

An iterator over the mutable subslices of the vector which are separated by elements that match pred. Unlike SplitMut, it contains the matched parts in the ends of the subslices.

SplitMut

An iterator over the mutable subslices of the vector which are separated by elements that match pred.

SplitN

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

SplitNMut

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

UniqueEntitySlice

A slice that contains only unique Entity.

Windows

An iterator over overlapping subslices of length size.