bevy_ecs::entity::unique_vec

Struct UniqueEntityEquivalentVec

pub struct UniqueEntityEquivalentVec<T: EntityEquivalent>(/* private fields */);

A Vec that contains only unique entities.

“Unique” means that x != y holds for any 2 entities in this collection. This is always true when less than 2 entities are present.

This type is best obtained by its FromEntitySetIterator impl, via either EntityIterator::collect_set or UniqueEntityEquivalentVec::from_entity_iter.

While this type can be constructed via Iterator::collect, doing so is inefficient, and not recommended.

When T is Entity, use the UniqueEntityVec alias.

Implementations

impl<T: EntityEquivalent> UniqueEntityEquivalentVec<T>

pub const fn new() -> Self

Constructs a new, empty UniqueEntityEquivalentVec<T>.

Equivalent to Vec::new.

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty UniqueEntityEquivalentVec<T> with at least the specified capacity.

Equivalent to Vec::with_capacity

pub unsafe fn from_raw_parts( ptr: *mut T, length: usize, capacity: usize, ) -> Self

Creates a UniqueEntityEquivalentVec<T> directly from a pointer, a length, and a capacity.

Equivalent to Vec::from_raw_parts.

Safety

It must be safe to call Vec::from_raw_parts with these inputs, and the resulting Vec must only contain unique elements.

pub unsafe fn from_vec_unchecked(vec: Vec<T>) -> Self

Constructs a UniqueEntityEquivalentVec from a Vec<T> unsafely.

Safety

vec must contain only unique elements.

pub fn into_inner(self) -> Vec<T>

Returns the inner Vec<T>.

pub fn as_vec(&self) -> &Vec<T>

Returns a reference to the inner Vec<T>.

pub unsafe fn as_mut_vec(&mut self) -> &mut Vec<T>

Returns a mutable reference to the inner Vec<T>.

Safety

The elements of this Vec must always remain unique, even while this mutable reference is live.

pub fn capacity(&self) -> usize

Returns the total number of elements the vector can hold without reallocating.

Equivalent to Vec::capacity.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given Vec<T>.

Equivalent to Vec::reserve.

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for at least additional more elements to be inserted in the given UniqueEntityEquivalentVec<T>.

Equivalent to Vec::reserve_exact.

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

Tries to reserve capacity for at least additional more elements to be inserted in the given Vec<T>.

Equivalent to Vec::try_reserve.

pub fn try_reserve_exact( &mut self, additional: usize, ) -> Result<(), TryReserveError>

Tries to reserve the minimum capacity for at least additional elements to be inserted in the given Vec<T>.

Equivalent to Vec::try_reserve_exact.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

Equivalent to Vec::shrink_to_fit.

pub fn shrink_to(&mut self, min_capacity: usize)

Shrinks the capacity of the vector with a lower bound.

Equivalent to Vec::shrink_to.

pub fn into_boxed_slice(self) -> Box<UniqueEntityEquivalentSlice<T>>

Converts the vector into Box<UniqueEntityEquivalentSlice<T>>.

pub fn as_slice(&self) -> &UniqueEntityEquivalentSlice<T>

Extracts a slice containing the entire vector.

pub fn as_mut_slice(&mut self) -> &mut UniqueEntityEquivalentSlice<T>

Extracts a mutable slice of the entire vector.

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len elements and dropping the rest.

Equivalent to Vec::truncate.

pub fn as_ptr(&self) -> *const T

Returns a raw pointer to the vector’s buffer, or a dangling raw pointer valid for zero sized reads if the vector didn’t allocate.

Equivalent to Vec::as_ptr.

pub fn as_mut_ptr(&mut self) -> *mut T

Returns a raw mutable pointer to the vector’s buffer, or a dangling raw pointer valid for zero sized reads if the vector didn’t allocate.

Equivalent to Vec::as_mut_ptr.

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of the vector to new_len.

Equivalent to Vec::set_len.

Safety

It must be safe to call Vec::set_len with these inputs, and the resulting Vec must only contain unique elements.

pub fn swap_remove(&mut self, index: usize) -> T

Removes an element from the vector and returns it.

Equivalent to Vec::swap_remove.

pub unsafe fn insert(&mut self, index: usize, element: T)

Inserts an element at position index within the vector, shifting all elements after it to the right.

Equivalent to Vec::insert.

Safety

No T contained by self may equal element.

pub fn remove(&mut self, index: usize) -> T

Removes and returns the element at position index within the vector, shifting all elements after it to the left.

Equivalent to Vec::remove.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

Equivalent to Vec::retain.

pub unsafe fn retain_mut<F>(&mut self, f: F)
where F: FnMut(&mut T) -> bool,

Retains only the elements specified by the predicate, passing a mutable reference to it.

Equivalent to Vec::retain_mut.

Safety

self must only contain unique elements after each individual execution of f.

pub unsafe fn dedup_by_key<F, K>(&mut self, key: F)
where F: FnMut(&mut T) -> K, K: PartialEq,

Removes all but the first of consecutive elements in the vector that resolve to the same key.

Equivalent to Vec::dedup_by_key.

Safety

self must only contain unique elements after each individual execution of key.

pub unsafe fn dedup_by<F>(&mut self, same_bucket: F)
where F: FnMut(&mut T, &mut T) -> bool,

Removes all but the first of consecutive elements in the vector satisfying a given equality relation.

Equivalent to Vec::dedup_by.

Safety

self must only contain unique elements after each individual execution of same_bucket.

pub unsafe fn push(&mut self, value: T)

Appends an element to the back of a collection.

Equivalent to Vec::push.

Safety

No T contained by self may equal element.

pub unsafe fn append(&mut self, other: &mut UniqueEntityEquivalentVec<T>)

Moves all the elements of other into self, leaving other empty.

Equivalent to Vec::append.

Safety

other must contain no elements that equal any element in self.

pub fn pop(&mut self) -> Option<T>

Removes the last element from a vector and returns it, or None if it is empty.

Equivalent to Vec::pop.

pub fn drain<R>(&mut self, range: R) -> Drain<'_, T>
where R: RangeBounds<usize>,

Removes the specified range from the vector in bulk, returning all removed elements as an iterator.

Equivalent to Vec::drain.

pub fn clear(&mut self)

Clears the vector, removing all values.

Equivalent to Vec::clear.

pub fn len(&self) -> usize

Returns the number of elements in the vector, also referred to as its ‘length’.

Equivalent to Vec::len.

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

Equivalent to Vec::is_empty.

pub fn split_off(&mut self, at: usize) -> Self

Splits the collection into two at the given index.

Equivalent to Vec::split_off.

pub unsafe fn resize_with<F>(&mut self, new_len: usize, f: F)
where F: FnMut() -> T,

Resizes the Vec in-place so that len is equal to new_len.

Equivalent to Vec::resize_with.

Safety

f must only produce unique T, and none of these may equal any T in self.

pub fn leak<'a>(self) -> &'a mut UniqueEntityEquivalentSlice<T>

Consumes and leaks the Vec, returning a mutable reference to the contents, &'a mut UniqueEntityEquivalentSlice<T>.

pub fn spare_capacity_mut(&mut self) -> &mut [MaybeUninit<T>]

Returns the remaining spare capacity of the vector as a slice of MaybeUninit<T>.

Equivalent to Vec::spare_capacity_mut.

pub unsafe fn splice<R, I>( &mut self, range: R, replace_with: I, ) -> Splice<'_, <I as IntoIterator>::IntoIter>
where R: RangeBounds<usize>, I: EntitySet<Item = T>,

Creates a splicing iterator that replaces the specified range in the vector with the given replace_with iterator and yields the removed items.

Equivalent to Vec::splice.

Safety

replace_with must not yield any elements that equal any elements in self, except for those in range.