bevy_ecs::entity_disabling

Struct Disabled

pub struct Disabled;

A marker component for disabled entities.

Semantically, this component is used to mark entities that are temporarily disabled (typically for gameplay reasons), but will likely be re-enabled at some point.

Like all disabling components, this only disables the entity itself, not its children or other entities that reference it. To disable an entire tree of entities, use EntityCommands::insert_recursive.

Every World has a default query filter that excludes entities with this component, registered in the DefaultQueryFilters resource. See the module docs for more info.