bevy_ecs::error

Trait CommandWithEntity

pub trait CommandWithEntity<Out> {
    // Required method
    fn with_entity(self, entity: Entity) -> impl Command<Out> + HandleError<Out>;
}

Passes in a specific entity to an EntityCommand, resulting in a Command that internally runs the EntityCommand on that entity.

Required Methods

fn with_entity(self, entity: Entity) -> impl Command<Out> + HandleError<Out>

Passes in a specific entity to an EntityCommand, resulting in a Command that internally runs the EntityCommand on that entity.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.