bevy_ecs::error

Trait HandleError

pub trait HandleError<Out = ()>: Send + 'static {
    // Required methods
    fn handle_error_with(self, error_handler: ErrorHandler) -> impl Command;
    fn handle_error(self) -> impl Command;
    fn ignore_error(self) -> impl Command;
}

Takes a Command that potentially returns a Result and uses a given error handler function to convert it into a Command that internally handles an error if it occurs and returns ().

Required Methods

fn handle_error_with(self, error_handler: ErrorHandler) -> impl Command

Takes a Command that returns a Result and uses a given error handler function to convert it into a Command that internally handles an error if it occurs and returns ().

fn handle_error(self) -> impl Command

Takes a Command that returns a Result and uses the default error handler function to convert it into a Command that internally handles an error if it occurs and returns ().

fn ignore_error(self) -> impl Command

Takes a Command that returns a Result and ignores any error that occurs.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.