Module bevy_ecs::event

Event handling types.

Structs

EventId

An EventId uniquely identifies an event stored in a specific World.

EventReader

Reads events of type T in order and tracks which events have already been read.

EventWriter

Sends events of type T.

Events

An event collection that represents the events that occurred within the last two Events::update calls. Events can be written to using an EventWriter and are typically cheaply read using an EventReader.

ManualEventIterator

An iterator that yields any unread events from an EventReader or ManualEventReader.

ManualEventIteratorWithId

An iterator that yields any unread events (and their IDs) from an EventReader or ManualEventReader.

ManualEventReader

Stores the state for an EventReader. Access to the Events<E> resource is required to read any incoming events.

Traits

Event

A type that can be stored in an Events<E> resource You can conveniently access events using the EventReader and EventWriter system parameter.

Derive Macros

Event