Struct bevy_ecs::event::EventReader

pub struct EventReader<'w, 's, E: Event> { /* private fields */ }

Reads events of type T in order and tracks which events have already been read.

Implementations

impl<'w, 's, E: Event> EventReader<'w, 's, E>

pub fn iter(&mut self) -> ManualEventIterator<'_, E>

Iterates over the events this EventReader has not seen yet. This updates the EventReader’s event counter, which means subsequent event reads will not include events that happened before now.

pub fn iter_with_id(&mut self) -> ManualEventIteratorWithId<'_, E>

Like iter, except also returning the EventId of the events.

pub fn len(&self) -> usize

Determines the number of events available to be read from this EventReader without consuming any.

pub fn is_empty(&self) -> bool

Returns true if there are no events available to read.

Example

The following example shows a useful pattern where some behavior is triggered if new events are available. EventReader::clear() is used so the same events don’t re-trigger the behavior the next time the system runs.

#[derive(Event)]
struct CollisionEvent;

fn play_collision_sound(mut events: EventReader<CollisionEvent>) {
    if !events.is_empty() {
        events.clear();
        // Play a sound
    }
}

pub fn clear(&mut self)

Consumes all available events.

This means these events will not appear in calls to EventReader::iter() or EventReader::iter_with_id() and EventReader::is_empty() will return true.

For usage, see EventReader::is_empty().