bevy_ecs::event
Trait Event

pub trait Event:
    Send
    + Sync
    + 'static { }

A type that can be stored in an Events<E> resource You can conveniently access events using the EventReader and EventWriter system parameter.

Events must be thread-safe.