bevy_ecs::hierarchy

Function validate_parent_has_component

pub fn validate_parent_has_component<C: Component>(
    world: DeferredWorld<'_>,
    _: HookContext,
)

An on_insert component hook that when run, will validate that the parent of a given entity contains component C. This will print a warning if the parent does not contain C.