bevy_ecs::hierarchy

Struct ChildOf

pub struct ChildOf(pub Entity);

Stores the parent entity of this child entity with this component.

This is a Relationship component, and creates the canonical “parent / child” hierarchy. This is the “source of truth” component, and it pairs with the Children RelationshipTarget.

This relationship should be used for things like:

Organizing entities in a scene

Propagating configuration or data inherited from a parent, such as “visibility” or “world-space global transforms”.

Ensuring a hierarchy is despawned when an entity is despawned.

ChildOf contains a single “target” Entity. When ChildOf is inserted on a “source” entity, the “target” entity will automatically (and immediately, via a component hook) have a Children component inserted, and the “source” entity will be added to that Children instance.

If the ChildOf component is replaced with a different “target” entity, the old target’s Children will be automatically (and immediately, via a component hook) be updated to reflect that change.

Likewise, when the ChildOf component is removed, the “source” entity will be removed from the old target’s Children. If this results in Children being empty, Children will be automatically removed.

When a parent is despawned, all children (and their descendants) will also be despawned.

You can create parent-child relationships in a variety of ways. The most direct way is to insert a ChildOf component:

let root = world.spawn_empty().id();
let child1 = world.spawn(ChildOf(root)).id();
let child2 = world.spawn(ChildOf(root)).id();
let grandchild = world.spawn(ChildOf(child1)).id();

assert_eq!(&**world.entity(root).get::<Children>().unwrap(), &[child1, child2]);
assert_eq!(&**world.entity(child1).get::<Children>().unwrap(), &[grandchild]);

world.entity_mut(child2).remove::<ChildOf>();
assert_eq!(&**world.entity(root).get::<Children>().unwrap(), &[child1]);

world.entity_mut(root).despawn();
assert!(world.get_entity(root).is_err());
assert!(world.get_entity(child1).is_err());
assert!(world.get_entity(grandchild).is_err());

However if you are spawning many children, you might want to use the EntityWorldMut::with_children helper instead:

let mut child1 = Entity::PLACEHOLDER;
let mut child2 = Entity::PLACEHOLDER;
let mut grandchild = Entity::PLACEHOLDER;
let root = world.spawn_empty().with_children(|p| {
    child1 = p.spawn_empty().with_children(|p| {
        grandchild = p.spawn_empty().id();
    }).id();
    child2 = p.spawn_empty().id();
}).id();

assert_eq!(&**world.entity(root).get::<Children>().unwrap(), &[child1, child2]);
assert_eq!(&**world.entity(child1).get::<Children>().unwrap(), &[grandchild]);

Tuple Fields
0: Entity
Implementations

impl ChildOf

pub fn parent(&self) -> Entity

The parent entity of this child entity.