bevy_ecs::hierarchy

Struct Children

pub struct Children(/* private fields */);

Tracks which entities are children of this parent entity.

A RelationshipTarget collection component that is populated with entities that “target” this entity with the ChildOf Relationship component.

Together, these components form the “canonical parent-child hierarchy”. See the ChildOf component for the full description of this relationship and instructions on how to use it.

Usage

Like all RelationshipTarget components, this data should not be directly manipulated to avoid desynchronization. Instead, modify the ChildOf components on the “source” entities.

To access the children of an entity, you can iterate over the Children component, using the IntoIterator trait. For more complex access patterns, see the RelationshipTarget trait.

Implementations

impl Children

pub fn swap(&mut self, a_index: usize, b_index: usize)

Swaps the child at a_index with the child at b_index.

pub fn sort_by<F>(&mut self, compare: F)
where F: FnMut(&Entity, &Entity) -> Ordering,

Sorts children stably in place using the provided comparator function.

For the underlying implementation, see slice::sort_by.

For the unstable version, see sort_unstable_by.

See also sort_by_key, sort_by_cached_key.

pub fn sort_by_key<K, F>(&mut self, compare: F)
where F: FnMut(&Entity) -> K, K: Ord,

Sorts children stably in place using the provided key extraction function.

For the underlying implementation, see slice::sort_by_key.

For the unstable version, see sort_unstable_by_key.

See also sort_by, sort_by_cached_key.

pub fn sort_by_cached_key<K, F>(&mut self, compare: F)
where F: FnMut(&Entity) -> K, K: Ord,

Sorts children stably in place using the provided key extraction function. Only evaluates each key at most once per sort, caching the intermediate results in memory.

For the underlying implementation, see slice::sort_by_cached_key.

See also sort_by, sort_by_key.

pub fn sort_unstable_by<F>(&mut self, compare: F)
where F: FnMut(&Entity, &Entity) -> Ordering,

Sorts children unstably in place using the provided comparator function.

For the underlying implementation, see slice::sort_unstable_by.

For the stable version, see sort_by.

See also sort_unstable_by_key.

pub fn sort_unstable_by_key<K, F>(&mut self, compare: F)
where F: FnMut(&Entity) -> K, K: Ord,

Sorts children unstably in place using the provided key extraction function.

For the underlying implementation, see slice::sort_unstable_by_key.

For the stable version, see sort_by_key.

See also sort_unstable_by.