Module bevy_ecs::identifier

A module for the unified Identifier ID struct, for use as a representation of multiple types of IDs in a single, packed type. Allows for describing an crate::entity::Entity, or other IDs that can be packed and expressed within a u64 sized type. Identifiers cannot be created directly, only able to be converted from other compatible IDs.

Modules

error

Error types for super::Identifier conversions. An ID can be converted to various kinds, but these can fail if they are not valid forms of those kinds. The error type in this module encapsulates the various failure modes.

Structs

Identifier

A unified identifier for all entity and similar IDs. Has the same size as a u64 integer, but the layout is split between a 32-bit low segment, a 31-bit high segment, and the significant bit reserved as type flags to denote entity kinds.