bevy_ecs::label

Trait DynEq

pub trait DynEq: Any {
    // Required method
    fn dyn_eq(&self, other: &dyn DynEq) -> bool;
}

An object safe version of Eq. This trait is automatically implemented for any 'static type that implements Eq.

Required Methods

fn dyn_eq(&self, other: &dyn DynEq) -> bool

This method tests for self and other values to be equal.

Implementers should avoid returning true when the underlying types are not the same.