bevy_ecs::lifecycle

Struct RemovedComponentMessages

pub struct RemovedComponentMessages { /* private fields */ }

Stores the RemovedComponents event buffers for all types of component in a given World.

Implementations

impl RemovedComponentMessages

pub fn new() -> Self

Creates an empty storage buffer for component removal messages.

pub fn update(&mut self)

For each type of component, swaps the event buffers and clears the oldest event buffer. In general, this should be called once per frame/update.

pub fn iter( &self, ) -> impl Iterator<Item = (&ComponentId, &Messages<RemovedComponentEntity>)>

Returns an iterator over components and their entity messages.

pub fn get( &self, component_id: impl Into<ComponentId>, ) -> Option<&Messages<RemovedComponentEntity>>

Gets the event storage for a given component.

pub fn send(&mut self, component_id: impl Into<ComponentId>, entity: Entity)

๐Deprecated since 0.17.0: Use RemovedComponentMessages:write instead.

Sends a removal message for the specified component.

pub fn write(&mut self, component_id: impl Into<ComponentId>, entity: Entity)

Writes a removal message for the specified component.