bevy_ecs::name

Struct NameOrEntity

pub struct NameOrEntity {
    pub name: Option<&'static Name>,
    pub entity: Entity,
}

Convenient query for giving a human friendly name to an entity.

fn increment_score(mut scores: Query<(NameOrEntity, &mut Score)>) {
    for (name, mut score) in &mut scores {
        score.0 += 1.0;
        if score.0.is_nan() {
            log::error!("Score for {name} is invalid");
        }
    }
}

Implementation

The Display impl for NameOrEntity returns the Name where there is one or {index}v{generation} for entities without one.

Fields
name: Option<&'static Name>

A Name that the entity might have that is displayed if available.

entity: Entity

The unique identifier of the entity as a fallback.