bevy_ecs::observer

Struct CachedObservers

pub struct CachedObservers { /* private fields */ }

Collection of ObserverRunner for Observer registered to a particular event.

This is stored inside of Observers, specialized for each kind of observer.

Implementations

impl CachedObservers

pub fn global_observers(&self) -> &ObserverMap

Observers watching for any time this event is triggered, regardless of target. These will also respond to events targeting specific components or entities

pub fn component_observers( &self, ) -> &HashMap<ComponentId, CachedComponentObservers>

Returns observers watching for triggers of events for a specific component.

pub fn entity_observers(&self) -> &EntityHashMap<ObserverMap>

Returns observers watching for triggers of events for a specific entity.