bevy_ecs::observer

Struct Trigger

pub struct Trigger<'w, E, B: Bundle = ()> { /* private fields */ }

Type containing triggered Event information for a given run of an Observer. This contains the Event data itself. If it was triggered for a specific Entity, it includes that as well.

Implementations

impl<'w, E, B: Bundle> Trigger<'w, E, B>

pub fn new(event: &'w mut E, trigger: ObserverTrigger) -> Self

Creates a new trigger for the given event and observer information.

pub fn event_type(&self) -> ComponentId

Returns the event type of this trigger.

pub fn event(&self) -> &E

Returns a reference to the triggered event.

pub fn event_mut(&mut self) -> &mut E

Returns a mutable reference to the triggered event.

pub fn event_ptr(&self) -> Ptr<'_>

Returns a pointer to the triggered event.

pub fn entity(&self) -> Entity

Returns the entity that triggered the observer, could be Entity::PLACEHOLDER.