Module bevy_ecs::prelude

Re-exports

pub use crate::reflect::ReflectComponent;

pub use crate::bundle::Bundle;

pub use crate::entity::Entity;

pub use crate::schedule::Schedule;

pub use crate::world::FromWorld;

pub use crate::world::World;

Structs

Added
Filter that retrieves components of type T that have been added since the last execution of this system

ChangeTrackers
Change trackers for component T

Changed
Filter that retrieves components of type T that have been changed since the last execution of this system

Commands
A list of commands that will be run to modify a World

In

Local

Mut
Unique borrow of an entity's component

NonSend
Shared borrow of a NonSend resource

NonSendMut
Unique borrow of a NonSend resource

Or

Query
Provides scoped access to a World according to a given WorldQuery and query filter

QuerySet

QueryState

RemovedComponents

Res
Shared borrow of a Resource

ResMut
Unique borrow of a Resource

RunCriteria

State
Stack based state machine

SystemSet
A builder for describing several systems at the same time.

SystemStage
Stores and executes systems. Execution order is not defined unless explicitly specified; see SystemDescriptor documentation.

With
Filter that selects entities with a component T

WithBundle

Without
Filter that selects entities without a component T

Traits

AmbiguitySetLabel

ExclusiveSystemDescriptorCoercion

IntoChainSystem

IntoExclusiveSystem

IntoSystem

ParallelSystemDescriptorCoercion

RunCriteriaDescriptorCoercion

RunCriteriaLabel

RunCriteriaPiping

Stage

StageLabel

System
An ECS system that can be added to a Schedule

SystemLabel

Derive Macros

AmbiguitySetLabel

Bundle

RunCriteriaLabel

StageLabel

SystemLabel