Struct bevy_ecs::query::Added

pub struct Added<T>(/* private fields */);

A filter on a component that only retains results added after the system last ran.

A common use for this filter is one-time initialization.

To retain all results without filtering but still check whether they were added after the system last ran, use Ref<T>.

Examples

fn print_add_name_component(query: Query<&Name, Added<Name>>) {
    for name in &query {
        println!("Named entity created: {:?}", name)
    }
}