Struct bevy_ecs::query::AnyOf

pub struct AnyOf<T>(_);

The AnyOf query parameter fetches entities with any of the component types included in T.

Query<AnyOf<(&A, &B, &mut C)>> is equivalent to Query<(Option<&A>, Option<&B>, Option<&mut C>), Or<(With<A>, With<B>, With<C>)>>. Each of the components in T is returned as an Option, as with Option<A> queries. Entities are guaranteed to have at least one of the components in T.