Struct bevy_ecs::query::ChangeTrackers

pub struct ChangeTrackers<T: Component> { /* private fields */ }

👎Deprecated: ChangeTrackers<T> will be removed in bevy 0.11. Use bevy_ecs::prelude::Ref<T> instead.

WorldQuery that tracks changes and additions for component T.

Wraps a Component to track whether the component changed for the corresponding entities in a query since the last time the system that includes these queries ran.

If you only care about entities that changed or that got added use the Changed and Added filters instead.

Examples

fn print_moving_objects_system(query: Query<(&Name, ChangeTrackers<Transform>)>) {
    for (name, tracker) in &query {
        if tracker.is_changed() {
            println!("Entity moved: {:?}", name);
        } else {
            println!("Entity stood still: {:?}", name);
        }
    }
}

Implementations

impl<T: Component> ChangeTrackers<T>

pub fn is_added(&self) -> bool

Returns true if this component has been added since the last execution of this system.

pub fn is_changed(&self) -> bool

Returns true if this component has been changed since the last execution of this system.