bevy_ecs
Module reflect

Types that enable reflection support.

Structs

AppTypeRegistry

A Resource storing TypeRegistry for type registrations relevant to a whole app.

ReflectBundle

A struct used to operate on reflected Bundle trait of a type.

ReflectBundleFns

The raw function pointers needed to make up a ReflectBundle.

ReflectComponent

A struct used to operate on reflected Component trait of a type.

ReflectComponentFns

The raw function pointers needed to make up a ReflectComponent.

ReflectFromWorld

A struct used to operate on the reflected FromWorld trait of a type.

ReflectFromWorldFns

The raw function pointers needed to make up a ReflectFromWorld.

ReflectMapEntities

For a specific type of component, this maps any fields with values of type Entity to a new world. Since a given Entity ID is only valid for the world it came from, when performing deserialization any stored IDs need to be re-allocated in the destination world.

ReflectResource

A struct used to operate on reflected Resource of a type.

ReflectResourceFns

The raw function pointers needed to make up a ReflectResource.

Traits

ReflectCommandExt

An extension trait for EntityCommands for reflection related functions