bevy_ecs::relationship

Function clone_relationship_target

pub fn clone_relationship_target<T: RelationshipTarget>(
    component: &T,
    cloned: &mut T,
    context: &mut ComponentCloneCtx<'_, '_>,
)

The “clone behavior” for RelationshipTarget. The RelationshipTarget will be populated with the proper components when the corresponding Relationship sources of truth are inserted. Cloning the actual entities in the original RelationshipTarget would result in duplicates, so we don’t do that!

This will also queue up clones of the relationship sources if the EntityCloner is configured to spawn recursively.