bevy_ecs

Module removal_detection

Alerting events when a component is removed from an entity.

Structs

RemovedComponentEntity

Wrapper around Entity for RemovedComponents. Internally, RemovedComponents uses these as an Events<RemovedComponentEntity>.

RemovedComponentEvents

Stores the RemovedComponents event buffers for all types of component in a given World.

RemovedComponentReader

Wrapper around a ManualEventReader<RemovedComponentEntity> so that we can differentiate events between components.

RemovedComponents

A SystemParam that yields entities that had their T Component removed or have been despawned with it.

Type Aliases

RemovedIter

Iterator over entities that had a specific component removed.

RemovedIterWithId

Iterator over entities that had a specific component removed.