bevy_ecs::schedule

Enum ScheduleConfigs

pub enum ScheduleConfigs<T: Schedulable> {
    ScheduleConfig(ScheduleConfig<T>),
    Configs {
        configs: Vec<ScheduleConfigs<T>>,
        collective_conditions: Vec<BoxedCondition>,
        metadata: T::GroupMetadata,
    },
}

Single or nested configurations for Schedulables.

Variants

ScheduleConfig(ScheduleConfig<T>)

Configuration for a single Schedulable.

Configs

Configuration for a tuple of nested Configs instances.

Fields

configs: Vec<ScheduleConfigs<T>>

Configuration for each element of the tuple.

collective_conditions: Vec<BoxedCondition>

Run conditions applied to everything in the tuple.

metadata: T::GroupMetadata

Metadata to be applied to all elements in the tuple.

Implementations

impl<T: Schedulable<Metadata = GraphInfo, GroupMetadata = Chain>> ScheduleConfigs<T>

pub fn in_set_inner(&mut self, set: InternedSystemSet)

Adds a new boxed system set to the systems.

pub fn run_if_dyn(&mut self, condition: BoxedCondition)

Adds a new boxed run condition to the systems.

This is useful if you have a run condition whose concrete type is unknown. Prefer run_if for run conditions whose type is known at compile time.