bevy_ecs::schedule

Module passes

Included optional schedule build passes.

Structs

AutoInsertApplyDeferredPass

A ScheduleBuildPass that inserts ApplyDeferred systems into the schedule graph when there are Deferred in one system and there are ordering dependencies on that system. Commands is one such deferred buffer.

IgnoreDeferred

If added to a dependency edge, the edge will not be considered for auto sync point insertions.