Struct bevy_ecs::schedule::NextState

pub struct NextState<S: States>(pub Option<S>);

The next state of State<S>.

To queue a transition, just set the contained value to Some(next_state). Note that these transitions can be overridden by other systems: only the actual value of this resource at the time of apply_state_transition matters.

Tuple Fields
0: Option<S>
Implementations

impl<S: States> NextState<S>

pub fn set(&mut self, state: S)

Tentatively set a planned state transition to Some(state).