Struct bevy_ecs::schedule::Schedule

pub struct Schedule { /* private fields */ }

A collection of systems, and the metadata and executor needed to run them in a certain order under certain conditions.

Example

Here is an example of a Schedule running a “Hello world” system:

fn hello_world() { println!("Hello world!") }

fn main() {
    let mut world = World::new();
    let mut schedule = Schedule::default();
    schedule.add_system(hello_world);

    schedule.run(&mut world);
}

A schedule can also run several systems in an ordered way:

fn system_one() { println!("System 1 works!") }
fn system_two() { println!("System 2 works!") }
fn system_three() { println!("System 3 works!") }
    
fn main() {
    let mut world = World::new();
    let mut schedule = Schedule::default();
    schedule.add_systems((
        system_two,
        system_one.before(system_two),
        system_three.after(system_two),
    ));

    schedule.run(&mut world);
}

Implementations

impl Schedule

pub fn new() -> Self

Constructs an empty Schedule.

pub fn set_default_base_set(
&mut self,
default_base_set: impl SystemSet
) -> &mut Self

pub fn add_system<M>(&mut self, system: impl IntoSystemConfig<M>) -> &mut Self

Add a system to the schedule.

pub fn add_systems<M>(&mut self, systems: impl IntoSystemConfigs<M>) -> &mut Self

Add a collection of systems to the schedule.

pub fn configure_set(&mut self, set: impl IntoSystemSetConfig) -> &mut Self

Configures a system set in this schedule, adding it if it does not exist.

pub fn configure_sets(&mut self, sets: impl IntoSystemSetConfigs) -> &mut Self

Configures a collection of system sets in this schedule, adding them if they does not exist.

pub fn set_build_settings(
&mut self,
settings: ScheduleBuildSettings
) -> &mut Self

Changes miscellaneous build settings.

pub fn get_executor_kind(&self) -> ExecutorKind

Returns the schedule’s current execution strategy.

pub fn set_executor_kind(&mut self, executor: ExecutorKind) -> &mut Self

Sets the schedule’s execution strategy.

pub fn set_apply_final_buffers(&mut self, apply_final_buffers: bool) -> &mut Self

Set whether the schedule applies buffers on final time or not. This is a catchall incase a system uses commands but was not explicitly ordered after a apply_system_buffers. By default this setting is true, but may be disabled if needed.

pub fn run(&mut self, world: &mut World)

Runs all systems in this schedule on the world, using its current execution strategy.

pub fn initialize(&mut self, world: &mut World) -> Result<(), ScheduleBuildError>

Initializes any newly-added systems and conditions, rebuilds the executable schedule, and re-initializes the executor.

Moves all systems and run conditions out of the ScheduleGraph.

pub fn graph(&self) -> &ScheduleGraph

Returns the ScheduleGraph.

pub fn graph_mut(&mut self) -> &mut ScheduleGraph

Returns a mutable reference to the ScheduleGraph.

pub fn apply_system_buffers(&mut self, world: &mut World)

Directly applies any accumulated system buffers (like Commands) to the world.

Like always, system buffers are applied in the “topological sort order” of the schedule graph. As a result, buffers from one system are only guaranteed to be applied before those of other systems if there is an explicit system ordering between the two systems.

This is used in rendering to extract data from the main world, storing the data in system buffers, before applying their buffers in a different world.