bevy_ecs::schedule

Struct ScheduleGraph

pub struct ScheduleGraph {
    pub systems: Systems,
    pub system_sets: SystemSets,
    pub ambiguous_with_all: HashSet<NodeId>,
    /* private fields */
}

Metadata for a Schedule.

The order isn’t optimized; calling ScheduleGraph::build_schedule will return a SystemSchedule where the order is optimized for execution.

Fields
systems: Systems

Container of systems in the schedule.

system_sets: SystemSets

Container of system sets in the schedule.

ambiguous_with_all: HashSet<NodeId>

Nodes that are allowed to have ambiguous ordering relationship with any other systems.

Implementations

impl ScheduleGraph

pub fn new() -> Self

Creates an empty ScheduleGraph with default settings.

pub fn hierarchy(&self) -> &Dag<NodeId>

Returns the Dag of the hierarchy.

The hierarchy is a directed acyclic graph of the systems and sets, where an edge denotes that a system or set is the child of another set.

pub fn dependency(&self) -> &Dag<NodeId>

Returns the Dag of the dependencies in the schedule.

Nodes in this graph are systems and sets, and edges denote that a system or set has to run before another system or set.

pub fn conflicting_systems(&self) -> &[(SystemKey, SystemKey, Vec<ComponentId>)]

Returns the list of systems that conflict with each other, i.e. have ambiguities in their access.

If the Vec<ComponentId> is empty, the systems conflict on World access. Must be called after ScheduleGraph::build_schedule to be non-empty.

pub fn initialize(&mut self, world: &mut World)

Initializes any newly-added systems and conditions by calling System::initialize.

pub fn build_schedule( &mut self, world: &mut World, ignored_ambiguities: &BTreeSet<ComponentId>, ) -> Result<(SystemSchedule, Vec<ScheduleBuildWarning>), ScheduleBuildError>

Builds an execution-optimized SystemSchedule from the current state of the graph. Also returns any warnings that were generated during the build process.

This method also

checks for dependency or hierarchy cycles

checks for system access conflicts and reports ambiguities

impl ScheduleGraph

pub fn get_node_name(&self, id: &NodeId) -> String

Returns the name of the node with the given NodeId. Resolves anonymous sets to a string that describes their contents.

Also displays the set(s) the node is contained in if ScheduleBuildSettings::report_sets is true, and shortens system names if ScheduleBuildSettings::use_shortnames is true.

pub fn topsort_graph<N: GraphNodeId + Into<NodeId>>( &self, graph: &DiGraph<N>, report: ReportCycles, ) -> Result<Vec<N>, ScheduleBuildError>

Tries to topologically sort graph.

If the graph is acyclic, returns Ok with the list of NodeId in a valid topological order. If the graph contains cycles, returns Err with the list of strongly-connected components that contain cycles (also in a valid topological order).

Errors

If the graph contain cycles, then an error is returned.

pub fn conflicts_to_string<'a>( &'a self, ambiguities: &'a [(SystemKey, SystemKey, Vec<ComponentId>)], components: &'a Components, ) -> impl Iterator<Item = (String, String, Vec<DebugName>)> + 'a

convert conflicts to human readable format