Struct bevy_ecs::schedule::Schedules

pub struct Schedules { /* private fields */ }

Resource that stores Schedules mapped to ScheduleLabels.

Implementations

impl Schedules

pub fn new() -> Self

Constructs an empty Schedules with zero initial capacity.

pub fn insert( &mut self, label: impl ScheduleLabel, schedule: Schedule ) -> Option<Schedule>

Inserts a labeled schedule into the map.

If the map already had an entry for label, schedule is inserted, and the old schedule is returned. Otherwise, None is returned.

pub fn remove(&mut self, label: &dyn ScheduleLabel) -> Option<Schedule>

Removes the schedule corresponding to the label from the map, returning it if it existed.

pub fn remove_entry( &mut self, label: &dyn ScheduleLabel ) -> Option<(Box<dyn ScheduleLabel>, Schedule)>

Removes the (schedule, label) pair corresponding to the label from the map, returning it if it existed.

pub fn contains(&self, label: &dyn ScheduleLabel) -> bool

Does a schedule with the provided label already exist?

pub fn get(&self, label: &dyn ScheduleLabel) -> Option<&Schedule>

Returns a reference to the schedule associated with label, if it exists.

pub fn get_mut(&mut self, label: &dyn ScheduleLabel) -> Option<&mut Schedule>

Returns a mutable reference to the schedule associated with label, if it exists.

pub fn iter(&self) -> impl Iterator<Item = (&dyn ScheduleLabel, &Schedule)>

Returns an iterator over all schedules. Iteration order is undefined.

pub fn iter_mut( &mut self ) -> impl Iterator<Item = (&dyn ScheduleLabel, &mut Schedule)>

Returns an iterator over mutable references to all schedules. Iteration order is undefined.