Trait bevy_ecs::schedule::IntoSystemConfigs

pub trait IntoSystemConfigs<Marker>where
    Self: Sized,{
    // Provided methods
    fn in_set(self, set: impl SystemSet) -> SystemConfigs { ... }
    fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemConfigs { ... }
    fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemConfigs { ... }
    fn distributive_run_if<M>(
        self,
        condition: impl Condition<M> + Clone
    ) -> SystemConfigs { ... }
    fn run_if<M>(self, condition: impl Condition<M>) -> SystemConfigs { ... }
    fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemConfigs { ... }
    fn ambiguous_with_all(self) -> SystemConfigs { ... }
    fn chain(self) -> SystemConfigs { ... }
    fn on_startup(self) -> SystemConfigs { ... }
    fn in_schedule(self, _schedule: impl ScheduleLabel) -> SystemConfigs { ... }
}

Types that can convert into a SystemConfigs.

Provided Methods

fn in_set(self, set: impl SystemSet) -> SystemConfigs

Add these systems to the provided set.

fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemConfigs

Run before all systems in set.

fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemConfigs

Run after all systems in set.

fn distributive_run_if<M>( self, condition: impl Condition<M> + Clone ) -> SystemConfigs

Add a run condition to each contained system.

Each system will receive its own clone of the Condition and will only run if the Condition is true.

Each individual condition will be evaluated at most once (per schedule run), right before the corresponding system prepares to run.

This is equivalent to calling run_if on each individual system, as shown below:

schedule.add_systems((a, b).distributive_run_if(condition));
schedule.add_systems((a.run_if(condition), b.run_if(condition)));

Note

Because the conditions are evaluated separately for each system, there is no guarantee that all evaluations in a single schedule run will yield the same result. If another system is run inbetween two evaluations it could cause the result of the condition to change.

Use run_if on a SystemSet if you want to make sure that either all or none of the systems are run, or you don’t want to evaluate the run condition for each contained system separately.

fn run_if<M>(self, condition: impl Condition<M>) -> SystemConfigs

Run the systems only if the Condition is true.

The Condition will be evaluated at most once (per schedule run), the first time a system in this set prepares to run.

If this set contains more than one system, calling run_if is equivalent to adding each system to a common set and configuring the run condition on that set, as shown below:

Examples

schedule.add_systems((a, b).run_if(condition));
schedule.add_systems((a, b).in_set(C)).configure_set(C.run_if(condition));

Note

Because the condition will only be evaluated once, there is no guarantee that the condition is upheld after the first system has run. You need to make sure that no other systems that could invalidate the condition are scheduled inbetween the first and last run system.

Use distributive_run_if if you want the condition to be evaluated for each individual system, right before one is run.

fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemConfigs

Suppress warnings and errors that would result from these systems having ambiguities (conflicting access but indeterminate order) with systems in set.

fn ambiguous_with_all(self) -> SystemConfigs

Suppress warnings and errors that would result from these systems having ambiguities (conflicting access but indeterminate order) with any other system.

fn chain(self) -> SystemConfigs

Treat this collection as a sequence of systems.

Ordering constraints will be applied between the successive elements.

fn on_startup(self) -> SystemConfigs

👎Deprecated since 0.11.0: app.add_system(foo.on_startup()) has been deprecated in favor of app.add_systems(Startup, foo). Please migrate to that API.

This used to add the system to CoreSchedule::Startup. This was a shorthand for self.in_schedule(CoreSchedule::Startup).

Panics

Always panics. Please migrate to the new App::add_systems with the Startup schedule: Ex: app.add_system(foo.on_startup()) -> app.add_systems(Startup, foo)

fn in_schedule(self, _schedule: impl ScheduleLabel) -> SystemConfigs

👎Deprecated since 0.11.0: app.add_system(foo.in_schedule(SomeSchedule)) has been deprecated in favor of app.add_systems(SomeSchedule, foo). Please migrate to that API.

This used to add the system to the provided schedule.

Panics

Always panics. Please migrate to the new App::add_systems: Ex: app.add_system(foo.in_schedule(SomeSchedule)) -> app.add_systems(SomeSchedule, foo)