Trait bevy_ecs::schedule::IntoSystemSetConfig

pub trait IntoSystemSetConfig: Sized {
    // Provided methods
    fn in_set(self, set: impl SystemSet) -> SystemSetConfig { ... }
    fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfig { ... }
    fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfig { ... }
    fn run_if<M>(self, condition: impl Condition<M>) -> SystemSetConfig { ... }
    fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfig { ... }
    fn ambiguous_with_all(self) -> SystemSetConfig { ... }
    fn on_startup(self) -> SystemSetConfigs { ... }
    fn in_schedule(self, _schedule: impl ScheduleLabel) -> SystemSetConfigs { ... }
}

Types that can be converted into a SystemSetConfig.

This has been implemented for all types that implement SystemSet and boxed trait objects.

Provided Methods

fn in_set(self, set: impl SystemSet) -> SystemSetConfig

Add to the provided set.

fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfig

Run before all systems in set.

fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfig

Run after all systems in set.

fn run_if<M>(self, condition: impl Condition<M>) -> SystemSetConfig

Run the systems in this set only if the Condition is true.

The Condition will be evaluated at most once (per schedule run), the first time a system in this set prepares to run.

fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfig

Suppress warnings and errors that would result from systems in this set having ambiguities (conflicting access but indeterminate order) with systems in set.

fn ambiguous_with_all(self) -> SystemSetConfig

Suppress warnings and errors that would result from systems in this set having ambiguities (conflicting access but indeterminate order) with any other system.

fn on_startup(self) -> SystemSetConfigs

👎Deprecated since 0.11.0: app.configure_set(MySet.on_startup()) has been deprecated in favor of app.configure_set(Startup, MySet). Please migrate to that API.

This used to configure the set in the CoreSchedule::Startup schedule. This was a shorthand for self.in_schedule(CoreSchedule::Startup).

Panics

Always panics. Please migrate to the new App::configure_set with the Startup schedule: Ex: app.configure_set(MySet.on_startup()) -> app.configure_set(Startup, MySet)

fn in_schedule(self, _schedule: impl ScheduleLabel) -> SystemSetConfigs

👎Deprecated since 0.11.0: app.configure_set(MySet.in_schedule(SomeSchedule)) has been deprecated in favor of app.configure_set(SomeSchedule, MySet). Please migrate to that API.

This used to configure the set in the provided schedule.

Panics

Always panics. Please migrate to the new App::configure_set: Ex: app.configure_set(MySet.in_schedule(SomeSchedule)) -> app.configure_set(SomeSchedule, MySet)