Trait bevy_ecs::schedule::ScheduleLabel

pub trait ScheduleLabel: 'static + Send + Sync + Debug {
    // Required methods
    fn dyn_clone(&self) -> Box<dyn ScheduleLabel>;
    fn as_dyn_eq(&self) -> &dyn DynEq;
    fn dyn_hash(&self, state: &mut dyn Hasher);

    // Provided method
    fn intern(&self) -> Interned<dyn ScheduleLabel>
       where Self: Sized { ... }
}

A strongly-typed class of labels used to identify a Schedule.

Required Methods

fn dyn_clone(&self) -> Box<dyn ScheduleLabel>

Clones this ScheduleLabel.

fn as_dyn_eq(&self) -> &dyn DynEq

Casts this value to a form where it can be compared with other type-erased values.

fn dyn_hash(&self, state: &mut dyn Hasher)

Feeds this value into the given Hasher.

Provided Methods

fn intern(&self) -> Interned<dyn ScheduleLabel>
where Self: Sized,

Returns an Interned value corresponding to self.