Type Alias bevy_ecs::schedule::SystemConfigs

pub type SystemConfigs = NodeConfigs<BoxedSystem>;

A collection of SystemConfig.

Aliased Type

enum SystemConfigs {
    NodeConfig(NodeConfig<Box<dyn System<Out = (), In = ()>, Global>>),
    Configs {
        configs: Vec<NodeConfigs<Box<dyn System<Out = (), In = ()>, Global>>, Global>,
        collective_conditions: Vec<Box<dyn ReadOnlySystem<Out = bool, In = ()>, Global>, Global>,
        chained: bool,
    },
}

Variants

NodeConfig(NodeConfig<Box<dyn System<Out = (), In = ()>, Global>>)

Configuration for a single node.

Configs

Fields

configs: Vec<NodeConfigs<Box<dyn System<Out = (), In = ()>, Global>>, Global>

Configuration for each element of the tuple.

collective_conditions: Vec<Box<dyn ReadOnlySystem<Out = bool, In = ()>, Global>, Global>

Run conditions applied to everything in the tuple.

chained: bool

If true, adds before -> after ordering constraints between the successive elements.

Configuration for a tuple of nested Configs instances.

Implementations

impl<T> NodeConfigs<T>

pub fn in_set_inner(&mut self, set: InternedSystemSet)

Adds a new boxed system set to the systems.

pub fn run_if_dyn(&mut self, condition: BoxedCondition)

Adds a new boxed run condition to the systems.

This is useful if you have a run condition whose concrete type is unknown. Prefer run_if for run conditions whose type is known at compile time.