bevy_ecs

Module spawn

Entity spawning abstractions, largely focused on spawning related hierarchies of entities. See related and SpawnRelated for the best entry points into these APIs and examples of how to use them.

Structs

Spawn

A wrapper over a Bundle indicating that an entity should be spawned with that Bundle. This is intended to be used for hierarchical spawning via traits like SpawnableList and SpawnRelated.

SpawnIter

A SpawnableList that spawns entities using an iterator of a given Bundle:

SpawnOneRelated

A Bundle that:

SpawnRelatedBundle

A Bundle that:

SpawnWith

A SpawnableList that spawns entities using a FnOnce with a RelatedSpawner as an argument:

WithOneRelated

A wrapper over an Entity indicating that an entity should be added. This is intended to be used for hierarchical spawning via traits like SpawnableList and SpawnRelated.

WithRelated

A SpawnableList that links already spawned entities to the root entity via relations of type I.

Traits

SpawnRelated

RelationshipTarget methods that create a Bundle with a DynamicBundle::Effect that:

SpawnableList

A spawn-able list of changes to a given World and relative to a given Entity. This is generally used for spawning “related” entities, such as children.