bevy_ecs::spawn

Trait SpawnableList

pub trait SpawnableList<R>: Sized {
    // Required methods
    fn spawn(this: MovingPtr<'_, Self>, world: &mut World, entity: Entity);
    fn size_hint(&self) -> usize;
}

A spawn-able list of changes to a given World and relative to a given Entity. This is generally used for spawning “related” entities, such as children.

Required Methods

fn spawn(this: MovingPtr<'_, Self>, world: &mut World, entity: Entity)

Spawn this list of changes in a given World and relative to a given Entity. This is generally used for spawning “related” entities, such as children.

fn size_hint(&self) -> usize

Returns a size hint, which is used to reserve space for this list in a RelationshipTarget. This should be less than or equal to the actual size of the list. When in doubt, just use 0.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.