bevy_ecs::storage

Struct ThinColumn

pub struct ThinColumn { /* private fields */ }

Very similar to a normal Column, but with the capacities and lengths cut out for performance reasons.

This type is used by Table, because all of the capacities and lengths of the Table’s columns must match.

Like many other low-level storage types, ThinColumn has a limited and highly unsafe interface. It’s highly advised to use higher level types and their safe abstractions instead of working directly with ThinColumn.

Implementations

impl ThinColumn

pub fn with_capacity(component_info: &ComponentInfo, capacity: usize) -> Self

Create a new ThinColumn with the given capacity.

pub unsafe fn get_data_slice<T>(&self, len: usize) -> &[UnsafeCell<T>]

Get a slice to the data stored in this ThinColumn.

Safety

T must match the type of data that’s stored in this ThinColumn

len must match the actual length of this column (number of elements stored)

pub unsafe fn get_added_ticks_slice(&self, len: usize) -> &[UnsafeCell<Tick>]

Get a slice to the added ticks in this ThinColumn.

Safety

len must match the actual length of this column (number of elements stored)

pub unsafe fn get_changed_ticks_slice(&self, len: usize) -> &[UnsafeCell<Tick>]

Get a slice to the changed ticks in this ThinColumn.

Safety

len must match the actual length of this column (number of elements stored)

pub unsafe fn get_changed_by_slice( &self, len: usize, ) -> MaybeLocation<&[UnsafeCell<&'static Location<'static>>]>

Get a slice to the calling locations that last changed each value in this ThinColumn

Safety

len must match the actual length of this column (number of elements stored)