bevy_ecs::system::command

Function spawn_batch

pub fn spawn_batch<I>(bundles_iter: I) -> impl Command
where
    I: IntoIterator + Send + Sync + 'static,
    I::Item: Bundle<Effect: NoBundleEffect>,

A Command that consumes an iterator of Bundles to spawn a series of entities.

This is more efficient than spawning the entities individually.