bevy_ecs::system::entity_command

Function clone_with_opt_out

pub fn clone_with_opt_out(
    target: Entity,
    config: impl FnOnce(&mut EntityClonerBuilder<'_, OptOut>) + Send + Sync + 'static,
) -> impl EntityCommand

An EntityCommand that clones parts of an entity onto another entity, configured through EntityClonerBuilder.

This builder tries to clone every component from the source entity except for components that were explicitly denied, for example by using the deny method.

Required components are not considered by denied components and must be explicitly denied as well if desired.