Type Alias bevy_ecs::system::lifetimeless::SResMut

pub type SResMut<T> = ResMut<'static, T>;

A ResMut with 'static lifetimes.

Aliased Type

struct SResMut<T> { /* private fields */ }

Implementations

impl<'a, T: ?Sized + Resource> ResMut<'a, T>

pub fn into_inner(self) -> &'a mut T

Consume self and return a mutable reference to the contained value while marking self as “changed”.

pub fn reborrow(&mut self) -> Mut<'_, T>

Returns a Mut<> with a smaller lifetime. This is useful if you have &mut ResMut <T>, but you need a Mut<T>.

pub fn map_unchanged<U: ?Sized>( self, f: impl FnOnce(&mut T) -> &mut U ) -> Mut<'a, U>

Maps to an inner value by applying a function to the contained reference, without flagging a change.

You should never modify the argument passed to the closure – if you want to modify the data without flagging a change, consider using DetectChangesMut::bypass_change_detection to make your intent explicit.

// When run, zeroes the translation of every entity.
fn reset_positions(mut transforms: Query<&mut Transform>) {
    for transform in &mut transforms {
        // We pinky promise not to modify `t` within the closure.
        // Breaking this promise will result in logic errors, but will never cause undefined behavior.
        let mut translation = transform.map_unchanged(|t| &mut t.translation);
        // Only reset the translation if it isn't already zero;
        translation.set_if_neq(Vec2::ZERO);
    }
}

pub fn as_deref_mut(&mut self) -> Mut<'_, <T as Deref>::Target>where T: DerefMut,

Allows you access to the dereferenced value of this pointer without immediately triggering change detection.