Struct bevy_ecs::system::AdapterSystem

pub struct AdapterSystem<Func, S> { /* private fields */ }

A System that takes the output of S and transforms it by applying Func to it.

Implementations

impl<Func, S> AdapterSystem<Func, S>
where Func: Adapt<S>, S: System,

pub const fn new(func: Func, system: S, name: Cow<'static, str>) -> Self

Creates a new System that uses func to adapt system, via the Adapt trait.