bevy_ecs::system

Struct DynSystemParam

pub struct DynSystemParam<'w, 's> { /* private fields */ }

A SystemParam with a type that can be configured at runtime.

To be useful, this must be configured using a DynParamBuilder to build the system using a SystemParamBuilder.

Examples

// If the inner parameter doesn't require any special building, use `ParamBuilder`.
// Either specify the type parameter on `DynParamBuilder::new()` ...
let system = (DynParamBuilder::new::<Res<A>>(ParamBuilder),)
    .build_state(&mut world)
    .build_system(expects_res_a);

// ... or use a factory method on `ParamBuilder` that returns a specific type.
let system = (DynParamBuilder::new(ParamBuilder::resource::<A>()),)
    .build_state(&mut world)
    .build_system(expects_res_a);

fn expects_res_a(mut param: DynSystemParam) {
    // Use the `downcast` methods to retrieve the inner parameter.
    // They will return `None` if the type does not match.
    assert!(param.is::<Res<A>>());
    assert!(!param.is::<Res<B>>());
    assert!(param.downcast_mut::<Res<B>>().is_none());
    let res = param.downcast_mut::<Res<A>>().unwrap();
    // The type parameter can be left out if it can be determined from use.
    let res: Res<A> = param.downcast().unwrap();
}

let system = (
    // If the inner parameter also requires building,
    // pass the appropriate `SystemParamBuilder`.
    DynParamBuilder::new(LocalBuilder(10usize)),
    // `DynSystemParam` is just an ordinary `SystemParam`,
    // and can be combined with other parameters as usual!
    ParamBuilder::query(),
)
    .build_state(&mut world)
    .build_system(|param: DynSystemParam, query: Query<()>| {
        let local: Local<usize> = param.downcast::<Local<usize>>().unwrap();
        assert_eq!(*local, 10);
    });

Implementations

impl<'w, 's> DynSystemParam<'w, 's>

pub fn is<T: SystemParam>(&self) -> bool
where T::Item<'static, 'static>: SystemParam<Item<'w, 's> = T> + 'static,

Returns true if the inner system param is the same as T.

pub fn downcast<T: SystemParam>(self) -> Option<T>
where T::Item<'static, 'static>: SystemParam<Item<'w, 's> = T> + 'static,

Returns the inner system param if it is the correct type. This consumes the dyn param, so the returned param can have its original world and state lifetimes.

pub fn downcast_mut<'a, T: SystemParam>(&'a mut self) -> Option<T>
where T::Item<'static, 'static>: SystemParam<Item<'a, 'a> = T> + 'static,

Returns the inner system parameter if it is the correct type. This borrows the dyn param, so the returned param is only valid for the duration of that borrow.

pub fn downcast_mut_inner<'a, T: ReadOnlySystemParam>(&'a mut self) -> Option<T>
where T::Item<'static, 'static>: SystemParam<Item<'w, 'a> = T> + 'static,

Returns the inner system parameter if it is the correct type. This borrows the dyn param, so the returned param is only valid for the duration of that borrow, but since it only performs read access it can keep the original world lifetime. This can be useful with methods like Query::iter_inner() or Res::into_inner() to obtain references with the original world lifetime.