bevy_ecs::system

Struct If

pub struct If<T>(pub T);

A SystemParam that wraps another parameter and causes its system to skip instead of failing when the parameter is invalid.

Example

// This system will fail if `SomeResource` is not present.
fn fails_on_missing_resource(res: Res<SomeResource>) {}

// This system will skip without error if `SomeResource` is not present.
fn skips_on_missing_resource(res: If<Res<SomeResource>>) {
    // The inner parameter is available using `Deref`
    let some_resource: &SomeResource = &res;
}

Tuple Fields
0: T
Implementations

impl<T> If<T>

pub fn into_inner(self) -> T

Returns the inner T.

The inner value is pub, so you can also obtain it by destructuring the parameter:

fn skips_on_missing_resource(If(res): If<Res<SomeResource>>) {
    let some_resource: Res<SomeResource> = res;
}