bevy_ecs::system
Struct Local

pub struct Local<'s, T: FromWorld + Send + 'static>(/* private fields */);

A system local SystemParam.

A local may only be accessed by the system itself and is therefore not visible to other systems. If two or more systems specify the same local type each will have their own unique local. If multiple SystemParams within the same system each specify the same local type each will get their own distinct data storage.

The supplied lifetime parameter is the SystemParams 's lifetime.

Examples

fn write_to_local(mut local: Local<usize>) {
    *local = 42;
}
fn read_from_local(local: Local<usize>) -> usize {
    *local
}
let mut write_system = IntoSystem::into_system(write_to_local);
let mut read_system = IntoSystem::into_system(read_from_local);
write_system.initialize(world);
read_system.initialize(world);

assert_eq!(read_system.run((), world), 0);
write_system.run((), world);
// Note how the read local is still 0 due to the locals not being shared.
assert_eq!(read_system.run((), world), 0);

N.B. A Locals value cannot be read or written to outside of the containing system. To add configuration to a system, convert a capturing closure into the system instead:

struct Config(u32);
#[derive(Resource)]
struct MyU32Wrapper(u32);
fn reset_to_system(value: Config) -> impl FnMut(ResMut<MyU32Wrapper>) {
    move |mut val| val.0 = value.0
}

// .add_systems(reset_to_system(my_config))