bevy_ecs::system
Struct Populated

pub struct Populated<'w, 's, D: QueryData, F: QueryFilter = ()>(/* private fields */);

System parameter that works very much like Query except it always contains at least one matching entity.

This SystemParam fails validation if no matching entities exist. /// This will cause a panic, but can be configured to do nothing or warn once.

Much like Query::is_empty the worst case runtime will be O(n) where n is the number of potential matches. This can be notably expensive for queries that rely on non-archetypal filters such as Added or Changed which must individually check each query result for a match.

See Query for more details.

Implementations

impl<'w, 's, D: QueryData, F: QueryFilter> Populated<'w, 's, D, F>

pub fn into_inner(self) -> Query<'w, 's, D, F>

Returns the inner item with ownership.