Struct bevy_ecs::system::Res

pub struct Res<'w, T: Resource> { /* private fields */ }

Shared borrow of a resource.

See the World documentation to see the usage of a resource.

If you need a unique mutable borrow, use ResMut instead.

Panics

Panics when used as a SystemParameter if the resource does not exist.

Use Option<Res<T>> instead if the resource might not always exist.

Implementations

impl<'w, T: Resource> Res<'w, T>

pub fn is_added(&self) -> bool

Returns true if the resource was added after the system last ran.

pub fn is_changed(&self) -> bool

Returns true if the resource was added or mutably dereferenced after the system last ran.

pub fn into_inner(self) -> &'w T