Struct bevy_ecs::system::SystemBuilder

pub struct SystemBuilder<'w, T: SystemParam = ()> { /* private fields */ }

Builder struct used to construct state for SystemParam passed to a system.

fn my_system(res: Res<R>, query: Query<&A>, param: MyParam) {
    // ...
}

// Create a builder from the world, helper methods exist to add `SystemParam`,
// alternatively use `.param::<T>()` for any other `SystemParam` types.
let system = SystemBuilder::<()>::new(&mut world)
    .resource::<R>()
    .query::<&A>()
    .param::<MyParam>()
    .build(my_system);

// Parameters that the builder is initialised with will appear first in the arguments.
let system = SystemBuilder::<(Res<R>, Query<&A>)>::new(&mut world)
    .param::<MyParam>()
    .build(my_system);

// Parameters that implement `BuildableSystemParam` can use `.builder::<T>()` to build in place.
let system = SystemBuilder::<()>::new(&mut world)
    .resource::<R>()
    .builder::<Query<&A>>(|builder| { builder.with::<B>(); })
    .param::<MyParam>()
    .build(my_system);

world.run_system_once(system);

Implementations

impl<'w, T: SystemParam> SystemBuilder<'w, T>

pub fn new(world: &'w mut World) -> Self

Construct a new builder with the default state for T

pub fn build<F, Marker>(self, func: F) -> FunctionSystem<Marker, F>
where F: SystemParamFunction<Marker, Param = T>,

Construct the a system with the built params

pub fn state(self) -> SystemState<T>

Return the constructed SystemState

impl<'w> SystemBuilder<'w, ()>

pub fn param<T: SystemParam>(self) -> SystemBuilder<'w, (T,)>

Add T as a parameter built from the world

pub fn resource<T: Resource>(self) -> SystemBuilder<'w, (Res<'static, T>,)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (ResMut<'static, T>,)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (Local<'static, T>,)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (Query<'static, 'static, D, ()>,)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (Query<'static, 'static, D, F>,)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (T,)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam> SystemBuilder<'w, (P0,)>

pub fn param<T: SystemParam>(self) -> SystemBuilder<'w, (P0, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>(self) -> SystemBuilder<'w, (P0, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam> SystemBuilder<'w, (P0, P1)>

pub fn param<T: SystemParam>(self) -> SystemBuilder<'w, (P0, P1, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam> SystemBuilder<'w, (P0, P1, P2)>

pub fn param<T: SystemParam>(self) -> SystemBuilder<'w, (P0, P1, P2, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3)>

pub fn param<T: SystemParam>(self) -> SystemBuilder<'w, (P0, P1, P2, P3, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4)>

pub fn param<T: SystemParam>(self) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam, P9: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam, P9: SystemParam, P10: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam, P9: SystemParam, P10: SystemParam, P11: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam, P9: SystemParam, P10: SystemParam, P11: SystemParam, P12: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam, P9: SystemParam, P10: SystemParam, P11: SystemParam, P12: SystemParam, P13: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, T)>

Add T as a parameter built with the given function

impl<'w, P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam, P8: SystemParam, P9: SystemParam, P10: SystemParam, P11: SystemParam, P12: SystemParam, P13: SystemParam, P14: SystemParam> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14)>

pub fn param<T: SystemParam>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, T)>

Add T as a parameter built from the world

pub fn resource<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, Res<'static, T>)>

Helper method for reading a Resource as a param, equivalent to .param::<Res<T>>()

pub fn resource_mut<T: Resource>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, ResMut<'static, T>)>

Helper method for mutably accessing a Resource as a param, equivalent to .param::<ResMut<T>>()

pub fn local<T: Send + FromWorld>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, Local<'static, T>)>

Helper method for adding a Local as a param, equivalent to .param::<Local<T>>()

pub fn query<D: QueryData>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, Query<'static, 'static, D, ()>)>

Helper method for adding a Query as a param, equivalent to .param::<Query<D>>()

pub fn query_filtered<D: QueryData, F: QueryFilter>( self ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, Query<'static, 'static, D, F>)>

Helper method for adding a filtered Query as a param, equivalent to .param::<Query<D, F>>()

pub fn builder<T: BuildableSystemParam>( self, func: impl FnOnce(&mut T::Builder<'_>) ) -> SystemBuilder<'w, (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, T)>

Add T as a parameter built with the given function