bevy_ecs::system

Struct SystemChangeTick

pub struct SystemChangeTick { /* private fields */ }

A SystemParam that reads the previous and current change ticks of the system.

A systemโs change ticks are updated each time it runs:

last_run copies the previous value of change_tick

this_run copies the current value of World::read_change_tick

Component change ticks that are more recent than last_run will be detected by the system. Those can be read by calling last_changed on a Mut<T> or ResMut<T>.

Implementations

impl SystemChangeTick

pub fn this_run(&self) -> Tick

Returns the current World change tick seen by the system.

pub fn last_run(&self) -> Tick

Returns the World change tick seen by the system the previous time it ran.