bevy_ecs::system

Struct SystemName

pub struct SystemName(/* private fields */);

SystemParam that returns the name of the system which it is used in.

This is not a reliable identifier, it is more so useful for debugging or logging.

Examples

#[derive(SystemParam)]
struct Logger {
    system_name: SystemName,
}

impl Logger {
    fn log(&mut self, message: &str) {
        eprintln!("{}: {}", self.system_name, message);
    }
}

fn system1(mut logger: Logger) {
    // Prints: "crate_name::mod_name::system1: Hello".
    logger.log("Hello");
}

Implementations

impl SystemName

pub fn name(&self) -> DebugName

Gets the name of the system.