bevy_ecs::system

Struct WithInputFromWrapper

pub struct WithInputFromWrapper<S, T> { /* private fields */ }

Constructed in IntoSystem::with_input_from.

Implementations

impl<S, T> WithInputFromWrapper<S, T>
where for<'i> S: System<In: SystemInput<Inner<'i> = &'i mut T>>, T: Send + Sync + 'static,

pub fn new<M>(system: impl IntoSystem<S::In, S::Out, M, System = S>) -> Self

Wraps the given system.

pub fn value(&self) -> Option<&T>

Returns a reference to the input value, if it has been initialized.

pub fn value_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the input value, if it has been initialized.