Trait bevy_ecs::system::IntoPipeSystem

pub trait IntoPipeSystem<ParamA, Payload, SystemB, ParamB, Out>: IntoSystem<(), Payload, ParamA> + Sizedwhere
    SystemB: IntoSystem<Payload, Out, ParamB>,{
    fn pipe(self, system: SystemB) -> PipeSystem<Self::System, SystemB::System>;
}

An extension trait providing the IntoPipeSystem::pipe method to pass input from one system into the next.

The first system must have return type T and the second system must have In<T> as its first system parameter.

This trait is blanket implemented for all system pairs that fulfill the type requirements.

See PipeSystem.

Required Methods

fn pipe(self, system: SystemB) -> PipeSystem<Self::System, SystemB::System>

Pass the output of this system A into a second system B, creating a new compound system.