Trait bevy_ecs::system::IntoSystem

pub trait IntoSystem<In, Out, Marker>: Sized {
    type System: System<In = In, Out = Out>;

    // Required method
    fn into_system(this: Self) -> Self::System;

    // Provided method
    fn pipe<B, Final, MarkerB>(
        self,
        system: B
    ) -> PipeSystem<Self::System, B::System>
       where B: IntoSystem<Out, Final, MarkerB> { ... }
}

Conversion trait to turn something into a System.

Use this to get a system from a function. Also note that every system implements this trait as well.

Examples

use bevy_ecs::prelude::*;

fn my_system_function(a_usize_local: Local<usize>) {}

let system = IntoSystem::into_system(my_system_function);

Required Associated Types

type System: System<In = In, Out = Out>

The type of System that this instance converts into.

Required Methods

fn into_system(this: Self) -> Self::System

Turns this value into its corresponding System.

Provided Methods

fn pipe<B, Final, MarkerB>( self, system: B ) -> PipeSystem<Self::System, B::System>where B: IntoSystem<Out, Final, MarkerB>,

Pass the output of this system A into a second system B, creating a new compound system.

The second system must have In<T> as its first parameter, where T is the return type of the first system.