Trait bevy_ecs::system::SystemParamFetch

pub trait SystemParamFetch<'world, 'state>: SystemParamState {
    type Item: SystemParam<Fetch = Self>;

    unsafe fn get_param(
        state: &'state mut Self, 
        system_meta: &SystemMeta, 
        world: &'world World, 
        change_tick: u32
    ) -> Self::Item;
}

Required Associated Types

type Item: SystemParam<Fetch = Self>

Required Methods

unsafe fn get_param(
state: &'state mut Self,
system_meta: &SystemMeta,
world: &'world World,
change_tick: u32
) -> Self::Item

Safety

This call might access any of the input parameters in an unsafe way. Make sure the data access is safe in the context of the system scheduler.