Type Alias bevy_ecs::system::RunSystem

pub type RunSystem = RunSystemWithInput<()>;

The Command type for World::run_system.

This command runs systems in an exclusive and single threaded way. Running slow systems can become a bottleneck.

If the system needs an In<_> input value to run, use the RunSystemWithInput type instead.

There is no way to get the output of a system when run as a command, because the execution of the system happens later. To get the output of a system, use World::run_system or World::run_system_with_input instead of running the system as a command.

Aliased Type

struct RunSystem { /* private fields */ }

Implementations

impl RunSystem

pub fn new(system_id: SystemId) -> Self

Creates a new Command struct, which can be added to Commands