bevy_ecs::world

Module error

Contains error types returned by bevy’s schedule.

Structs

EntityDespawnError

An error that occurs when a specified Entity could not be despawned.

TryInsertBatchError

The error type returned by World::try_insert_batch and World::try_insert_batch_if_new if any of the provided entities do not exist.

TryRunScheduleError

The error type returned by World::try_run_schedule if the provided schedule does not exist.

Enums

EntityComponentError

An error that occurs when dynamically retrieving components from an entity.

EntityMutableFetchError

An error that occurs when fetching entities mutably from a world.

ResourceFetchError

An error that occurs when getting a resource of a given type in a world.