bevy_ecs::world

Struct EntityRefExcept

pub struct EntityRefExcept<'w, 's, B>
where
    B: Bundle,
{ /* private fields */ }

Provides read-only access to a single entity and all its components, save for an explicitly-enumerated set.

Implementations

impl<'w, 's, B> EntityRefExcept<'w, 's, B>
where B: Bundle,

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn get<C>(&self) -> Option<&'w C>
where C: Component,

Gets access to the component of type C for the current entity. Returns None if the component doesn’t have a component of that type or if the type is one of the excluded components.

pub fn get_ref<C>(&self) -> Option<Ref<'w, C>>
where C: Component,

Gets access to the component of type C for the current entity, including change detection information. Returns None if the component doesn’t have a component of that type or if the type is one of the excluded components.

pub fn spawned_by(&self) -> MaybeLocation

Returns the source code location from which this entity has been spawned.

pub fn spawn_tick(&self) -> Tick

Returns the Tick at which this entity has been spawned.

pub fn get_by_id(&self, component_id: ComponentId) -> Option<Ptr<'w>>

Gets the component of the given ComponentId from the entity.

You should prefer to use the typed API Self::get where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityRefExcept::get, this returns a raw pointer to the component, which is only valid while the EntityRefExcept is alive.

pub fn contains<T: Component>(&self) -> bool

Returns true if the current entity has a component of type T. Otherwise, this returns false.

Notes

If you do not know the concrete type of a component, consider using Self::contains_id or Self::contains_type_id.

pub fn contains_id(&self, component_id: ComponentId) -> bool

Returns true if the current entity has a component identified by component_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you know the component’s TypeId but not its ComponentId, consider using Self::contains_type_id.

pub fn contains_type_id(&self, type_id: TypeId) -> bool

Returns true if the current entity has a component with the type identified by type_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you have a ComponentId instead of a TypeId, consider using Self::contains_id.

pub fn get_change_ticks<T: Component>(&self) -> Option<ComponentTicks>

Retrieves the change ticks for the given component. This can be useful for implementing change detection in custom runtimes.

pub fn get_change_ticks_by_id( &self, component_id: ComponentId, ) -> Option<ComponentTicks>

Retrieves the change ticks for the given ComponentId. This can be useful for implementing change detection in custom runtimes.

You should prefer to use the typed API Self::get_change_ticks where possible and only use this in cases where the actual component types are not known at compile time.