bevy_ecs::world
Struct FilteredEntityRef

pub struct FilteredEntityRef<'w> { /* private fields */ }

Provides read-only access to a single entity and some of its components defined by the contained Access.

To define the access when used as a QueryData, use a QueryBuilder or QueryParamBuilder. The FilteredEntityRef must be the entire QueryData, and not nested inside a tuple with other data.

// This gives the `FilteredEntityRef` access to `&A`.
let mut query = QueryBuilder::<FilteredEntityRef>::new(&mut world)
    .data::<&A>()
    .build();

let filtered_entity: FilteredEntityRef = query.single(&mut world);
let component: &A = filtered_entity.get().unwrap();

// Here `FilteredEntityRef` is nested in a tuple, so it does not have access to `&A`.
let mut query = QueryBuilder::<(Entity, FilteredEntityRef)>::new(&mut world)
    .data::<&A>()
    .build();

let (_, filtered_entity) = query.single(&mut world);
assert!(filtered_entity.get::<A>().is_none());

Implementations

impl<'w> FilteredEntityRef<'w>

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn location(&self) -> EntityLocation

Gets metadata indicating the location where the current entity is stored.

pub fn archetype(&self) -> &Archetype

Returns the archetype that the current entity belongs to.

pub fn access(&self) -> &Access<ComponentId>

Returns a reference to the underlying Access.

pub fn contains<T: Component>(&self) -> bool

Returns true if the current entity has a component of type T. Otherwise, this returns false.

Notes

If you do not know the concrete type of a component, consider using Self::contains_id or Self::contains_type_id.

pub fn contains_id(&self, component_id: ComponentId) -> bool

Returns true if the current entity has a component identified by component_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you know the component’s TypeId but not its ComponentId, consider using Self::contains_type_id.

pub fn contains_type_id(&self, type_id: TypeId) -> bool

Returns true if the current entity has a component with the type identified by type_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you have a ComponentId instead of a TypeId, consider using Self::contains_id.

pub fn get<T: Component>(&self) -> Option<&'w T>

Gets access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

pub fn get_ref<T: Component>(&self) -> Option<Ref<'w, T>>

Gets access to the component of type T for the current entity, including change detection information as a Ref.

Returns None if the entity does not have a component of type T.

pub fn get_change_ticks<T: Component>(&self) -> Option<ComponentTicks>

Retrieves the change ticks for the given component. This can be useful for implementing change detection in custom runtimes.

pub fn get_change_ticks_by_id( &self, component_id: ComponentId, ) -> Option<ComponentTicks>

Retrieves the change ticks for the given ComponentId. This can be useful for implementing change detection in custom runtimes.

You should prefer to use the typed API Self::get_change_ticks where possible and only use this in cases where the actual component types are not known at compile time.

pub fn get_by_id(&self, component_id: ComponentId) -> Option<Ptr<'w>>

Gets the component of the given ComponentId from the entity.

You should prefer to use the typed API Self::get where possible and only use this in cases where the actual component types are not known at compile time.

Unlike FilteredEntityRef::get, this returns a raw pointer to the component, which is only valid while the FilteredEntityRef is alive.