bevy_ecs::world

Struct OccupiedComponentEntry

pub struct OccupiedComponentEntry<'w, 'a, T: Component> { /* private fields */ }

A view into an occupied entry in a EntityWorldMut. It is part of the OccupiedComponentEntry enum.

The contained entity must have the component type parameter if we have this struct.

Implementations

impl<'w, 'a, T: Component> OccupiedComponentEntry<'w, 'a, T>

pub fn get(&self) -> &T

Gets a reference to the component in the entry.

Examples

#[derive(Component, Default, Clone, Copy, Debug, PartialEq)]
struct Comp(u32);

let mut entity = world.spawn(Comp(5));

if let ComponentEntry::Occupied(o) = entity.entry::<Comp>() {
    assert_eq!(o.get().0, 5);
}

pub fn insert(&mut self, component: T)

Replaces the component of the entry.

Examples

#[derive(Component, Default, Clone, Copy, Debug, PartialEq)]
struct Comp(u32);

let mut entity = world.spawn(Comp(5));

if let ComponentEntry::Occupied(mut o) = entity.entry::<Comp>() {
    o.insert(Comp(10));
}

assert_eq!(world.query::<&Comp>().single(&world).unwrap().0, 10);

pub fn take(self) -> T

Removes the component from the entry and returns it.

Examples

#[derive(Component, Default, Clone, Copy, Debug, PartialEq)]
struct Comp(u32);

let mut entity = world.spawn(Comp(5));

if let ComponentEntry::Occupied(o) = entity.entry::<Comp>() {
    assert_eq!(o.take(), Comp(5));
}

assert_eq!(world.query::<&Comp>().iter(&world).len(), 0);

impl<'w, 'a, T: Component<Mutability = Mutable>> OccupiedComponentEntry<'w, 'a, T>

pub fn get_mut(&mut self) -> Mut<'_, T>

Gets a mutable reference to the component in the entry.

If you need a reference to the OccupiedComponentEntry which may outlive the destruction of the OccupiedComponentEntry value, see into_mut.

Examples

#[derive(Component, Default, Clone, Copy, Debug, PartialEq)]
struct Comp(u32);

let mut entity = world.spawn(Comp(5));

if let ComponentEntry::Occupied(mut o) = entity.entry::<Comp>() {
    o.get_mut().0 += 10;
    assert_eq!(o.get().0, 15);

    // We can use the same Entry multiple times.
    o.get_mut().0 += 2
}

assert_eq!(world.query::<&Comp>().single(&world).unwrap().0, 17);

pub fn into_mut(self) -> Mut<'a, T>

Converts the OccupiedComponentEntry into a mutable reference to the value in the entry with a lifetime bound to the EntityWorldMut.

If you need multiple references to the OccupiedComponentEntry, see get_mut.

Examples

#[derive(Component, Default, Clone, Copy, Debug, PartialEq)]
struct Comp(u32);

let mut entity = world.spawn(Comp(5));

if let ComponentEntry::Occupied(o) = entity.entry::<Comp>() {
    o.into_mut().0 += 10;
}

assert_eq!(world.query::<&Comp>().single(&world).unwrap().0, 15);