bevy_ecs

Module bundle

Types for handling Bundles.

This module contains the Bundle trait and some other helper types.

Structs

BundleId

For a specific World, this stores a unique value identifying a type of a registered Bundle.

BundleInfo

Stores metadata associated with a specific type of Bundle for a given World.

Bundles

Metadata for bundles. Stores a BundleInfo for each type of Bundle in a given world.

Traits

Bundle

The Bundle trait enables insertion and removal of Components from an entity.

DynamicBundle

The parts from Bundle that don’t require statically knowing the components of the bundle.

Derive Macros

Bundle