bevy_ecs::change_detection

Trait DetectChanges

pub trait DetectChanges {
    // Required methods
    fn is_added(&self) -> bool;
    fn is_changed(&self) -> bool;
    fn last_changed(&self) -> Tick;
    fn added(&self) -> Tick;
    fn changed_by(&self) -> MaybeLocation;
}

Types that can read change detection information. This change detection is controlled by DetectChangesMut types such as ResMut.

Example

Using types that implement DetectChanges, such as Res, provide a way to query if a value has been mutated in another system.

use bevy_ecs::prelude::*;

#[derive(Resource)]
struct MyResource(u32);

fn my_system(mut resource: Res<MyResource>) {
    if resource.is_changed() {
        println!("My component was mutated!");
    }
}

Required Methods

fn is_added(&self) -> bool

Returns true if this value was added after the system last ran.

fn is_changed(&self) -> bool

Returns true if this value was added or mutably dereferenced either since the last time the system ran or, if the system never ran, since the beginning of the program.

To check if the value was mutably dereferenced only, use this.is_changed() && !this.is_added().

fn last_changed(&self) -> Tick

Returns the change tick recording the time this data was most recently changed.

Note that components and resources are also marked as changed upon insertion.

For comparison, the previous change tick of a system can be read using the SystemChangeTick SystemParam.

fn added(&self) -> Tick

Returns the change tick recording the time this data was added.

fn changed_by(&self) -> MaybeLocation

The location that last caused this to change.