bevy_ecs

Module event

Event handling types.

Structs

EventId

An EventId uniquely identifies an event stored in a specific World.

EventIterator

An iterator that yields any unread events from an EventReader or ManualEventReader.

EventIteratorWithId

An iterator that yields any unread events (and their IDs) from an EventReader or ManualEventReader.

EventParIter

A parallel iterator over Events.

EventReader

Reads events of type T in order and tracks which events have already been read.

EventRegistry

A registry of all of the Events in the World, used by event_update_system to update all of the events.

EventWriter

Sends events of type T.

Events

An event collection that represents the events that occurred within the last two Events::update calls. Events can be written to using an EventWriter and are typically cheaply read using an EventReader.

ManualEventReader

Stores the state for an EventReader.

SendBatchIds

Iterator over sent EventIds from a batch.

Enums

ShouldUpdateEvents

Controls whether or not the events in an EventRegistry should be updated.

Traits

Event

Something that “happens” and might be read / observed by app logic.

Functions

event_update_condition

A run condition for event_update_system.

event_update_system

A system that calls Events::update on all registered Events in the world.

signal_event_update_system

Signals the event_update_system to run after FixedUpdate systems.

Derive Macros

Event