bevy_ecs

Module observer

Observers are a push-based tool for responding to Events. The Observer component holds a System that runs whenever a matching Event is triggered.

See Event and Observer for in-depth documentation and usage examples.

Structs

CachedComponentObservers

Collection of ObserverRunner for Observer registered to a particular event targeted at a specific component.

CachedObservers

Collection of ObserverRunner for Observer registered to a particular event.

ObservedBy

Tracks a list of entity observers for the Entity ObservedBy is added to.

Observer

An Observer system. Add this Component to an Entity to turn it into an “observer”.

ObserverDescriptor

Store information about what an Observer observes.

Observers

An internal lookup table tracking all of the observers in the world.

On

A system parameter used by an observer to process events. See Observer and Event for examples.

TriggerContext

Metadata about a specific Event that triggered an observer.

Type Aliases

ObserverMap

Map between an observer entity and its ObserverRunner

ObserverRunner

Type for function that is run when an observer is triggered.

TriggerDeprecated

Deprecated in favor of On.