bevy_ecs::query
Struct Access

pub struct Access<T: SparseSetIndex> { /* private fields */ }

Tracks read and write access to specific elements in a collection.

Used internally to ensure soundness during system initialization and execution. See the is_compatible and get_conflicts functions.

Implementations

impl<T: SparseSetIndex> Access<T>

pub const fn new() -> Self

Creates an empty Access collection.

pub fn add_component_read(&mut self, index: T)

Adds access to the component given by index.

pub fn add_component_write(&mut self, index: T)

Adds exclusive access to the component given by index.

pub fn add_resource_read(&mut self, index: T)

Adds access to the resource given by index.

pub fn add_resource_write(&mut self, index: T)

Adds exclusive access to the resource given by index.

pub fn remove_component_read(&mut self, index: T)

Removes both read and write access to the component given by index.

Because this method corresponds to the set difference operator ∖, it can create complicated logical formulas that you should verify correctness of. For example, A ∪ (B ∖ A) isn’t equivalent to (A ∪ B) ∖ A, so you can’t replace a call to remove_component_read followed by a call to extend with a call to extend followed by a call to remove_component_read.

pub fn remove_component_write(&mut self, index: T)

Removes write access to the component given by index.

Because this method corresponds to the set difference operator ∖, it can create complicated logical formulas that you should verify correctness of. For example, A ∪ (B ∖ A) isn’t equivalent to (A ∪ B) ∖ A, so you can’t replace a call to remove_component_write followed by a call to extend with a call to extend followed by a call to remove_component_write.

pub fn add_archetypal(&mut self, index: T)

Adds an archetypal (indirect) access to the component given by index.

This is for components whose values are not accessed (and thus will never cause conflicts), but whose presence in an archetype may affect query results.

Currently, this is only used for Has<T>.

pub fn has_component_read(&self, index: T) -> bool

Returns true if this can access the component given by index.

pub fn has_any_component_read(&self) -> bool

Returns true if this can access any component.

pub fn has_component_write(&self, index: T) -> bool

Returns true if this can exclusively access the component given by index.

pub fn has_any_component_write(&self) -> bool

Returns true if this accesses any component mutably.

pub fn has_resource_read(&self, index: T) -> bool

Returns true if this can access the resource given by index.

pub fn has_any_resource_read(&self) -> bool

Returns true if this can access any resource.

pub fn has_resource_write(&self, index: T) -> bool

Returns true if this can exclusively access the resource given by index.

pub fn has_any_resource_write(&self) -> bool

Returns true if this accesses any resource mutably.

pub fn has_archetypal(&self, index: T) -> bool

Returns true if this has an archetypal (indirect) access to the component given by index.

This is a component whose value is not accessed (and thus will never cause conflicts), but whose presence in an archetype may affect query results.

Currently, this is only used for Has<T>.

pub fn read_all_components(&mut self)

Sets this as having access to all components (i.e. EntityRef).

pub fn write_all_components(&mut self)

Sets this as having mutable access to all components (i.e. EntityMut).

pub fn read_all_resources(&mut self)

Sets this as having access to all resources (i.e. &World).

pub fn write_all_resources(&mut self)

Sets this as having mutable access to all resources (i.e. &mut World).

pub fn read_all(&mut self)

Sets this as having access to all indexed elements (i.e. &World).

pub fn write_all(&mut self)

Sets this as having mutable access to all indexed elements (i.e. &mut World).

pub fn has_read_all_components(&self) -> bool

Returns true if this has access to all components (i.e. EntityRef).

pub fn has_write_all_components(&self) -> bool

Returns true if this has write access to all components (i.e. EntityMut).

pub fn has_read_all_resources(&self) -> bool

Returns true if this has access to all resources (i.e. EntityRef).

pub fn has_write_all_resources(&self) -> bool

Returns true if this has write access to all resources (i.e. EntityMut).

pub fn has_read_all(&self) -> bool

Returns true if this has access to all indexed elements (i.e. &World).

pub fn has_write_all(&self) -> bool

Returns true if this has write access to all indexed elements (i.e. &mut World).

pub fn clear_writes(&mut self)

Removes all writes.

pub fn clear(&mut self)

Removes all accesses.

pub fn extend(&mut self, other: &Access<T>)

Adds all access from other.

pub fn is_components_compatible(&self, other: &Access<T>) -> bool

Returns true if the access and other can be active at the same time, only looking at their component access.

Access instances are incompatible if one can write an element that the other can read or write.

pub fn is_resources_compatible(&self, other: &Access<T>) -> bool

Returns true if the access and other can be active at the same time, only looking at their resource access.

Access instances are incompatible if one can write an element that the other can read or write.

pub fn is_compatible(&self, other: &Access<T>) -> bool

Returns true if the access and other can be active at the same time.

Access instances are incompatible if one can write an element that the other can read or write.

pub fn is_subset_components(&self, other: &Access<T>) -> bool

Returns true if the set’s component access is a subset of another, i.e. other’s component access contains at least all the values in self.

pub fn is_subset_resources(&self, other: &Access<T>) -> bool

Returns true if the set’s resource access is a subset of another, i.e. other’s resource access contains at least all the values in self.

pub fn is_subset(&self, other: &Access<T>) -> bool

Returns true if the set is a subset of another, i.e. other contains at least all the values in self.

pub fn get_conflicts(&self, other: &Access<T>) -> AccessConflicts

Returns a vector of elements that the access and other cannot access at the same time.

pub fn resource_reads_and_writes(&self) -> impl Iterator<Item = T> + '_

Returns the indices of the resources this has access to.

pub fn resource_reads(&self) -> impl Iterator<Item = T> + '_

Returns the indices of the resources this has non-exclusive access to.

pub fn resource_writes(&self) -> impl Iterator<Item = T> + '_

Returns the indices of the resources this has exclusive access to.

pub fn archetypal(&self) -> impl Iterator<Item = T> + '_

Returns the indices of the components that this has an archetypal access to.

These are components whose values are not accessed (and thus will never cause conflicts), but whose presence in an archetype may affect query results.

Currently, this is only used for Has<T>.