Trait bevy_ecs::query::ArchetypeFilter

pub trait ArchetypeFilter: QueryFilter { }

A marker trait to indicate that the filter works at an archetype level.

This is needed to implement ExactSizeIterator for QueryIter that contains archetype-level filters.

The trait must only be implemented for filters where its corresponding QueryFilter::IS_ARCHETYPAL is true. As such, only the With and Without filters can implement the trait. Tuples and Or filters are automatically implemented with the trait only if its containing types also implement the same trait.

Added and Changed works with entities, and therefore are not archetypal. As such they do not implement ArchetypeFilter.

Object Safety

This trait is not object safe.