bevy_ecs::removal_detection
Struct RemovedComponents

pub struct RemovedComponents<'w, 's, T: Component> { /* private fields */ }

A SystemParam that yields entities that had their T Component removed or have been despawned with it.

This acts effectively the same as an EventReader.

Note that this does not allow you to see which data existed before removal. If you need this, you will need to track the component data value on your own, using a regularly scheduled system that requests Query<(Entity, &T), Changed<T>> and stores the data somewhere safe to later cross-reference.

If you are using bevy_ecs as a standalone crate, note that the RemovedComponents list will not be automatically cleared for you, and will need to be manually flushed using World::clear_trackers.

For users of bevy and bevy_app, World::clear_trackers is automatically called by bevy_app::App::update and bevy_app::SubApp::update. For the main world, this is delayed until after all SubApps have run.

Examples

Basic usage:

fn react_on_removal(mut removed: RemovedComponents<MyComponent>) {
    removed.read().for_each(|removed_entity| println!("{:?}", removed_entity));
}

Implementations

impl<'w, 's, T: Component> RemovedComponents<'w, 's, T>

pub fn reader(&self) -> &EventCursor<RemovedComponentEntity>

Fetch underlying EventCursor.

pub fn reader_mut(&mut self) -> &mut EventCursor<RemovedComponentEntity>

Fetch underlying EventCursor mutably.

pub fn events(&self) -> Option<&Events<RemovedComponentEntity>>

Fetch underlying Events.

pub fn reader_mut_with_events( &mut self, ) -> Option<(&mut RemovedComponentReader<T>, &Events<RemovedComponentEntity>)>

Destructures to get a mutable reference to the EventCursor and a reference to Events.

This is necessary since Rust can’t detect destructuring through methods and most usecases of the reader uses the Events as well.

pub fn read(&mut self) -> RemovedIter<'_>

Iterates over the events this RemovedComponents has not seen yet. This updates the RemovedComponents’s event counter, which means subsequent event reads will not include events that happened before now.

pub fn read_with_id(&mut self) -> RemovedIterWithId<'_>

Like read, except also returning the EventId of the events.

pub fn len(&self) -> usize

Determines the number of removal events available to be read from this RemovedComponents without consuming any.

pub fn is_empty(&self) -> bool

Returns true if there are no events available to read.

pub fn clear(&mut self)

Consumes all available events.

This means these events will not appear in calls to RemovedComponents::read() or RemovedComponents::read_with_id() and RemovedComponents::is_empty() will return true.