bevy_ecs::schedule

Module common_conditions

A collection of run conditions that may be useful in any bevy app.

Functions

any_component_removed

Generates a Condition-satisfying closure that returns true if there are any entity with a component of the given type removed.

any_with_component

A Condition-satisfying system that returns true if there are any entities with the given component type.

not

Generates a Condition that inverses the result of passed one.

on_event

Generates a Condition-satisfying closure that returns true if there are any new events of the given type since it was last called.

resource_added

A Condition-satisfying system that returns true if the resource of the given type has been added since the condition was last checked.

resource_changed

A Condition-satisfying system that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_changed_or_removed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_equals

Generates a Condition-satisfying closure that returns true if the resource is equal to value.

resource_exists

A Condition-satisfying system that returns true if the resource exists.

resource_exists_and_changed

A Condition-satisfying system that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_exists_and_equals

Generates a Condition-satisfying closure that returns true if the resource exists and is equal to value.

resource_removed

Generates a Condition-satisfying closure that returns true if the resource of the given type has been removed since the condition was last checked.

run_once

Generates a Condition-satisfying closure that returns true if the first time the condition is run and false every time after