bevy_ecs::schedule

Module common_conditions

A collection of run conditions that may be useful in any bevy app.

Functions

any_component_removed

A SystemCondition-satisfying system that returns true if there are any entity with a component of the given type removed.

any_match_filter

A SystemCondition-satisfying system that returns true if there are any entities that match the given QueryFilter.

any_with_component

A SystemCondition-satisfying system that returns true if there are any entities with the given component type.

condition_changed

Generates a SystemCondition that returns true when the passed one changes.

condition_changed_to

Generates a SystemCondition that returns true when the result of the passed one went from false to true since the last time this was called.

not

Generates a SystemCondition that inverses the result of passed one.

on_eventDeprecated

A SystemCondition-satisfying system that returns true if there are any new events of the given type since it was last called.

on_message

A SystemCondition-satisfying system that returns true if there are any new messages of the given type since it was last called.

resource_added

A SystemCondition-satisfying system that returns true if the resource of the given type has been added since the condition was last checked.

resource_changed

A SystemCondition-satisfying system that returns true if the resource of the given type has been added or mutably dereferenced since the condition was last checked.

resource_changed_or_removed

A SystemCondition-satisfying system that returns true if the resource of the given type has been added, removed or mutably dereferenced since the condition was last checked.

resource_equals

Generates a SystemCondition-satisfying closure that returns true if the resource is equal to value.

resource_exists

A SystemCondition-satisfying system that returns true if the resource exists.

resource_exists_and_changed

A SystemCondition-satisfying system that returns true if the resource of the given type has been added or mutably dereferenced since the condition was last checked.

resource_exists_and_equals

Generates a SystemCondition-satisfying closure that returns true if the resource exists and is equal to value.

resource_removed

A SystemCondition-satisfying system that returns true if the resource of the given type has been removed since the condition was last checked.

run_once

A SystemCondition-satisfying system that returns true on the first time the condition is run and false every time after.