Trait bevy_ecs::schedule::IntoSystemSetConfigs

pub trait IntoSystemSetConfigswhere
    Self: Sized,{
    // Provided methods
    fn in_set(self, set: impl SystemSet) -> SystemSetConfigs { ... }
    fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs { ... }
    fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs { ... }
    fn run_if<M>(self, condition: impl Condition<M>) -> SystemSetConfigs { ... }
    fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs { ... }
    fn ambiguous_with_all(self) -> SystemSetConfigs { ... }
    fn chain(self) -> SystemSetConfigs { ... }
}

Types that can convert into a SystemSetConfigs.

Provided Methods

fn in_set(self, set: impl SystemSet) -> SystemSetConfigs

Add these system sets to the provided set.

fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs

Run before all systems in set.

fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs

Run after all systems in set.

fn run_if<M>(self, condition: impl Condition<M>) -> SystemSetConfigs

Run the systems in this set(s) only if the Condition is true.

The Condition will be evaluated at most once (per schedule run), the first time a system in this set(s) prepares to run.

fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs

Suppress warnings and errors that would result from systems in these sets having ambiguities (conflicting access but indeterminate order) with systems in set.

fn ambiguous_with_all(self) -> SystemSetConfigs

Suppress warnings and errors that would result from systems in these sets having ambiguities (conflicting access but indeterminate order) with any other system.

fn chain(self) -> SystemSetConfigs

Treat this collection as a sequence of system sets.

Ordering constraints will be applied between the successive elements.