Trait bevy_ecs::schedule::IntoSystemSetConfigs

pub trait IntoSystemSetConfigswhere
    Self: Sized,{
    // Provided methods
    fn in_set(self, set: impl SystemSet) -> SystemSetConfigs { ... }
    fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs { ... }
    fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs { ... }
    fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs { ... }
    fn ambiguous_with_all(self) -> SystemSetConfigs { ... }
    fn chain(self) -> SystemSetConfigs { ... }
    fn on_startup(self) -> SystemSetConfigs { ... }
    fn in_schedule(self, _schedule: impl ScheduleLabel) -> SystemSetConfigs { ... }
}

Types that can convert into a SystemSetConfigs.

Provided Methods

fn in_set(self, set: impl SystemSet) -> SystemSetConfigs

Add these system sets to the provided set.

fn before<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs

Run before all systems in set.

fn after<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs

Run after all systems in set.

fn ambiguous_with<M>(self, set: impl IntoSystemSet<M>) -> SystemSetConfigs

Suppress warnings and errors that would result from systems in these sets having ambiguities (conflicting access but indeterminate order) with systems in set.

fn ambiguous_with_all(self) -> SystemSetConfigs

Suppress warnings and errors that would result from systems in these sets having ambiguities (conflicting access but indeterminate order) with any other system.

fn chain(self) -> SystemSetConfigs

Treat this collection as a sequence of system sets.

Ordering constraints will be applied between the successive elements.

fn on_startup(self) -> SystemSetConfigs

👎Deprecated since 0.11.0: app.configure_sets((A, B).on_startup()) has been deprecated in favor of app.configure_sets(Startup, (A, B)). Please migrate to that API.

This used to configure the sets in the CoreSchedule::Startup schedule. This was a shorthand for self.in_schedule(CoreSchedule::Startup).

Panics

Always panics. Please migrate to the new App::configure_sets with the Startup schedule: Ex: app.configure_sets((A, B).on_startup()) -> app.configure_sets(Startup, (A, B))

fn in_schedule(self, _schedule: impl ScheduleLabel) -> SystemSetConfigs

👎Deprecated since 0.11.0: app.configure_sets((A, B).in_schedule(SomeSchedule)) has been deprecated in favor of app.configure_sets(SomeSchedule, (A, B)). Please migrate to that API.

This used to configure the sets in the provided schedule.

Panics

Always panics. Please migrate to the new App::configure_set: Ex: app.configure_sets((A, B).in_schedule(SomeSchedule)) -> app.configure_sets(SomeSchedule, (A, B))