bevy_ecs::system

Module lifetimeless

Contains type aliases for built-in SystemParams with 'static lifetimes. This makes it more convenient to refer to these types in contexts where explicit lifetime annotations are required.

Note that this is entirely safe and tracks lifetimes correctly. This purely exists for convenience.

You can’t instantiate a static SystemParam, you’ll always end up with Res<'w, T>, ResMut<'w, T> or &'w T bound to the lifetime of the provided &'w World.

Type Aliases

Read

A shorthand for writing &'static T.

SCommands

Commands with 'static lifetimes.

SQuery

A Query with 'static lifetimes.

SRes

A Res with 'static lifetimes.

SResMut

A ResMut with 'static lifetimes.

Write

A shorthand for writing &'static mut T.