bevy_ecs::system
Struct ExclusiveFunctionSystem

pub struct ExclusiveFunctionSystem<Marker, F>
where
    F: ExclusiveSystemParamFunction<Marker>,
{ /* private fields */ }

A function system that runs with exclusive World access.

You get this by calling IntoSystem::into_system on a function that only accepts ExclusiveSystemParams.

ExclusiveFunctionSystem must be .initialized before they can be run.

Implementations

impl<Marker, F> ExclusiveFunctionSystem<Marker, F>
where F: ExclusiveSystemParamFunction<Marker>,

pub fn with_name(self, new_name: impl Into<Cow<'static, str>>) -> Self

Return this system with a new name.

Useful to give closure systems more readable and unique names for debugging and tracing.