bevy_ecs::system
Struct ParamSetBuilder

pub struct ParamSetBuilder<T>(pub T);

A SystemParamBuilder for a ParamSet.

To build a ParamSet with a tuple of system parameters, pass a tuple of matching SystemParamBuilders. To build a ParamSet with a Vec of system parameters, pass a Vec of matching SystemParamBuilders.

Examples

let system = (ParamSetBuilder((
    QueryParamBuilder::new(|builder| {
        builder.with::<Enemy>();
    }),
    QueryParamBuilder::new(|builder| {
        builder.with::<Ally>();
    }),
    ParamBuilder,
)),)
    .build_state(&mut world)
    .build_system(buildable_system_with_tuple);

fn buildable_system_with_tuple(
    mut set: ParamSet<(Query<&mut Health>, Query<&mut Health>, &World)>,
) {
    // The first parameter is built from the first builder,
    // so this will iterate over enemies.
    for mut health in set.p0().iter_mut() {}
    // And the second parameter is built from the second builder,
    // so this will iterate over allies.
    for mut health in set.p1().iter_mut() {}
    // Parameters that don't need special building can use `ParamBuilder`.
    let entities = set.p2().entities();
}

let system = (ParamSetBuilder(vec![
    QueryParamBuilder::new_box(|builder| {
        builder.with::<Enemy>();
    }),
    QueryParamBuilder::new_box(|builder| {
        builder.with::<Ally>();
    }),
]),)
    .build_state(&mut world)
    .build_system(buildable_system_with_vec);

fn buildable_system_with_vec(mut set: ParamSet<Vec<Query<&mut Health>>>) {
    // As with tuples, the first parameter is built from the first builder,
    // so this will iterate over enemies.
    for mut health in set.get_mut(0).iter_mut() {}
    // And the second parameter is built from the second builder,
    // so this will iterate over allies.
    for mut health in set.get_mut(1).iter_mut() {}
    // You can iterate over the parameters either by index,
    // or using the `for_each` method.
    set.for_each(|mut query| for mut health in query.iter_mut() {});
}

Tuple Fields
0: T