bevy_ecs::system

Struct Query

pub struct Query<'world, 'state, D: QueryData, F: QueryFilter = ()> { /* private fields */ }

A system parameter that provides selective access to the Component data stored in a World.

Queries enable systems to access entity identifiers and components without requiring direct access to the World. Its iterators and getter methods return query items, which are types containing data related to an entity.

Query is a generic data structure that accepts two type parameters:

D (query data): The type of data fetched by the query, which will be returned as the query item. Only entities that match the requested data will generate an item. Must implement the QueryData trait.

F (query filter): An optional set of conditions that determine whether query items should be kept or discarded. This defaults to unit, which means no additional filters will be applied. Must implement the QueryFilter trait.

Similar parameters

Query has few sibling SystemParams, which perform additional validation:

Single - Exactly one matching query item.

Option<Single> - Zero or one matching query item.

Populated - At least one matching query item.

These parameters will prevent systems from running if their requirements are not met.

System parameter declaration

A query should always be declared as a system parameter. This section shows the most common idioms involving the declaration of Query.

Component access

You can fetch an entity’s component by specifying a reference to that component in the query’s data parameter:

// A component can be accessed by a shared reference...
fn immutable_query(query: Query<&ComponentA>) {
    // ...
}

// ...or by a mutable reference.
fn mutable_query(query: Query<&mut ComponentA>) {
    // ...
}

Note that components need to be behind a reference (& or &mut), or the query will not compile:

// This needs to be `&ComponentA` or `&mut ComponentA` in order to compile.
fn invalid_query(query: Query<ComponentA>) {
    // ...
}

Query filtering

Setting the query filter type parameter will ensure that each query item satisfies the given condition:

// `ComponentA` data will be accessed, but only for entities that also contain `ComponentB`.
fn filtered_query(query: Query<&ComponentA, With<ComponentB>>) {
    // ...
}

Note that the filter is With<ComponentB>, not With<&ComponentB>. Unlike query data, With does not require components to be behind a reference.

QueryData or QueryFilter tuples

Using tuples, each Query type parameter can contain multiple elements.

In the following example two components are accessed simultaneously, and the query items are filtered on two conditions:

fn complex_query(
    query: Query<(&mut ComponentA, &ComponentB), (With<ComponentC>, Without<ComponentD>)>
) {
    // ...
}

Note that this currently only works on tuples with 15 or fewer items. You may nest tuples to get around this limit:

fn nested_query(
    query: Query<(&ComponentA, &ComponentB, (&mut ComponentC, &mut ComponentD))>
) {
    // ...
}

Entity identifier access

You can access Entity, the entity identifier, by including it in the query data parameter:

fn entity_id_query(query: Query<(Entity, &ComponentA)>) {
    // ...
}

Be aware that Entity is not a component, so it does not need to be behind a reference.

Optional component access

A component can be made optional by wrapping it into an Option. In the following example, a query item will still be generated even if the queried entity does not contain ComponentB. When this is the case, Option<&ComponentB>’s corresponding value will be None.

// Queried items must contain `ComponentA`. If they also contain `ComponentB`, its value will
// be fetched as well.
fn optional_component_query(query: Query<(&ComponentA, Option<&ComponentB>)>) {
    // ...
}

Optional components can hurt performance in some cases, so please read the performance section to learn more about them. Additionally, if you need to declare several optional components, you may be interested in using AnyOf.

Disjoint queries

A system cannot contain two queries that break Rust’s mutability rules, or else it will panic when initialized. This can often be fixed with the Without filter, which makes the queries disjoint.

In the following example, the two queries can mutably access the same &mut Health component if an entity has both the Player and Enemy components. Bevy will catch this and panic, however, instead of breaking Rust’s mutability rules:

fn randomize_health(
    player_query: Query<&mut Health, With<Player>>,
    enemy_query: Query<&mut Health, With<Enemy>>,
) {
    // ...
}

Adding a Without filter will disjoint the queries. In the following example, any entity that has both the Player and Enemy components will be excluded from both queries:

fn randomize_health(
    player_query: Query<&mut Health, (With<Player>, Without<Enemy>)>,
    enemy_query: Query<&mut Health, (With<Enemy>, Without<Player>)>,
) {
    // ...
}

An alternative solution to this problem would be to wrap the conflicting queries in ParamSet.

Whole Entity Access

EntityRef can be used in a query to gain read-only access to all components of an entity. This is useful when dynamically fetching components instead of baking them into the query type.

fn all_components_query(query: Query<(EntityRef, &ComponentA)>) {
    // ...
}

As EntityRef can read any component on an entity, a query using it will conflict with any mutable component access.

// `EntityRef` provides read access to *all* components on an entity. When combined with
// `&mut ComponentA` in the same query, it creates a conflict because `EntityRef` could read
// `&ComponentA` while `&mut ComponentA` attempts to modify it - violating Rust's borrowing
// rules.
fn invalid_query(query: Query<(EntityRef, &mut ComponentA)>) {
    // ...
}

It is strongly advised to couple EntityRef queries with the use of either With / Without filters or ParamSets. Not only does this improve the performance and parallelization of the system, but it enables systems to gain mutable access to other components:

// The first query only reads entities that have `ComponentA`, while the second query only
// modifies entities that *don't* have `ComponentA`. Because neither query will access the same
// entity, this system does not conflict.
fn disjoint_query(
    query_a: Query<EntityRef, With<ComponentA>>,
    query_b: Query<&mut ComponentB, Without<ComponentA>>,
) {
    // ...
}

The fundamental rule: EntityRef’s ability to read all components means it can never coexist with mutable access. With / Without filters can guarantee this by keeping the queries on completely separate entities.

Accessing query items

The following table summarizes the behavior of safe methods that can be used to get query items:

Query methods	Effect

iter[_mut]	Returns an iterator over all query items.

iter[_mut]().for_each(),
par_iter[_mut]	Runs a specified function for each query item.

iter_many[_unique][_mut]	Iterates over query items that match a list of entities.

iter_combinations[_mut]	Iterates over all combinations of query items.

single[_mut]	Returns a single query item if only one exists.

get[_mut]	Returns the query item for a specified entity.

get_many[_unique][_mut]	Returns all query items that match a list of entities.

There are two methods for each type of query operation: immutable and mutable (ending with _mut). When using immutable methods, the query items returned are of type ROQueryItem, a read-only version of the query item. In this circumstance, every mutable reference in the query fetch type parameter is substituted by a shared reference.

Performance

Creating a Query is a low-cost constant operation. Iterating it, on the other hand, fetches data from the world and generates items, which can have a significant computational cost.

Two systems cannot be executed in parallel if both access the same component type where at least one of the accesses is mutable. Because of this, it is recommended for queries to only fetch mutable access to components when necessary, since immutable access can be parallelized.

Query filters (With / Without) can improve performance because they narrow the kinds of entities that can be fetched. Systems that access fewer kinds of entities are more likely to be parallelized by the scheduler.

On the other hand, be careful using optional components (Option<&ComponentA>) and EntityRef because they broaden the amount of entities kinds that can be accessed. This is especially true of a query that only fetches optional components or EntityRef, as the query would iterate over all entities in the world.

There are two types of component storage types: Table and SparseSet. Table offers fast iteration speeds, but slower insertion and removal speeds. SparseSet is the opposite: it offers fast component insertion and removal speeds, but slower iteration speeds.

The following table compares the computational complexity of the various methods and operations, where:

n is the number of entities that match the query.

r is the number of elements in a combination.

k is the number of involved entities in the operation.

a is the number of archetypes in the world.

C is the binomial coefficient, used to count combinations. nCr is read as “n choose r” and is equivalent to the number of distinct unordered subsets of r elements that can be taken from a set of n elements.

Query operation	Computational complexity

iter[_mut]	O(n)

iter[_mut]().for_each(),
par_iter[_mut]	O(n)

iter_many[_mut]	O(k)

iter_combinations[_mut]	O(nCr)

single[_mut]	O(a)

get[_mut]	O(1)

get_many	O(k)

get_many_mut	O(k2)

Archetype-based filtering (With, Without, Or)	O(a)

Change detection filtering (Added, Changed, Spawned)	O(a + n)

Iterator::for_each

The for_each methods appear to be generally faster than for-loops when run on worlds with high archetype fragmentation, and may enable additional optimizations like autovectorization. It is strongly advised to only use Iterator::for_each if it tangibly improves performance. Always profile or benchmark before and after the change!

fn system(query: Query<&ComponentA>) {
    // This may result in better performance...
    query.iter().for_each(|component| {
        // ...
    });

    // ...than this. Always benchmark to validate the difference!
    for component in query.iter() {
        // ...
    }
}

Implementations

impl<'w, 's, D: QueryData, F: QueryFilter> Query<'w, 's, D, F>

pub fn related<R: Relationship>(&'w self, entity: Entity) -> Option<Entity>
where <D as QueryData>::ReadOnly: QueryData<Item<'w, 's> = &'w R>,

If the given entity contains the R Relationship component, returns the target entity of that relationship.

pub fn relationship_sources<S: RelationshipTarget>( &'w self, entity: Entity, ) -> impl Iterator<Item = Entity> + 'w
where <D as QueryData>::ReadOnly: QueryData<Item<'w, 's> = &'w S>,

If the given entity contains the S RelationshipTarget component, returns the source entities stored on that component.

pub fn root_ancestor<R: Relationship>(&'w self, entity: Entity) -> Entity
where <D as QueryData>::ReadOnly: QueryData<Item<'w, 's> = &'w R>,

Recursively walks up the tree defined by the given R Relationship until there are no more related entities, returning the “root entity” of the relationship hierarchy.

Warning

For relationship graphs that contain loops, this could loop infinitely. If your relationship is not a tree (like Bevy’s hierarchy), be sure to stop if you encounter a duplicate entity.

pub fn iter_leaves<S: RelationshipTarget>( &'w self, entity: Entity, ) -> impl Iterator<Item = Entity> + use<'w, 's, S, D, F>
where <D as QueryData>::ReadOnly: QueryData<Item<'w, 's> = &'w S>, SourceIter<'w, S>: DoubleEndedIterator,

Iterates all “leaf entities” as defined by the RelationshipTarget hierarchy.

Warning

For relationship graphs that contain loops, this could loop infinitely. If your relationship is not a tree (like Bevy’s hierarchy), be sure to stop if you encounter a duplicate entity.

pub fn iter_siblings<R: Relationship>( &'w self, entity: Entity, ) -> impl Iterator<Item = Entity> + 'w
where D::ReadOnly: QueryData<Item<'w, 's> = (Option<&'w R>, Option<&'w R::RelationshipTarget>)>,

Iterates all sibling entities that also have the R Relationship with the same target entity.

pub fn iter_descendants<S: RelationshipTarget>( &'w self, entity: Entity, ) -> DescendantIter<'w, 's, D, F, S>
where D::ReadOnly: QueryData<Item<'w, 's> = &'w S>,

Iterates all descendant entities as defined by the given entity’s RelationshipTarget and their recursive RelationshipTarget.

Warning

For relationship graphs that contain loops, this could loop infinitely. If your relationship is not a tree (like Bevy’s hierarchy), be sure to stop if you encounter a duplicate entity.

pub fn iter_descendants_depth_first<S: RelationshipTarget>( &'w self, entity: Entity, ) -> DescendantDepthFirstIter<'w, 's, D, F, S>
where D::ReadOnly: QueryData<Item<'w, 's> = &'w S>, SourceIter<'w, S>: DoubleEndedIterator,

Iterates all descendant entities as defined by the given entity’s RelationshipTarget and their recursive RelationshipTarget in depth-first order.

Warning

For relationship graphs that contain loops, this could loop infinitely. If your relationship is not a tree (like Bevy’s hierarchy), be sure to stop if you encounter a duplicate entity.

pub fn iter_ancestors<R: Relationship>( &'w self, entity: Entity, ) -> AncestorIter<'w, 's, D, F, R>
where D::ReadOnly: QueryData<Item<'w, 's> = &'w R>,

Iterates all ancestors of the given entity as defined by the R Relationship.

Warning

For relationship graphs that contain loops, this could loop infinitely. If your relationship is not a tree (like Bevy’s hierarchy), be sure to stop if you encounter a duplicate entity.

impl<'w, 's, D: QueryData, F: QueryFilter> Query<'w, 's, D, F>

pub fn as_readonly(&self) -> Query<'_, 's, D::ReadOnly, F>

Returns another Query from this that fetches the read-only version of the query items.

For example, Query<(&mut D1, &D2, &mut D3), With<F>> will become Query<(&D1, &D2, &D3), With<F>>. This can be useful when working around the borrow checker, or reusing functionality between systems via functions that accept query types.

See also

into_readonly for a version that consumes the Query to return one with the full 'world lifetime.

pub fn into_readonly(self) -> Query<'w, 's, D::ReadOnly, F>

Returns another Query from this that fetches the read-only version of the query items.

For example, Query<(&mut D1, &D2, &mut D3), With<F>> will become Query<(&D1, &D2, &D3), With<F>>. This can be useful when working around the borrow checker, or reusing functionality between systems via functions that accept query types.

See also

as_readonly for a version that borrows the Query instead of consuming it.

pub fn reborrow(&mut self) -> Query<'_, 's, D, F>

Returns a new Query reborrowing the access from this one. The current query will be unusable while the new one exists.

Example

For example this allows to call other methods or other systems that require an owned Query without completely giving up ownership of it.

fn helper_system(query: Query<&ComponentA>) { /* ... */}

fn system(mut query: Query<&ComponentA>) {
    helper_system(query.reborrow());
    // Can still use query here:
    for component in &query {
        // ...
    }
}

pub unsafe fn reborrow_unsafe(&self) -> Query<'_, 's, D, F>

Returns a new Query reborrowing the access from this one. The current query will still be usable while the new one exists, but must not be used in a way that violates aliasing.

Safety

This function makes it possible to violate Rust’s aliasing guarantees. You must make sure this call does not result in a mutable or shared reference to a component with a mutable reference.

See also

reborrow for the safe versions.

pub fn iter(&self) -> QueryIter<'_, 's, D::ReadOnly, F>

Returns an Iterator over the read-only query items.

This iterator is always guaranteed to return results from each matching entity once and only once. Iteration order is not guaranteed.

Example

Here, the report_names_system iterates over the Player component of every entity that contains it:

fn report_names_system(query: Query<&Player>) {
    for player in &query {
        println!("Say hello to {}!", player.name);
    }
}

See also

iter_mut for mutable query items.

pub fn iter_mut(&mut self) -> QueryIter<'_, 's, D, F>

Returns an Iterator over the query items.

This iterator is always guaranteed to return results from each matching entity once and only once. Iteration order is not guaranteed.

Example

Here, the gravity_system updates the Velocity component of every entity that contains it:

fn gravity_system(mut query: Query<&mut Velocity>) {
    const DELTA: f32 = 1.0 / 60.0;
    for mut velocity in &mut query {
        velocity.y -= 9.8 * DELTA;
    }
}

See also

iter for read-only query items.

pub fn iter_combinations<const K: usize>( &self, ) -> QueryCombinationIter<'_, 's, D::ReadOnly, F, K>

Returns a QueryCombinationIter over all combinations of K read-only query items without repetition.

This iterator is always guaranteed to return results from each unique pair of matching entities. Iteration order is not guaranteed.

Example

fn some_system(query: Query<&ComponentA>) {
    for [a1, a2] in query.iter_combinations() {
        // ...
    }
}

See also

iter_combinations_mut for mutable query item combinations.

iter_combinations_inner for mutable query item combinations with the full 'world lifetime.

pub fn iter_combinations_mut<const K: usize>( &mut self, ) -> QueryCombinationIter<'_, 's, D, F, K>

Returns a QueryCombinationIter over all combinations of K query items without repetition.

This iterator is always guaranteed to return results from each unique pair of matching entities. Iteration order is not guaranteed.

Example

fn some_system(mut query: Query<&mut ComponentA>) {
    let mut combinations = query.iter_combinations_mut();
    while let Some([mut a1, mut a2]) = combinations.fetch_next() {
        // mutably access components data
    }
}

See also

iter_combinations for read-only query item combinations.

iter_combinations_inner for mutable query item combinations with the full 'world lifetime.

pub fn iter_combinations_inner<const K: usize>( self, ) -> QueryCombinationIter<'w, 's, D, F, K>

Returns a QueryCombinationIter over all combinations of K query items without repetition. This consumes the Query to return results with the actual “inner” world lifetime.

This iterator is always guaranteed to return results from each unique pair of matching entities. Iteration order is not guaranteed.

Example

fn some_system(query: Query<&mut ComponentA>) {
    let mut combinations = query.iter_combinations_inner();
    while let Some([mut a1, mut a2]) = combinations.fetch_next() {
        // mutably access components data
    }
}

See also

iter_combinations for read-only query item combinations.

iter_combinations_mut for mutable query item combinations.

pub fn iter_many<EntityList: IntoIterator<Item: EntityEquivalent>>( &self, entities: EntityList, ) -> QueryManyIter<'_, 's, D::ReadOnly, F, EntityList::IntoIter>

Returns an Iterator over the read-only query items generated from an Entity list.

Items are returned in the order of the list of entities, and may not be unique if the input doesn’t guarantee uniqueness. Entities that don’t match the query are skipped.

Example

// A component containing an entity list.
#[derive(Component)]
struct Friends {
    list: Vec<Entity>,
}

fn system(
    friends_query: Query<&Friends>,
    counter_query: Query<&Counter>,
) {
    for friends in &friends_query {
        for counter in counter_query.iter_many(&friends.list) {
            println!("Friend's counter: {}", counter.value);
        }
    }
}

See also

iter_many_mut to get mutable query items.

iter_many_inner to get mutable query items with the full 'world lifetime.

pub fn iter_many_mut<EntityList: IntoIterator<Item: EntityEquivalent>>( &mut self, entities: EntityList, ) -> QueryManyIter<'_, 's, D, F, EntityList::IntoIter>

Returns an iterator over the query items generated from an Entity list.

Items are returned in the order of the list of entities, and may not be unique if the input doesn’t guarantee uniqueness. Entities that don’t match the query are skipped.

Examples

#[derive(Component)]
struct Counter {
    value: i32
}

#[derive(Component)]
struct Friends {
    list: Vec<Entity>,
}

fn system(
    friends_query: Query<&Friends>,
    mut counter_query: Query<&mut Counter>,
) {
    for friends in &friends_query {
        let mut iter = counter_query.iter_many_mut(&friends.list);
        while let Some(mut counter) = iter.fetch_next() {
            println!("Friend's counter: {}", counter.value);
            counter.value += 1;
        }
    }
}

See also

iter_many to get read-only query items.

iter_many_inner to get mutable query items with the full 'world lifetime.

pub fn iter_many_inner<EntityList: IntoIterator<Item: EntityEquivalent>>( self, entities: EntityList, ) -> QueryManyIter<'w, 's, D, F, EntityList::IntoIter>

Returns an iterator over the query items generated from an Entity list. This consumes the Query to return results with the actual “inner” world lifetime.

Items are returned in the order of the list of entities, and may not be unique if the input doesn’t guarantee uniqueness. Entities that don’t match the query are skipped.

See also

iter_many to get read-only query items.

iter_many_mut to get mutable query items.

pub fn iter_many_unique<EntityList: EntitySet>( &self, entities: EntityList, ) -> QueryManyUniqueIter<'_, 's, D::ReadOnly, F, EntityList::IntoIter>

Returns an Iterator over the unique read-only query items generated from an EntitySet.

Items are returned in the order of the list of entities. Entities that don’t match the query are skipped.

Example

// `Friends` ensures that it only lists unique entities.
#[derive(Component)]
struct Friends {
    unique_list: Vec<Entity>,
}

impl<'a> IntoIterator for &'a Friends {

    type Item = &'a Entity;
    type IntoIter = UniqueEntityIter<slice::Iter<'a, Entity>>;

    fn into_iter(self) -> Self::IntoIter {
        // SAFETY: `Friends` ensures that it unique_list contains only unique entities.
       unsafe { UniqueEntityIter::from_iterator_unchecked(self.unique_list.iter()) }
    }
}

fn system(
    friends_query: Query<&Friends>,
    counter_query: Query<&Counter>,
) {
    for friends in &friends_query {
        for counter in counter_query.iter_many_unique(friends) {
            println!("Friend's counter: {:?}", counter.value);
        }
    }
}

See also

iter_many_unique_mut to get mutable query items.

iter_many_unique_inner to get with the actual “inner” world lifetime.

pub fn iter_many_unique_mut<EntityList: EntitySet>( &mut self, entities: EntityList, ) -> QueryManyUniqueIter<'_, 's, D, F, EntityList::IntoIter>

Returns an iterator over the unique query items generated from an EntitySet.

Items are returned in the order of the list of entities. Entities that don’t match the query are skipped.

Examples

#[derive(Component)]
struct Counter {
    value: i32
}

// `Friends` ensures that it only lists unique entities.
#[derive(Component)]
struct Friends {
    unique_list: Vec<Entity>,
}

impl<'a> IntoIterator for &'a Friends {
    type Item = &'a Entity;
    type IntoIter = UniqueEntityIter<slice::Iter<'a, Entity>>;

    fn into_iter(self) -> Self::IntoIter {
        // SAFETY: `Friends` ensures that it unique_list contains only unique entities.
        unsafe { UniqueEntityIter::from_iterator_unchecked(self.unique_list.iter()) }
    }
}

fn system(
    friends_query: Query<&Friends>,
    mut counter_query: Query<&mut Counter>,
) {
    for friends in &friends_query {
        for mut counter in counter_query.iter_many_unique_mut(friends) {
            println!("Friend's counter: {:?}", counter.value);
            counter.value += 1;
        }
    }
}

See also

iter_many_unique to get read-only query items.

iter_many_unique_inner to get with the actual “inner” world lifetime.

pub fn iter_many_unique_inner<EntityList: EntitySet>( self, entities: EntityList, ) -> QueryManyUniqueIter<'w, 's, D, F, EntityList::IntoIter>

Returns an iterator over the unique query items generated from an EntitySet. This consumes the Query to return results with the actual “inner” world lifetime.

Items are returned in the order of the list of entities. Entities that don’t match the query are skipped.

Examples

#[derive(Component)]
struct Counter {
    value: i32
}

// `Friends` ensures that it only lists unique entities.
#[derive(Component)]
struct Friends {
    unique_list: Vec<Entity>,
}

impl<'a> IntoIterator for &'a Friends {
    type Item = &'a Entity;
    type IntoIter = UniqueEntityIter<slice::Iter<'a, Entity>>;

    fn into_iter(self) -> Self::IntoIter {
        // SAFETY: `Friends` ensures that it unique_list contains only unique entities.
        unsafe { UniqueEntityIter::from_iterator_unchecked(self.unique_list.iter()) }
    }
}

fn system(
    friends_query: Query<&Friends>,
    mut counter_query: Query<&mut Counter>,
) {
    let friends = friends_query.single().unwrap();
    for mut counter in counter_query.iter_many_unique_inner(friends) {
        println!("Friend's counter: {:?}", counter.value);
        counter.value += 1;
    }
}

See also

iter_many_unique to get read-only query items.

iter_many_unique_mut to get mutable query items.

pub unsafe fn iter_unsafe(&self) -> QueryIter<'_, 's, D, F>

Returns an Iterator over the query items.

This iterator is always guaranteed to return results from each matching entity once and only once. Iteration order is not guaranteed.

Safety

This function makes it possible to violate Rust’s aliasing guarantees. You must make sure this call does not result in multiple mutable references to the same component.

See also

iter and iter_mut for the safe versions.

pub unsafe fn iter_combinations_unsafe<const K: usize>( &self, ) -> QueryCombinationIter<'_, 's, D, F, K>

Iterates over all possible combinations of K query items without repetition.

This iterator is always guaranteed to return results from each unique pair of matching entities. Iteration order is not guaranteed.

Safety

This allows aliased mutability. You must make sure this call does not result in multiple mutable references to the same component.

See also

iter_combinations and iter_combinations_mut for the safe versions.

pub unsafe fn iter_many_unsafe<EntityList: IntoIterator<Item: EntityEquivalent>>( &self, entities: EntityList, ) -> QueryManyIter<'_, 's, D, F, EntityList::IntoIter>

Returns an Iterator over the query items generated from an Entity list.

Items are returned in the order of the list of entities, and may not be unique if the input doesnn’t guarantee uniqueness. Entities that don’t match the query are skipped.

Safety

This allows aliased mutability and does not check for entity uniqueness. You must make sure this call does not result in multiple mutable references to the same component. Particular care must be taken when collecting the data (rather than iterating over it one item at a time) such as via Iterator::collect.

See also

iter_many_mut to safely access the query items.

pub unsafe fn iter_many_unique_unsafe<EntityList: EntitySet>( &self, entities: EntityList, ) -> QueryManyUniqueIter<'_, 's, D, F, EntityList::IntoIter>

Returns an Iterator over the unique query items generated from an Entity list.

Items are returned in the order of the list of entities. Entities that don’t match the query are skipped.

Safety

This allows aliased mutability. You must make sure this call does not result in multiple mutable references to the same component.

See also

iter_many_unique to get read-only query items.

iter_many_unique_mut to get mutable query items.

iter_many_unique_inner to get with the actual “inner” world lifetime.

pub fn par_iter(&self) -> QueryParIter<'_, 's, D::ReadOnly, F>

Returns a parallel iterator over the query results for the given World.

This parallel iterator is always guaranteed to return results from each matching entity once and only once. Iteration order and thread assignment is not guaranteed.

If the multithreaded feature is disabled, iterating with this operates identically to Iterator::for_each on QueryIter.

This can only be called for read-only queries, see par_iter_mut for write-queries.

Note that you must use the for_each method to iterate over the results, see par_iter_mut for an example.

pub fn par_iter_mut(&mut self) -> QueryParIter<'_, 's, D, F>

Returns a parallel iterator over the query results for the given World.

This parallel iterator is always guaranteed to return results from each matching entity once and only once. Iteration order and thread assignment is not guaranteed.

If the multithreaded feature is disabled, iterating with this operates identically to Iterator::for_each on QueryIter.

This can only be called for mutable queries, see par_iter for read-only-queries.

Example

Here, the gravity_system updates the Velocity component of every entity that contains it:

fn gravity_system(mut query: Query<&mut Velocity>) {
    const DELTA: f32 = 1.0 / 60.0;
    query.par_iter_mut().for_each(|mut velocity| {
        velocity.y -= 9.8 * DELTA;
    });
}

pub fn par_iter_inner(self) -> QueryParIter<'w, 's, D, F>

Returns a parallel iterator over the query results for the given World. This consumes the Query to return results with the actual “inner” world lifetime.

This parallel iterator is always guaranteed to return results from each matching entity once and only once. Iteration order and thread assignment is not guaranteed.

If the multithreaded feature is disabled, iterating with this operates identically to Iterator::for_each on QueryIter.

Example

Here, the gravity_system updates the Velocity component of every entity that contains it:

fn gravity_system(query: Query<&mut Velocity>) {
    const DELTA: f32 = 1.0 / 60.0;
    query.par_iter_inner().for_each(|mut velocity| {
        velocity.y -= 9.8 * DELTA;
    });
}

pub fn par_iter_many<EntityList: IntoIterator<Item: EntityEquivalent>>( &self, entities: EntityList, ) -> QueryParManyIter<'_, 's, D::ReadOnly, F, EntityList::Item>

Returns a parallel iterator over the read-only query items generated from an Entity list.

Entities that don’t match the query are skipped. Iteration order and thread assignment is not guaranteed.

If the multithreaded feature is disabled, iterating with this operates identically to Iterator::for_each on QueryManyIter.

This can only be called for read-only queries. To avoid potential aliasing, there is no par_iter_many_mut equivalent. See par_iter_many_unique_mut for an alternative using EntitySet.

Note that you must use the for_each method to iterate over the results, see par_iter_mut for an example.

pub fn par_iter_many_unique<EntityList: EntitySet<Item: Sync>>( &self, entities: EntityList, ) -> QueryParManyUniqueIter<'_, 's, D::ReadOnly, F, EntityList::Item>

Returns a parallel iterator over the unique read-only query items generated from an EntitySet.

Entities that don’t match the query are skipped. Iteration order and thread assignment is not guaranteed.

If the multithreaded feature is disabled, iterating with this operates identically to Iterator::for_each on QueryManyUniqueIter.

This can only be called for read-only queries, see par_iter_many_unique_mut for write-queries.

Note that you must use the for_each method to iterate over the results, see par_iter_mut for an example.

pub fn par_iter_many_unique_mut<EntityList: EntitySet<Item: Sync>>( &mut self, entities: EntityList, ) -> QueryParManyUniqueIter<'_, 's, D, F, EntityList::Item>

Returns a parallel iterator over the unique query items generated from an EntitySet.

Entities that don’t match the query are skipped. Iteration order and thread assignment is not guaranteed.

If the multithreaded feature is disabled, iterating with this operates identically to Iterator::for_each on QueryManyUniqueIter.

This can only be called for mutable queries, see par_iter_many_unique for read-only-queries.

Note that you must use the for_each method to iterate over the results, see par_iter_mut for an example.

pub fn get( &self, entity: Entity, ) -> Result<ROQueryItem<'_, 's, D>, QueryEntityError>

Returns the read-only query item for the given Entity.

In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

This is always guaranteed to run in O(1) time.

Example

Here, get is used to retrieve the exact query item of the entity specified by the SelectedCharacter resource.

fn print_selected_character_name_system(
       query: Query<&Character>,
       selection: Res<SelectedCharacter>
)
{
    if let Ok(selected_character) = query.get(selection.entity) {
        println!("{}", selected_character.name);
    }
}

See also

get_mut to get a mutable query item.

pub fn get_many<const N: usize>( &self, entities: [Entity; N], ) -> Result<[ROQueryItem<'_, 's, D>; N], QueryEntityError>

Returns the read-only query items for the given array of Entity.

The returned query items are in the same order as the input. In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead. The elements of the array do not need to be unique, unlike get_many_mut.

Examples

use bevy_ecs::prelude::*;
use bevy_ecs::query::QueryEntityError;

#[derive(Component, PartialEq, Debug)]
struct A(usize);

let mut world = World::new();
let entity_vec: Vec<Entity> = (0..3).map(|i| world.spawn(A(i)).id()).collect();
let entities: [Entity; 3] = entity_vec.try_into().unwrap();

world.spawn(A(73));

let mut query_state = world.query::<&A>();
let query = query_state.query(&world);

let component_values = query.get_many(entities).unwrap();

assert_eq!(component_values, [&A(0), &A(1), &A(2)]);

let wrong_entity = Entity::from_raw_u32(365).unwrap();

assert_eq!(
    match query.get_many([wrong_entity]).unwrap_err() {
        QueryEntityError::EntityDoesNotExist(error) => error.entity,
        _ => panic!(),
    },
    wrong_entity
);

See also

get_many_mut to get mutable query items.

get_many_unique to only handle unique inputs.

pub fn get_many_unique<const N: usize>( &self, entities: UniqueEntityArray<N>, ) -> Result<[ROQueryItem<'_, 's, D>; N], QueryEntityError>

Returns the read-only query items for the given UniqueEntityArray.

The returned query items are in the same order as the input. In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

Examples

use bevy_ecs::{prelude::*, query::QueryEntityError, entity::{EntitySetIterator, UniqueEntityArray, UniqueEntityVec}};

#[derive(Component, PartialEq, Debug)]
struct A(usize);

let mut world = World::new();
let entity_set: UniqueEntityVec = world.spawn_batch((0..3).map(A)).collect_set();
let entity_set: UniqueEntityArray<3> = entity_set.try_into().unwrap();

world.spawn(A(73));

let mut query_state = world.query::<&A>();
let query = query_state.query(&world);

let component_values = query.get_many_unique(entity_set).unwrap();

assert_eq!(component_values, [&A(0), &A(1), &A(2)]);

let wrong_entity = Entity::from_raw_u32(365).unwrap();

assert_eq!(
    match query.get_many_unique(UniqueEntityArray::from([wrong_entity])).unwrap_err() {
        QueryEntityError::EntityDoesNotExist(error) => error.entity,
        _ => panic!(),
    },
    wrong_entity
);

See also

get_many_unique_mut to get mutable query items.

get_many to handle inputs with duplicates.

pub fn get_mut( &mut self, entity: Entity, ) -> Result<D::Item<'_, 's>, QueryEntityError>

Returns the query item for the given Entity.

In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

This is always guaranteed to run in O(1) time.

Example

Here, get_mut is used to retrieve the exact query item of the entity specified by the PoisonedCharacter resource.

fn poison_system(mut query: Query<&mut Health>, poisoned: Res<PoisonedCharacter>) {
    if let Ok(mut health) = query.get_mut(poisoned.character_id) {
        health.0 -= 1;
    }
}

See also

get to get a read-only query item.

pub fn get_inner( self, entity: Entity, ) -> Result<D::Item<'w, 's>, QueryEntityError>

Returns the query item for the given Entity. This consumes the Query to return results with the actual “inner” world lifetime.

In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

This is always guaranteed to run in O(1) time.

See also

get_mut to get the item using a mutable borrow of the Query.

pub fn get_many_mut<const N: usize>( &mut self, entities: [Entity; N], ) -> Result<[D::Item<'_, 's>; N], QueryEntityError>

Returns the query items for the given array of Entity.

The returned query items are in the same order as the input. In case of a nonexisting entity, duplicate entities or mismatched component, a QueryEntityError is returned instead.

Examples

use bevy_ecs::prelude::*;
use bevy_ecs::query::QueryEntityError;

#[derive(Component, PartialEq, Debug)]
struct A(usize);

let mut world = World::new();

let entities: Vec<Entity> = (0..3).map(|i| world.spawn(A(i)).id()).collect();
let entities: [Entity; 3] = entities.try_into().unwrap();

world.spawn(A(73));
let wrong_entity = Entity::from_raw_u32(57).unwrap();
let invalid_entity = world.spawn_empty().id();

let mut query_state = world.query::<&mut A>();
let mut query = query_state.query_mut(&mut world);

let mut mutable_component_values = query.get_many_mut(entities).unwrap();

for mut a in &mut mutable_component_values {
    a.0 += 5;
}

let component_values = query.get_many(entities).unwrap();

assert_eq!(component_values, [&A(5), &A(6), &A(7)]);

assert_eq!(
    match query
        .get_many_mut([wrong_entity])
        .unwrap_err()
    {
        QueryEntityError::EntityDoesNotExist(error) => error.entity,
        _ => panic!(),
    },
    wrong_entity
);
assert_eq!(
    match query
        .get_many_mut([invalid_entity])
        .unwrap_err()
    {
        QueryEntityError::QueryDoesNotMatch(entity, _) => entity,
        _ => panic!(),
    },
    invalid_entity
);
assert_eq!(
    query
        .get_many_mut([entities[0], entities[0]])
        .unwrap_err(),
    QueryEntityError::AliasedMutability(entities[0])
);

See also

get_many to get read-only query items without checking for duplicate entities.

pub fn get_many_unique_mut<const N: usize>( &mut self, entities: UniqueEntityArray<N>, ) -> Result<[D::Item<'_, 's>; N], QueryEntityError>

Returns the query items for the given UniqueEntityArray.

The returned query items are in the same order as the input. In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

Examples

use bevy_ecs::{prelude::*, query::QueryEntityError, entity::{EntitySetIterator, UniqueEntityArray, UniqueEntityVec}};

#[derive(Component, PartialEq, Debug)]
struct A(usize);

let mut world = World::new();

let entity_set: UniqueEntityVec = world.spawn_batch((0..3).map(A)).collect_set();
let entity_set: UniqueEntityArray<3> = entity_set.try_into().unwrap();

world.spawn(A(73));
let wrong_entity = Entity::from_raw_u32(57).unwrap();
let invalid_entity = world.spawn_empty().id();

let mut query_state = world.query::<&mut A>();
let mut query = query_state.query_mut(&mut world);

let mut mutable_component_values = query.get_many_unique_mut(entity_set).unwrap();

for mut a in &mut mutable_component_values {
    a.0 += 5;
}

let component_values = query.get_many_unique(entity_set).unwrap();

assert_eq!(component_values, [&A(5), &A(6), &A(7)]);

assert_eq!(
    match query
        .get_many_unique_mut(UniqueEntityArray::from([wrong_entity]))
        .unwrap_err()
    {
        QueryEntityError::EntityDoesNotExist(error) => error.entity,
        _ => panic!(),
    },
    wrong_entity
);
assert_eq!(
    match query
        .get_many_unique_mut(UniqueEntityArray::from([invalid_entity]))
        .unwrap_err()
    {
        QueryEntityError::QueryDoesNotMatch(entity, _) => entity,
        _ => panic!(),
    },
    invalid_entity
);

See also

get_many_unique to get read-only query items.

pub fn get_many_mut_inner<const N: usize>( self, entities: [Entity; N], ) -> Result<[D::Item<'w, 's>; N], QueryEntityError>

Returns the query items for the given array of Entity. This consumes the Query to return results with the actual “inner” world lifetime.

The returned query items are in the same order as the input. In case of a nonexisting entity, duplicate entities or mismatched component, a QueryEntityError is returned instead.

See also

get_many to get read-only query items without checking for duplicate entities.

get_many_mut to get items using a mutable reference.

get_many_inner to get read-only query items with the actual “inner” world lifetime.

pub fn get_many_inner<const N: usize>( self, entities: [Entity; N], ) -> Result<[D::Item<'w, 's>; N], QueryEntityError>
where D: ReadOnlyQueryData,

Returns the query items for the given array of Entity. This consumes the Query to return results with the actual “inner” world lifetime.

The returned query items are in the same order as the input. In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

See also

get_many to get read-only query items without checking for duplicate entities.

get_many_mut to get items using a mutable reference.

get_many_mut_inner to get mutable query items with the actual “inner” world lifetime.

pub fn get_many_unique_inner<const N: usize>( self, entities: UniqueEntityArray<N>, ) -> Result<[D::Item<'w, 's>; N], QueryEntityError>

Returns the query items for the given UniqueEntityArray. This consumes the Query to return results with the actual “inner” world lifetime.

The returned query items are in the same order as the input. In case of a nonexisting entity, duplicate entities or mismatched component, a QueryEntityError is returned instead.

See also

get_many_unique to get read-only query items without checking for duplicate entities.

get_many_unique_mut to get items using a mutable reference.

pub unsafe fn get_unchecked( &self, entity: Entity, ) -> Result<D::Item<'_, 's>, QueryEntityError>

Returns the query item for the given Entity.

In case of a nonexisting entity or mismatched component, a QueryEntityError is returned instead.

This is always guaranteed to run in O(1) time.

Safety

This function makes it possible to violate Rust’s aliasing guarantees. You must make sure this call does not result in multiple mutable references to the same component.

See also

get_mut for the safe version.

pub fn single(&self) -> Result<ROQueryItem<'_, 's, D>, QuerySingleError>

Returns a single read-only query item when there is exactly one entity matching the query.

If the number of query items is not exactly one, a QuerySingleError is returned instead.

Example

fn player_scoring_system(query: Query<&PlayerScore>) {
    match query.single() {
        Ok(PlayerScore(score)) => {
            println!("Score: {}", score);
        }
        Err(QuerySingleError::NoEntities(_)) => {
            println!("Error: There is no player!");
        }
        Err(QuerySingleError::MultipleEntities(_)) => {
            println!("Error: There is more than one player!");
        }
    }
}

See also

single_mut to get the mutable query item.

pub fn single_mut(&mut self) -> Result<D::Item<'_, 's>, QuerySingleError>

Returns a single query item when there is exactly one entity matching the query.

If the number of query items is not exactly one, a QuerySingleError is returned instead.

Example

fn regenerate_player_health_system(mut query: Query<&mut Health, With<Player>>) {
    let mut health = query.single_mut().expect("Error: Could not find a single player.");
    health.0 += 1;
}

See also

single to get the read-only query item.

pub fn single_inner(self) -> Result<D::Item<'w, 's>, QuerySingleError>

Returns a single query item when there is exactly one entity matching the query. This consumes the Query to return results with the actual “inner” world lifetime.

If the number of query items is not exactly one, a QuerySingleError is returned instead.

Example

fn regenerate_player_health_system(query: Query<&mut Health, With<Player>>) {
    let mut health = query.single_inner().expect("Error: Could not find a single player.");
    health.0 += 1;
}

See also

single to get the read-only query item.

single_mut to get the mutable query item.

pub fn is_empty(&self) -> bool

Returns true if there are no query items.

This is equivalent to self.iter().next().is_none(), and thus the worst case runtime will be O(n) where n is the number of potential matches. This can be notably expensive for queries that rely on non-archetypal filters such as Added, Changed or Spawned which must individually check each query result for a match.

Example

Here, the score is increased only if an entity with a Player component is present in the world:

fn update_score_system(query: Query<(), With<Player>>, mut score: ResMut<Score>) {
    if !query.is_empty() {
        score.0 += 1;
    }
}

pub fn contains(&self, entity: Entity) -> bool

Returns true if the given Entity matches the query.

This is always guaranteed to run in O(1) time.

Example

fn targeting_system(in_range_query: Query<&InRange>, target: Res<Target>) {
    if in_range_query.contains(target.entity) {
        println!("Bam!")
    }
}

pub fn count(&self) -> usize

Counts the number of entities that match the query.

This is equivalent to self.iter().count() but may be more efficient in some cases.

If F::IS_ARCHETYPAL is true, this will do work proportional to the number of matched archetypes or tables, but will not iterate each entity. If it is false, it will have to do work for each entity.

Example

fn targeting_system(in_range_query: Query<&InRange>) {
    let count = in_range_query.count();
    println!("{count} targets in range!");
}

pub fn transmute_lens<NewD: QueryData>(&mut self) -> QueryLens<'_, NewD>

Returns a QueryLens that can be used to construct a new Query giving more restrictive access to the entities matched by the current query.

A transmute is valid only if NewD has a subset of the read, write, and required access of the current query. A precise description of the access required by each parameter type is given in the table below, but typical uses are to:

Remove components, e.g. Query<(&A, &B)> to Query<&A>.

Retrieve an existing component with reduced or equal access, e.g. Query<&mut A> to Query<&A> or Query<&T> to Query<Ref<T>>.

Add parameters with no new access, for example adding an Entity parameter.

Note that since filter terms are dropped, non-archetypal filters like Added, Changed and Spawned will not be respected. To maintain or change filter terms see Self::transmute_lens_filtered.

QueryData parameter type	Access required

Entity, EntityLocation, SpawnDetails, &Archetype, Has<T>, PhantomData<T>	No access

EntityMut	Read and write access to all components, but no required access

EntityRef	Read access to all components, but no required access

&T, Ref<T>	Read and required access to T

&mut T, Mut<T>	Read, write and required access to T

Option<T>, AnyOf<(D, ...)>	Read and write access to T, but no required access

Tuples of query data and
#[derive(QueryData)] structs	The union of the access of their subqueries

FilteredEntityRef, FilteredEntityMut	Determined by the QueryBuilder used to construct them. Any query can be transmuted to them, and they will receive the access of the source query. When combined with other QueryData, they will receive any access of the source query that does not conflict with the other data

transmute_lens drops filter terms, but Self::transmute_lens_filtered supports returning a QueryLens with a new filter type - the access required by filter parameters are as follows.

QueryFilter parameter type	Access required

Added<T>, Changed<T>	Read and required access to T

With<T>, Without<T>	No access

Or<(T, ...)>	Read access of the subqueries, but no required access

Tuples of query filters and #[derive(QueryFilter)] structs	The union of the access of their subqueries

Panics

This will panic if the access required by NewD is not a subset of that required by the original fetch D.

Example

fn reusable_function(lens: &mut QueryLens<&A>) {
    assert_eq!(lens.query().single().unwrap().0, 10);
}

// We can use the function in a system that takes the exact query.
fn system_1(mut query: Query<&A>) {
    reusable_function(&mut query.as_query_lens());
}

// We can also use it with a query that does not match exactly
// by transmuting it.
fn system_2(mut query: Query<(&mut A, &B)>) {
    let mut lens = query.transmute_lens::<&A>();
    reusable_function(&mut lens);
}

Examples of valid transmutes

// `&mut T` and `Mut<T>` access the same data and can be transmuted to each other,
// `&T` and `Ref<T>` access the same data and can be transmuted to each other,
// and mutable versions can be transmuted to read-only versions
assert_valid_transmute::<&mut T, &T>();
assert_valid_transmute::<&mut T, Mut<T>>();
assert_valid_transmute::<Mut<T>, &mut T>();
assert_valid_transmute::<&T, Ref<T>>();
assert_valid_transmute::<Ref<T>, &T>();

// The structure can be rearranged, or subqueries dropped
assert_valid_transmute::<(&T, &U), &T>();
assert_valid_transmute::<((&T, &U), &V), (&T, (&U, &V))>();
assert_valid_transmute::<Option<(&T, &U)>, (Option<&T>, Option<&U>)>();

// Queries with no access can be freely added
assert_valid_transmute::<
    &T,
    (&T, Entity, EntityLocation, &Archetype, Has<U>, PhantomData<T>),
>();

// Required access can be transmuted to optional,
// and optional access can be transmuted to other optional access
assert_valid_transmute::<&T, Option<&T>>();
assert_valid_transmute::<AnyOf<(&mut T, &mut U)>, Option<&T>>();
// Note that removing subqueries from `AnyOf` will result
// in an `AnyOf` where all subqueries can yield `None`!
assert_valid_transmute::<AnyOf<(&T, &U, &V)>, AnyOf<(&T, &U)>>();
assert_valid_transmute::<EntityMut, Option<&mut T>>();

// Anything can be transmuted to `FilteredEntityRef` or `FilteredEntityMut`
// This will create a `FilteredEntityMut` that only has read access to `T`
assert_valid_transmute::<&T, FilteredEntityMut>();
// This will create a `FilteredEntityMut` that has no access to `T`,
// read access to `U`, and write access to `V`.
assert_valid_transmute::<(&mut T, &mut U, &mut V), (&mut T, &U, FilteredEntityMut)>();

// `Added<T>` and `Changed<T>` filters have the same access as `&T` data
// Remember that they are only evaluated on the transmuted query, not the original query!
assert_valid_transmute_filtered::<Entity, Changed<T>, &T, ()>();
assert_valid_transmute_filtered::<&mut T, (), &T, Added<T>>();
// Nested inside of an `Or` filter, they have the same access as `Option<&T>`.
assert_valid_transmute_filtered::<Option<&T>, (), Entity, Or<(Changed<T>, With<U>)>>();

pub fn transmute_lens_inner<NewD: QueryData>(self) -> QueryLens<'w, NewD>

Returns a QueryLens that can be used to construct a new Query giving more restrictive access to the entities matched by the current query.

This consumes the Query to return results with the actual “inner” world lifetime.

See Self::transmute_lens for a description of allowed transmutes.

Panics

This will panic if NewD is not a subset of the original fetch D

Example

fn reusable_function(mut lens: QueryLens<&A>) {
    assert_eq!(lens.query().single().unwrap().0, 10);
}

// We can use the function in a system that takes the exact query.
fn system_1(query: Query<&A>) {
    reusable_function(query.into_query_lens());
}

// We can also use it with a query that does not match exactly
// by transmuting it.
fn system_2(query: Query<(&mut A, &B)>) {
    let mut lens = query.transmute_lens_inner::<&A>();
    reusable_function(lens);
}

See also

transmute_lens to convert to a lens using a mutable borrow of the Query.

pub fn transmute_lens_filtered<NewD: QueryData, NewF: QueryFilter>( &mut self, ) -> QueryLens<'_, NewD, NewF>

Equivalent to Self::transmute_lens but also includes a QueryFilter type.

See Self::transmute_lens for a description of allowed transmutes.

Note that the lens will iterate the same tables and archetypes as the original query. This means that additional archetypal query terms like With and Without will not necessarily be respected and non-archetypal terms like Added, Changed and Spawned will only be respected if they are in the type signature.

pub fn transmute_lens_filtered_inner<NewD: QueryData, NewF: QueryFilter>( self, ) -> QueryLens<'w, NewD, NewF>

Equivalent to Self::transmute_lens_inner but also includes a QueryFilter type. This consumes the Query to return results with the actual “inner” world lifetime.

See Self::transmute_lens for a description of allowed transmutes.

Note that the lens will iterate the same tables and archetypes as the original query. This means that additional archetypal query terms like With and Without will not necessarily be respected and non-archetypal terms like Added, Changed and Spawned will only be respected if they are in the type signature.

See also

transmute_lens_filtered to convert to a lens using a mutable borrow of the Query.

pub fn as_query_lens(&mut self) -> QueryLens<'_, D>

Gets a QueryLens with the same accesses as the existing query

pub fn into_query_lens(self) -> QueryLens<'w, D>

Gets a QueryLens with the same accesses as the existing query

See also

as_query_lens to convert to a lens using a mutable borrow of the Query.

pub fn join<'a, OtherD: QueryData, NewD: QueryData>( &'a mut self, other: &'a mut Query<'_, '_, OtherD>, ) -> QueryLens<'a, NewD>

Returns a QueryLens that can be used to get a query with the combined fetch.

For example, this can take a Query<&A> and a Query<&B> and return a Query<(&A, &B)>. The returned query will only return items with both A and B. Note that since filters are dropped, non-archetypal filters like Added, Changed and Spawned will not be respected. To maintain or change filter terms see Self::join_filtered.

Example

fn system(
    mut transforms: Query<&Transform>,
    mut players: Query<&Player>,
    mut enemies: Query<&Enemy>
) {
    let mut players_transforms: QueryLens<(&Transform, &Player)> = transforms.join(&mut players);
    for (transform, player) in &players_transforms.query() {
        // do something with a and b
    }

    let mut enemies_transforms: QueryLens<(&Transform, &Enemy)> = transforms.join(&mut enemies);
    for (transform, enemy) in &enemies_transforms.query() {
        // do something with a and b
    }
}

Panics

This will panic if NewD is not a subset of the union of the original fetch Q and OtherD.

Allowed Transmutes

Like transmute_lens the query terms can be changed with some restrictions. See Self::transmute_lens for more details.

pub fn join_inner<OtherD: QueryData, NewD: QueryData>( self, other: Query<'w, '_, OtherD>, ) -> QueryLens<'w, NewD>

Returns a QueryLens that can be used to get a query with the combined fetch. This consumes the Query to return results with the actual “inner” world lifetime.

For example, this can take a Query<&A> and a Query<&B> and return a Query<(&A, &B)>. The returned query will only return items with both A and B. Note that since filters are dropped, non-archetypal filters like Added, Changed and Spawned will not be respected. To maintain or change filter terms see Self::join_filtered.

Panics

This will panic if NewD is not a subset of the union of the original fetch Q and OtherD.

Allowed Transmutes

Like transmute_lens the query terms can be changed with some restrictions. See Self::transmute_lens for more details.

See also

join to join using a mutable borrow of the Query.

pub fn join_filtered<'a, OtherD: QueryData, OtherF: QueryFilter, NewD: QueryData, NewF: QueryFilter>( &'a mut self, other: &'a mut Query<'_, '_, OtherD, OtherF>, ) -> QueryLens<'a, NewD, NewF>

Equivalent to Self::join but also includes a QueryFilter type.

Note that the lens with iterate a subset of the original queries’ tables and archetypes. This means that additional archetypal query terms like With and Without will not necessarily be respected and non-archetypal terms like Added, Changed and Spawned will only be respected if they are in the type signature.

pub fn join_filtered_inner<OtherD: QueryData, OtherF: QueryFilter, NewD: QueryData, NewF: QueryFilter>( self, other: Query<'w, '_, OtherD, OtherF>, ) -> QueryLens<'w, NewD, NewF>

Equivalent to Self::join_inner but also includes a QueryFilter type. This consumes the Query to return results with the actual “inner” world lifetime.

Note that the lens with iterate a subset of the original queries’ tables and archetypes. This means that additional archetypal query terms like With and Without will not necessarily be respected and non-archetypal terms like Added, Changed and Spawned will only be respected if they are in the type signature.

See also

join_filtered to join using a mutable borrow of the Query.

impl<'w, 's, D: ReadOnlyQueryData, F: QueryFilter> Query<'w, 's, D, F>

pub fn iter_inner(&self) -> QueryIter<'w, 's, D::ReadOnly, F>

Returns an Iterator over the query items, with the actual “inner” world lifetime.

This can only return immutable data (mutable data will be cast to an immutable form). See Self::iter_mut for queries that contain at least one mutable component.

Example

Here, the report_names_system iterates over the Player component of every entity that contains it:

fn report_names_system(query: Query<&Player>) {
    for player in &query {
        println!("Say hello to {}!", player.name);
    }
}