Trait bevy_ecs::system::SystemParamFetch

pub trait SystemParamFetch<'a>: SystemParamState {
    type Item;
    unsafe fn get_param(
        state: &'a mut Self, 
        system_state: &'a SystemState, 
        world: &'a World, 
        change_tick: u32
    ) -> Self::Item;
}

Associated Types

type Item

Loading content...
Required methods

unsafe fn get_param(
state: &'a mut Self,
system_state: &'a SystemState,
world: &'a World,
change_tick: u32
) -> Self::Item

Safety

This call might access any of the input parameters in an unsafe way. Make sure the data access is safe in the context of the system scheduler

Loading content...