Module bevy_ecs::world

Defines the World and APIs for accessing it directly.

Re-exports

pub use crate::change_detection::Mut;

pub use crate::change_detection::Ref;

pub use crate::change_detection::CHECK_TICK_THRESHOLD;

Modules

error

Contains error types returned by bevy’s schedule.

unsafe_world_cell

Contains types that allow disjoint mutable access to a World.

Structs

EntityMut

A mutable reference to a particular Entity and all of its components

EntityRef

A read-only reference to a particular Entity and all of its components

SpawnBatchIter

An iterator that spawns a series of entities and returns the ID of each spawned entity.

World

Stores and exposes operations on entities, components, resources, and their associated metadata.

WorldBorrow

A read-only borrow of some data stored in a World. This type is returned by WorldCell, which uses run-time checks to ensure that the borrow does not violate Rust’s aliasing rules.

WorldBorrowMut

A mutable borrow of some data stored in a World. This type is returned by WorldCell, which uses run-time checks to ensure that the borrow does not violate Rust’s aliasing rules.

WorldCell

Exposes safe mutable access to multiple resources at a time in a World. Attempting to access World in a way that violates Rust’s mutability rules will panic thanks to runtime checks.

WorldId

A unique identifier for a World.

Traits

FromWorld

Creates an instance of the type this trait is implemented for using data from the supplied World.