bevy_ecs::world

Struct DeferredWorld

pub struct DeferredWorld<'w> { /* private fields */ }

A World reference that disallows structural ECS changes. This includes initializing resources, registering components or spawning entities.

This means that in order to add entities, for example, you will need to use commands instead of the world directly.

Implementations

impl<'w> DeferredWorld<'w>

pub fn reborrow(&mut self) -> DeferredWorld<'_>

Reborrow self as a new instance of DeferredWorld

pub fn commands(&mut self) -> Commands<'_, '_>

Creates a Commands instance that pushes to the world’s command queue

pub fn get_mut<T: Component<Mutability = Mutable>>( &mut self, entity: Entity, ) -> Option<Mut<'_, T>>

Retrieves a mutable reference to the given entity’s Component of the given type. Returns None if the entity does not have a Component of the given type.

pub fn get_entity_mut<F: WorldEntityFetch>( &mut self, entities: F, ) -> Result<F::DeferredMut<'_>, EntityMutableFetchError>

Returns EntityMuts that expose read and write operations for the given entities, returning Err if any of the given entities do not exist. Instead of immediately unwrapping the value returned from this function, prefer World::entity_mut.

This function supports fetching a single entity or multiple entities:

Pass an Entity to receive a single EntityMut.

Pass a slice of Entitys to receive a Vec<EntityMut>.

Pass an array of Entitys to receive an equally-sized array of EntityMuts.

Pass an &EntityHashSet to receive an EntityHashMap<EntityMut>.

As DeferredWorld does not allow structural changes, all returned references are EntityMuts, which do not allow structural changes (i.e. adding/removing components or despawning the entity).

Errors

Returns EntityMutableFetchError::EntityDoesNotExist if any of the given entities do not exist in the world.

Only the first entity found to be missing will be returned.

Returns EntityMutableFetchError::AliasedMutability if the same entity is requested multiple times.

Examples

For examples, see DeferredWorld::entity_mut.

pub fn entity_mut<F: WorldEntityFetch>( &mut self, entities: F, ) -> F::DeferredMut<'_>

Returns EntityMuts that expose read and write operations for the given entities. This will panic if any of the given entities do not exist. Use DeferredWorld::get_entity_mut if you want to check for entity existence instead of implicitly panicking.

This function supports fetching a single entity or multiple entities:

Pass an Entity to receive a single EntityMut.

Pass a slice of Entitys to receive a Vec<EntityMut>.

Pass an array of Entitys to receive an equally-sized array of EntityMuts.

Pass an &EntityHashSet to receive an EntityHashMap<EntityMut>.

As DeferredWorld does not allow structural changes, all returned references are EntityMuts, which do not allow structural changes (i.e. adding/removing components or despawning the entity).

Panics

If any of the given entities do not exist in the world.

Examples

Single Entity

#[derive(Component)]
struct Position {
  x: f32,
  y: f32,
}

let mut world: DeferredWorld = // ...

let mut entity_mut = world.entity_mut(entity);
let mut position = entity_mut.get_mut::<Position>().unwrap();
position.y = 1.0;
assert_eq!(position.x, 0.0);

Array of Entitys

#[derive(Component)]
struct Position {
  x: f32,
  y: f32,
}

let mut world: DeferredWorld = // ...

let [mut e1_ref, mut e2_ref] = world.entity_mut([e1, e2]);
let mut e1_position = e1_ref.get_mut::<Position>().unwrap();
e1_position.x = 1.0;
assert_eq!(e1_position.x, 1.0);
let mut e2_position = e2_ref.get_mut::<Position>().unwrap();
e2_position.x = 2.0;
assert_eq!(e2_position.x, 2.0);

Slice of Entitys

#[derive(Component)]
struct Position {
  x: f32,
  y: f32,
}

let mut world: DeferredWorld = // ...

let ids = vec![e1, e2, e3];
for mut eref in world.entity_mut(&ids[..]) {
    let mut pos = eref.get_mut::<Position>().unwrap();
    pos.y = 2.0;
    assert_eq!(pos.y, 2.0);
}

&EntityHashSet

#[derive(Component)]
struct Position {
  x: f32,
  y: f32,
}

let mut world: DeferredWorld = // ...

let ids = EntityHashSet::from_iter([e1, e2, e3]);
for (_id, mut eref) in world.entity_mut(&ids) {
    let mut pos = eref.get_mut::<Position>().unwrap();
    pos.y = 2.0;
    assert_eq!(pos.y, 2.0);
}

pub fn entities_and_commands(&mut self) -> (EntityFetcher<'_>, Commands<'_, '_>)

Simultaneously provides access to entity data and a command queue, which will be applied when the World is next flushed.

This allows using borrowed entity data to construct commands where the borrow checker would otherwise prevent it.

See World::entities_and_commands for the non-deferred version.

Example

#[derive(Component)]
struct Targets(Vec<Entity>);
#[derive(Component)]
struct TargetedBy(Entity);

let mut world: DeferredWorld = // ...
let (entities, mut commands) = world.entities_and_commands();

let entity = entities.get(eid).unwrap();
for &target in entity.get::<Targets>().unwrap().0.iter() {
    commands.entity(target).insert(TargetedBy(eid));
}

pub fn query<'s, D: QueryData, F: QueryFilter>( &mut self, state: &'s mut QueryState<D, F>, ) -> Query<'_, 's, D, F>

Returns Query for the given QueryState, which is used to efficiently run queries on the World by storing and reusing the QueryState.

Panics

If state is from a different world then self

pub fn resource_mut<R: Resource>(&mut self) -> Mut<'_, R>

Gets a mutable reference to the resource of the given type

Panics

Panics if the resource does not exist. Use get_resource_mut instead if you want to handle this case.

pub fn get_resource_mut<R: Resource>(&mut self) -> Option<Mut<'_, R>>

Gets a mutable reference to the resource of the given type if it exists

pub fn non_send_resource_mut<R: 'static>(&mut self) -> Mut<'_, R>

Gets a mutable reference to the non-send resource of the given type, if it exists.

Panics

Panics if the resource does not exist. Use get_non_send_resource_mut instead if you want to handle this case.

This function will panic if it isn’t called from the same thread that the resource was inserted from.

pub fn get_non_send_resource_mut<R: 'static>(&mut self) -> Option<Mut<'_, R>>

Gets a mutable reference to the non-send resource of the given type, if it exists. Otherwise returns None.

Panics

This function will panic if it isn’t called from the same thread that the resource was inserted from.

pub fn write_message<M: Message>(&mut self, message: M) -> Option<MessageId<M>>

Writes a Message. This method returns the MessageId of the written message, or None if the message could not be written.

pub fn send_event<E: Message>(&mut self, event: E) -> Option<MessageId<E>>

👎Deprecated since 0.17.0: Use DeferredWorld::write_message instead.

Writes a Message. This method returns the ID of the written event, or None if the event could not be written.

pub fn write_message_default<E: Message + Default>( &mut self, ) -> Option<MessageId<E>>

Writes the default value of the Message of type E. This method returns the MessageId of the written event, or None if the event could not be written.

pub fn send_event_default<E: Message + Default>( &mut self, ) -> Option<MessageId<E>>

👎Deprecated since 0.17.0: Use DeferredWorld::write_message_default instead.

Writes the default value of the Message of type E. This method returns the ID of the written event, or None if the event could not be written.

pub fn write_message_batch<E: Message>( &mut self, events: impl IntoIterator<Item = E>, ) -> Option<WriteBatchIds<E>>

Writes a batch of Messages from an iterator. This method returns the IDs of the written events, or None if the event could not be written.

pub fn send_event_batch<E: Message>( &mut self, events: impl IntoIterator<Item = E>, ) -> Option<WriteBatchIds<E>>

👎Deprecated since 0.17.0: Use DeferredWorld::write_message_batch instead.

Writes a batch of Messages from an iterator. This method returns the IDs of the written events, or None if the event could not be written.

pub fn get_resource_mut_by_id( &mut self, component_id: ComponentId, ) -> Option<MutUntyped<'_>>

Gets a pointer to the resource with the id ComponentId if it exists. The returned pointer may be used to modify the resource, as long as the mutable borrow of the World is still valid.

You should prefer to use the typed API World::get_resource_mut where possible and only use this in cases where the actual types are not known at compile time.

pub fn get_non_send_mut_by_id( &mut self, component_id: ComponentId, ) -> Option<MutUntyped<'_>>

Gets a !Send resource to the resource with the id ComponentId if it exists. The returned pointer may be used to modify the resource, as long as the mutable borrow of the World is still valid.

You should prefer to use the typed API World::get_resource_mut where possible and only use this in cases where the actual types are not known at compile time.

Panics

This function will panic if it isn’t called from the same thread that the resource was inserted from.

pub fn get_mut_by_id( &mut self, entity: Entity, component_id: ComponentId, ) -> Option<MutUntyped<'_>>

Retrieves a mutable untyped reference to the given entity’s Component of the given ComponentId. Returns None if the entity does not have a Component of the given type.

You should prefer to use the typed API World::get_mut where possible and only use this in cases where the actual types are not known at compile time.

pub unsafe fn trigger_raw<'a, E: Event>( &mut self, event_key: EventKey, event: &mut E, trigger: &mut E::Trigger<'a>, caller: MaybeLocation, )

Triggers all event observers for the given targets

Safety

Caller must ensure E is accessible as the type represented by event_key

pub fn trigger<'a>(&mut self, event: impl Event<Trigger<'a>: Default>)

Sends a global Event without any targets.

This will run any Observer of the given Event that isn’t scoped to specific targets.

pub fn as_unsafe_world_cell(&mut self) -> UnsafeWorldCell<'_>

Gets an UnsafeWorldCell containing the underlying world.

Safety

must only be used to make non-structural ECS changes