bevy_ecs::world
Struct EntityMut

pub struct EntityMut<'w>(/* private fields */);

Provides mutable access to a single entity and all of its components.

Contrast with EntityWorldMut, which allows adding and removing components, despawning the entity, and provides mutable access to the entire world. Because of this, EntityWorldMut cannot coexist with any other world accesses.

Examples

Disjoint mutable access.

fn disjoint_system(
    query1: Query<EntityMut, With<A>>,
    query2: Query<EntityMut, Without<A>>,
) {
    // ...
}

Implementations

impl<'w> EntityMut<'w>

pub fn reborrow(&mut self) -> EntityMut<'_>

Returns a new instance with a shorter lifetime. This is useful if you have &mut EntityMut, but you need EntityMut.

pub fn as_readonly(&self) -> EntityRef<'_>

Gets read-only access to all of the entity’s components.

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn location(&self) -> EntityLocation

Gets metadata indicating the location where the current entity is stored.

pub fn archetype(&self) -> &Archetype

Returns the archetype that the current entity belongs to.

pub fn contains<T: Component>(&self) -> bool

Returns true if the current entity has a component of type T. Otherwise, this returns false.

Notes

If you do not know the concrete type of a component, consider using Self::contains_id or Self::contains_type_id.

pub fn contains_id(&self, component_id: ComponentId) -> bool

Returns true if the current entity has a component identified by component_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you know the component’s TypeId but not its ComponentId, consider using Self::contains_type_id.

pub fn contains_type_id(&self, type_id: TypeId) -> bool

Returns true if the current entity has a component with the type identified by type_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you have a ComponentId instead of a TypeId, consider using Self::contains_id.

pub fn get<T: Component>(&self) -> Option<&T>

Gets access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

pub fn components<Q: ReadOnlyQueryData>(&self) -> Q::Item<'_>

Returns read-only components for the current entity that match the query Q.

Panics

If the entity does not have the components required by the query Q.

pub fn get_components<Q: ReadOnlyQueryData>(&self) -> Option<Q::Item<'_>>

Returns read-only components for the current entity that match the query Q, or None if the entity does not have the components required by the query Q.

pub fn into_borrow<T: Component>(self) -> Option<&'w T>

Consumes self and gets access to the component of type T with the world 'w lifetime for the current entity.

Returns None if the entity does not have a component of type T.

pub fn get_ref<T: Component>(&self) -> Option<Ref<'_, T>>

Gets access to the component of type T for the current entity, including change detection information as a Ref.

Returns None if the entity does not have a component of type T.

pub fn into_ref<T: Component>(self) -> Option<Ref<'w, T>>

Consumes self and gets access to the component of type T with world 'w lifetime for the current entity, including change detection information as a [Ref<'w>].

Returns None if the entity does not have a component of type T.

pub fn get_mut<T: Component>(&mut self) -> Option<Mut<'_, T>>

Gets mutable access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

pub fn into_mut<T: Component>(self) -> Option<Mut<'w, T>>

Consumes self and gets mutable access to the component of type T with the world 'w lifetime for the current entity. Returns None if the entity does not have a component of type T.

pub fn get_change_ticks<T: Component>(&self) -> Option<ComponentTicks>

Retrieves the change ticks for the given component. This can be useful for implementing change detection in custom runtimes.

pub fn get_change_ticks_by_id( &self, component_id: ComponentId, ) -> Option<ComponentTicks>

Retrieves the change ticks for the given ComponentId. This can be useful for implementing change detection in custom runtimes.

You should prefer to use the typed API EntityWorldMut::get_change_ticks where possible and only use this in cases where the actual component types are not known at compile time.

pub fn get_by_id<F: DynamicComponentFetch>( &self, component_ids: F, ) -> Result<F::Ref<'_>, EntityComponentError>

Returns untyped read-only reference(s) to component(s) for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityMut::get where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityMut::get, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Examples

For examples on how to use this method, see EntityRef::get_by_id.

pub fn into_borrow_by_id<F: DynamicComponentFetch>( self, component_ids: F, ) -> Result<F::Ref<'w>, EntityComponentError>

Consumes self and returns untyped read-only reference(s) to component(s) with lifetime 'w for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityMut::into_borrow where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityMut::into_borrow, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Examples

For examples on how to use this method, see EntityRef::get_by_id.

pub fn get_mut_by_id<F: DynamicComponentFetch>( &mut self, component_ids: F, ) -> Result<F::Mut<'_>, EntityComponentError>

Returns untyped mutable reference(s) to component(s) for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityMut::get_mut where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityMut::get_mut, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Returns EntityComponentError::AliasedMutability if a component is requested multiple times.

Examples

Single ComponentId

let entity = world.spawn(Foo(42)).id();

// Grab the component ID for `Foo` in whatever way you like.
let component_id = world.register_component::<Foo>();

// Then, get the component by ID.
let mut entity_mut = world.entity_mut(entity);
let mut ptr = entity_mut.get_mut_by_id(component_id)

Array of ComponentIds

let entity = world.spawn((X(42), Y(10))).id();

// Grab the component IDs for `X` and `Y` in whatever way you like.
let x_id = world.register_component::<X>();
let y_id = world.register_component::<Y>();

// Then, get the components by ID. You'll receive a same-sized array.
let mut entity_mut = world.entity_mut(entity);
let Ok([mut x_ptr, mut y_ptr]) = entity_mut.get_mut_by_id([x_id, y_id]) else {
    // Up to you to handle if a component is missing from the entity.
};

Slice of ComponentIds

let entity = world.spawn((X(42), Y(10))).id();

// Grab the component IDs for `X` and `Y` in whatever way you like.
let x_id = world.register_component::<X>();
let y_id = world.register_component::<Y>();

// Then, get the components by ID. You'll receive a vec of ptrs.
let mut entity_mut = world.entity_mut(entity);
let ptrs = entity_mut.get_mut_by_id(&[x_id, y_id] as &[ComponentId])

HashSet of ComponentIds

let entity = world.spawn((X(42), Y(10))).id();

// Grab the component IDs for `X` and `Y` in whatever way you like.
let x_id = world.register_component::<X>();
let y_id = world.register_component::<Y>();

// Then, get the components by ID. You'll receive a `HashMap` of ptrs.
let mut entity_mut = world.entity_mut(entity);
let mut ptrs = entity_mut.get_mut_by_id(&HashSet::from_iter([x_id, y_id]))

pub fn into_mut_by_id<F: DynamicComponentFetch>( self, component_ids: F, ) -> Result<F::Mut<'w>, EntityComponentError>

Consumes self and returns untyped mutable reference(s) to component(s) with lifetime 'w for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityMut::into_mut where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityMut::into_mut, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Returns EntityComponentError::AliasedMutability if a component is requested multiple times.

Examples

For examples on how to use this method, see EntityMut::get_mut_by_id.