bevy_ecs::world

Struct EntityRef

pub struct EntityRef<'w> { /* private fields */ }

A read-only reference to a particular Entity and all of its components.

Examples

Read-only access disjoint with mutable access.

fn disjoint_system(
    query1: Query<&mut A>,
    query2: Query<EntityRef, Without<A>>,
) {
    // ...
}

Implementations

impl<'w> EntityRef<'w>

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn location(&self) -> EntityLocation

Gets metadata indicating the location where the current entity is stored.

pub fn archetype(&self) -> &Archetype

Returns the archetype that the current entity belongs to.

pub fn contains<T: Component>(&self) -> bool

Returns true if the current entity has a component of type T. Otherwise, this returns false.

Notes

If you do not know the concrete type of a component, consider using Self::contains_id or Self::contains_type_id.

pub fn contains_id(&self, component_id: ComponentId) -> bool

Returns true if the current entity has a component identified by component_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you know the component’s TypeId but not its ComponentId, consider using Self::contains_type_id.

pub fn contains_type_id(&self, type_id: TypeId) -> bool

Returns true if the current entity has a component with the type identified by type_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you have a ComponentId instead of a TypeId, consider using Self::contains_id.

pub fn get<T: Component>(&self) -> Option<&'w T>

Gets access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

pub fn get_ref<T: Component>(&self) -> Option<Ref<'w, T>>

Gets access to the component of type T for the current entity, including change detection information as a Ref.

Returns None if the entity does not have a component of type T.

pub fn get_change_ticks<T: Component>(&self) -> Option<ComponentTicks>

Retrieves the change ticks for the given component. This can be useful for implementing change detection in custom runtimes.

pub fn get_change_ticks_by_id( &self, component_id: ComponentId, ) -> Option<ComponentTicks>

Retrieves the change ticks for the given ComponentId. This can be useful for implementing change detection in custom runtimes.

You should prefer to use the typed API EntityRef::get_change_ticks where possible and only use this in cases where the actual component types are not known at compile time.

pub fn get_by_id<F: DynamicComponentFetch>( &self, component_ids: F, ) -> Result<F::Ref<'w>, EntityComponentError>

Returns untyped read-only reference(s) to component(s) for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityRef::get where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityRef::get, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Examples

Single ComponentId

let entity = world.spawn(Foo(42)).id();

// Grab the component ID for `Foo` in whatever way you like.
let component_id = world.register_component::<Foo>();

// Then, get the component by ID.
let ptr = world.entity(entity).get_by_id(component_id);

Array of ComponentIds

let entity = world.spawn((X(42), Y(10))).id();

// Grab the component IDs for `X` and `Y` in whatever way you like.
let x_id = world.register_component::<X>();
let y_id = world.register_component::<Y>();

// Then, get the components by ID. You'll receive a same-sized array.
let Ok([x_ptr, y_ptr]) = world.entity(entity).get_by_id([x_id, y_id]) else {
    // Up to you to handle if a component is missing from the entity.
};

Slice of ComponentIds

let entity = world.spawn((X(42), Y(10))).id();

// Grab the component IDs for `X` and `Y` in whatever way you like.
let x_id = world.register_component::<X>();
let y_id = world.register_component::<Y>();

// Then, get the components by ID. You'll receive a vec of ptrs.
let ptrs = world.entity(entity).get_by_id(&[x_id, y_id] as &[ComponentId]);

HashSet of ComponentIds

let entity = world.spawn((X(42), Y(10))).id();

// Grab the component IDs for `X` and `Y` in whatever way you like.
let x_id = world.register_component::<X>();
let y_id = world.register_component::<Y>();

// Then, get the components by ID. You'll receive a vec of ptrs.
let ptrs = world.entity(entity).get_by_id(&HashSet::from_iter([x_id, y_id]));

pub fn components<Q: ReadOnlyQueryData + ReleaseStateQueryData>( &self, ) -> Q::Item<'w, 'static>

Returns read-only components for the current entity that match the query Q.

Panics

If the entity does not have the components required by the query Q.

pub fn get_components<Q: ReadOnlyQueryData + ReleaseStateQueryData>( &self, ) -> Option<Q::Item<'w, 'static>>

Returns read-only components for the current entity that match the query Q, or None if the entity does not have the components required by the query Q.

pub fn spawned_by(&self) -> MaybeLocation

Returns the source code location from which this entity has been spawned.

pub fn spawn_tick(&self) -> Tick

Returns the Tick at which this entity has been spawned.