bevy_ecs::world
Struct EntityRefExcept

pub struct EntityRefExcept<'w, B>
where
    B: Bundle,
{ /* private fields */ }

Provides read-only access to a single entity and all its components, save for an explicitly-enumerated set.

Implementations

impl<'w, B> EntityRefExcept<'w, B>
where B: Bundle,

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn get<C>(&self) -> Option<&'w C>
where C: Component,

Gets access to the component of type C for the current entity. Returns None if the component doesnโt have a component of that type or if the type is one of the excluded components.

pub fn get_ref<C>(&self) -> Option<Ref<'w, C>>
where C: Component,

Gets access to the component of type C for the current entity, including change detection information. Returns None if the component doesnโt have a component of that type or if the type is one of the excluded components.