bevy_ecs::world
Struct FilteredEntityMut

pub struct FilteredEntityMut<'w> { /* private fields */ }

Provides mutable access to a single entity and some of its components defined by the contained Access.

To define the access when used as a QueryData, use a QueryBuilder or QueryParamBuilder. The FilteredEntityMut must be the entire QueryData, and not nested inside a tuple with other data.

// This gives the `FilteredEntityMut` access to `&mut A`.
let mut query = QueryBuilder::<FilteredEntityMut>::new(&mut world)
    .data::<&mut A>()
    .build();

let mut filtered_entity: FilteredEntityMut = query.single_mut(&mut world);
let component: Mut<A> = filtered_entity.get_mut().unwrap();

// Here `FilteredEntityMut` is nested in a tuple, so it does not have access to `&mut A`.
let mut query = QueryBuilder::<(Entity, FilteredEntityMut)>::new(&mut world)
    .data::<&mut A>()
    .build();

let (_, mut filtered_entity) = query.single_mut(&mut world);
assert!(filtered_entity.get_mut::<A>().is_none());

Implementations

impl<'w> FilteredEntityMut<'w>

pub fn reborrow(&mut self) -> FilteredEntityMut<'_>

Returns a new instance with a shorter lifetime. This is useful if you have &mut FilteredEntityMut, but you need FilteredEntityMut.

pub fn as_readonly(&self) -> FilteredEntityRef<'_>

Gets read-only access to all of the entity’s components.

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn location(&self) -> EntityLocation

Gets metadata indicating the location where the current entity is stored.

pub fn archetype(&self) -> &Archetype

Returns the archetype that the current entity belongs to.

pub fn access(&self) -> &Access<ComponentId>

Returns a reference to the underlying Access.

pub fn contains<T: Component>(&self) -> bool

Returns true if the current entity has a component of type T. Otherwise, this returns false.

Notes

If you do not know the concrete type of a component, consider using Self::contains_id or Self::contains_type_id.

pub fn contains_id(&self, component_id: ComponentId) -> bool

Returns true if the current entity has a component identified by component_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you know the component’s TypeId but not its ComponentId, consider using Self::contains_type_id.

pub fn contains_type_id(&self, type_id: TypeId) -> bool

Returns true if the current entity has a component with the type identified by type_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you have a ComponentId instead of a TypeId, consider using Self::contains_id.

pub fn get<T: Component>(&self) -> Option<&T>

Gets access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

pub fn get_ref<T: Component>(&self) -> Option<Ref<'_, T>>

Gets access to the component of type T for the current entity, including change detection information as a Ref.

Returns None if the entity does not have a component of type T.

pub fn get_mut<T: Component>(&mut self) -> Option<Mut<'_, T>>

Gets mutable access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

pub fn into_mut<T: Component>(self) -> Option<Mut<'w, T>>

Consumes self and gets mutable access to the component of type T with the world 'w lifetime for the current entity. Returns None if the entity does not have a component of type T.

pub fn get_change_ticks<T: Component>(&self) -> Option<ComponentTicks>

Retrieves the change ticks for the given component. This can be useful for implementing change detection in custom runtimes.

pub fn get_change_ticks_by_id( &self, component_id: ComponentId, ) -> Option<ComponentTicks>

Retrieves the change ticks for the given ComponentId. This can be useful for implementing change detection in custom runtimes.

You should prefer to use the typed API Self::get_change_ticks where possible and only use this in cases where the actual component types are not known at compile time.

pub fn get_by_id(&self, component_id: ComponentId) -> Option<Ptr<'_>>

Gets the component of the given ComponentId from the entity.

You should prefer to use the typed API Self::get where possible and only use this in cases where the actual component types are not known at compile time.

Unlike FilteredEntityMut::get, this returns a raw pointer to the component, which is only valid while the FilteredEntityMut is alive.

pub fn get_mut_by_id( &mut self, component_id: ComponentId, ) -> Option<MutUntyped<'_>>

Gets a MutUntyped of the component of the given ComponentId from the entity.

You should prefer to use the typed API Self::get_mut where possible and only use this in cases where the actual component types are not known at compile time.

Unlike FilteredEntityMut::get_mut, this returns a raw pointer to the component, which is only valid while the FilteredEntityMut is alive.